% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{addFirstCdsOnLeaderEnds}
\alias{addFirstCdsOnLeaderEnds}
\title{When reassigning Transcript start site, often you want to add downstream too.
This is a simple way to do that}
\usage{
addFirstCdsOnLeaderEnds(fiveUTRs, cds)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}

\item{cds}{If you want to extend 5' leaders downstream, to catch uorfs going into cds, include it.}
}
\value{
a GRangesList of cds exons added to ends
}
\description{
When reassigning Transcript start site, often you want to add downstream too.
This is a simple way to do that
}
