% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{allFeatures}
\alias{allFeatures}
\title{Get all possible feature in ORFik}
\usage{
allFeatures(grl, RFP, RNA = NULL, Gtf = NULL, tx = NULL,
  fiveUTRs = NULL, cds = NULL, threeUTRs = NULL, faFile = NULL,
  riboStart = 26, riboStop = 34, extension = NULL, orfFeatures = T,
  cageFiveUTRs = NULL)
}
\arguments{
\item{grl}{a GRangesList object with usually ORFs, but can also be
either leaders, cds', 3' utrs or  ORFs are a special case,
see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}

\item{Gtf}{a TxDb object of a gtf file,}

\item{tx}{a GrangesList of transcripts,
normally called from: exonsBy(Gtf, by = "tx", use.names = T)
only add this if you are not including Gtf file
You do not need to reassign these to the cage peaks, it will do it for you.}

\item{fiveUTRs}{fiveUTRs as GRangesList, must be original unchanged fiveUTRs}

\item{cds}{a GRangesList of coding sequences}

\item{threeUTRs}{a GrangesList of transcript 3' utrs,
normally called from: threeUTRsByTranscript(Gtf, use.names = T)}

\item{faFile}{a FaFile from the fasta file, see ?FaFile}

\item{riboStart}{usually 26, the start of the floss interval, see ?floss}

\item{riboStop}{usually 34, the end of the floss interval}

\item{extension}{a numeric/integer needs to be set! set to 0 if you did not
use cage, if you used cage to change tss' when finding the orfs,
standard cage extension is 1000}

\item{orfFeatures}{a logical,  is the grl a list of orfs? Must be assigned.}

\item{cageFiveUTRs}{a GRangesList, if you used cage-data to extend 5' utrs,
include this, also extension must match with the extension used for these.}
}
\value{
a data.table with scores, each column is one score type, name of
 columns are the names of the scores, i.g floss or fpkmRFP.
}
\description{
If you want to get all the features easily, run this function.
Each feature have a link to an article describing feature,
try ?floss
}
\examples{
 \dontrun{
 #The easiest way to run the method is to include as this:
 allFeatures(grl = grl, orfFeatures =  T, RFP = RFP, RNA = RNA, Gtf = Gtf,
  faFile = faFile, extension = 0)
 #The other arguments will then be found by the Gtf file,
 #and extension = 0 means
 #you did not use cage to extend 5' utrs
 }
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
