% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFik.R
\docType{package}
\name{amplican}
\alias{ORFik}
\alias{ORFik-package}
\alias{amplican-package}
\title{ORFik for analysis of open reading frames.}
\description{
Main goals:
\enumerate{
\item Find open reading frames
\item Compatible with GRanges
\item Precise quantification of mutation rates.
\item Prepare automatic reports as .Rmd files that are flexible
and open for manipulation.
\item Provide specialized plots for deletions, insertions, mismatches,
variants, heterogeneity of the reads.
}
}
\details{
To learn more about amplican, start with the vignettes:
\code{browseVignettes(package = "amplican")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JokingHero/ORFik}
  \item Report bugs at \url{https://github.com/JokingHero/ORFik/issues}
}

}
