% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_annotations.R
\name{artificial.orfs}
\alias{artificial.orfs}
\title{Create small artificial orfs from cds}
\usage{
artificial.orfs(cds, start5 = 1, end5 = 4, start3 = -4, end3 = 0)
}
\arguments{
\item{cds}{a GRangesList of orfs, must have width
\%\% 3 == 0 and length >= 6}

\item{start5}{integer, default: 1 (start of orf)}

\item{end5}{integer, default: 4 (max 4 codons from start codon)}

\item{start3}{integer, default -4 (max 4 codons from stop codon)}

\item{end3}{integer, default: 0 (end of orf)}
}
\value{
GRangesList of new ORFs
}
\description{
Split cds first in two, a start part and stop part.
Then say how large the two parts can be and merge them together.
It will sample a value in range give.\cr
Parts will be forced to not overlap and can not extend outside
original cds
}
