% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{as.character,GRangesList-method}
\alias{as.character,GRangesList-method}
\title{Convert GRangesList to character vector}
\usage{
\S4method{as.character}{GRangesList}(x, ...)
}
\arguments{
\item{x}{A \code{\link{GRangesList}}}

\item{...}{Not used for now, to preserve generic requirement}
}
\value{
a character vector, 1 element per element in GRangesList
}
\description{
Single exon format:\cr
"1:14598834-14598914:+"\cr
Multi-exon format (exon separator: ';'):\cr
"1:15210514-15210562:+;1:15214895-15215025:+"
}
