% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{bamVarName}
\alias{bamVarName}
\title{Get variable names from experiment}
\usage{
bamVarName(df, skip.replicate = length(unique(df$rep)) == 1,
  skip.condition = length(unique(df$condition)) == 1,
  skip.stage = length(unique(df$stage)) == 1,
  skip.fraction = length(unique(df$fraction)) == 1,
  skip.experiment = !df@expInVarName)
}
\arguments{
\item{df}{an ORFik experiment data.frame}

\item{skip.replicate}{a logical (FALSE), don't include replicate
in variable name.}

\item{skip.condition}{a logical (FALSE), don't include condition
in variable name.}

\item{skip.stage}{a logical (FALSE), don't include stage
in variable name.}

\item{skip.fraction}{a logical (FALSE), don't include fraction}

\item{skip.experiment}{a logical (FALSE), don't include experiment}
}
\value{
variable names of libraries (character vector)
}
\description{
Get variable names from experiment
}
