% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{bamVarNamePicker}
\alias{bamVarNamePicker}
\title{Get variable name per filepath in experiment}
\usage{
bamVarNamePicker(
  df,
  skip.replicate = FALSE,
  skip.condition = FALSE,
  skip.stage = FALSE,
  skip.fraction = FALSE,
  skip.experiment = FALSE,
  skip.libtype = FALSE
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{skip.replicate}{a logical (FALSE), don't include replicate
in variable name.}

\item{skip.condition}{a logical (FALSE), don't include condition
in variable name.}

\item{skip.stage}{a logical (FALSE), don't include stage
in variable name.}

\item{skip.fraction}{a logical (FALSE), don't include fraction}

\item{skip.experiment}{a logical (FALSE), don't include experiment}

\item{skip.libtype}{a logical (FALSE), don't include libtype}
}
\value{
variable name of library (character vector)
}
\description{
Get variable name per filepath in experiment
}
