% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{biotype}
\alias{biotype}
\title{Get ORFik experiment gene biotype}
\usage{
biotype(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a data.table with gene id, gene symbols and tx ids (3 columns)
}
\description{
Loads premade fst table at path:
file.path(refFolder(x), "biotype_per_gene.fst")
}
\examples{
df <- ORFik.template.experiment()
symbols(df)
}
