% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{cageFromFile}
\alias{cageFromFile}
\title{Get Cage-Data From a file-path}
\usage{
cageFromFile(filePath)
}
\arguments{
\item{filePath}{The location of the cage-file}
}
\description{
Get Cage-Data From a file-path
}
