% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{calculateEntropy}
\alias{calculateEntropy}
\title{Calucalte Entropy value of input reads. Based on ...}
\usage{
calculateEntropy(countsOver)
}
\arguments{
\item{countsOver}{A numeric vector containing reads over given ORF}
}
\value{
numeric value of entropy
}
\description{
Calucalte Entropy value of input reads. Based on ...
}
\examples{
calculateEntropy(c(30,15,5,45,23,4,23,4,1,14,7,3,24,6,2,10,6,2))
}

