% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{canonical_isoforms}
\alias{canonical_isoforms}
\title{Get canonical isoforms of organism}
\usage{
canonical_isoforms(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character vector
}
\description{
Search for a txt file at location:
file.path(refFolder(x), "canonical_isoforms.txt"), where x is an
ORFik experiment.
}
\examples{
df <- ORFik.template.experiment()
canonical_isoforms(df)
}
