% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{change_point_analysis}
\alias{change_point_analysis}
\title{Get the offset for specific ribo-seq read width}
\usage{
change_point_analysis(df, feature = "start")
}
\arguments{
\item{df}{a data.frame with points to analyse,
from function get_start_stop}

\item{feature}{a character, (start) or stop}
}
\value{
a single numeric offset
}
\description{
Get the offset for specific ribo-seq read width
}
