% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{changePointAnalysis}
\alias{changePointAnalysis}
\title{Get the offset for specific RiboSeq read width}
\usage{
changePointAnalysis(x, feature = "start")
}
\arguments{
\item{x}{a vector with points to analyse, assumes the zero is in the
middle + 1}

\item{feature}{(character) either "start" or "stop"}
}
\value{
a single numeric offset
}
\description{
Get the offset for specific RiboSeq read width
}
