% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{checkRFP}
\alias{checkRFP}
\title{Helper Function to check valid RFP input}
\usage{
checkRFP(class)
}
\arguments{
\item{class, }{the given class of RFP object}
}
\description{
Helper Function to check valid RFP input
}
