% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{checkRNA}
\alias{checkRNA}
\title{Helper Function to check valid RNA input}
\usage{
checkRNA(class)
}
\arguments{
\item{class, }{the given class of RNA object}
}
\description{
Helper Function to check valid RNA input
}
