% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{codon_dotplot}
\alias{codon_dotplot}
\title{Codon dwell time plot}
\usage{
codon_dotplot(
  dt,
  codon_score = "percentage",
  min_total_N_codons = ifelse(!is.null(attr(dt, "min_total_N_codons")), attr(dt,
    "min_total_N_codons"), 100L),
  add_plotly_tooltip = TRUE
)
}
\arguments{
\item{dt}{a data.table created from ORFik codon dwell time analysis}

\item{codon_score}{character, default columns to use in dt,
default "percentage" (i.e. relative_to_max_score)}

\item{min_total_N_codons}{= 100}

\item{add_plotly_tooltip}{logical, default TRUE.
If you code \code{plotly::ggplotly(plot, tooltip = "text")}, you will get a nice
tooltip by default.}
}
\value{
a data.table with additional attributes
}
\description{
Will create a fold change (ratio score) over your defined 'codon_score'.\cr
It then runs a wilcox p-value test on the differential dwell times of your scores.
It is set to signifcant = TRUE, if p-value < 0.05 & codon > min_total_N_codons &
(score > min_ratio_change | score < 1/min_ratio_change)
}
\details{
If a more rigerous model with estimated interaction terms between P and A sites is needed,
we advice RiboDT (see references).
}
\examples{
df <- ORFik.template.experiment()[9:10,] # Subset to 2 Ribo-seq libs
## For multiple libs
dt <- codon_usage_exp(df, outputLibs(df, type = "pshifted", output.mode = "list"),
                 min_counts_cds_filter = 10)
codon_dotplot(dt) # Sort on A-sites of RFP_r1.
}
\references{
https://doi.org/10.1016/j.ymeth.2021.10.004
}
\seealso{
Other codon: 
\code{\link{codon_usage}()},
\code{\link{codon_usage_exp}()},
\code{\link{codon_usage_plot}()},
\code{\link{diff_exp_codon_plot}()}
}
