% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{codon_usage_plot}
\alias{codon_usage_plot}
\title{Plot codon_usage}
\usage{
codon_usage_plot(
  res,
  score_column = res$relative_to_max_score,
  ylab = "Ribo-seq library",
  legend.position = "none",
  coord_flip = TRUE
)
}
\arguments{
\item{res}{a data.table of output from a codon_usage function}

\item{score_column}{numeric, default: res$relative_to_max_score.
Which parameter to use as score column.}

\item{ylab}{character vector, names for libraries to show on Y axis}

\item{legend.position}{character, default "none", do not display legend.}

\item{coord_flip}{logical, default TRUE. Flip so that seqs are y coordinates}
}
\value{
a ggplot object
}
\description{
Plot codon_usage
}
\examples{
df <- ORFik.template.experiment()[9:10,] # Subset to 2 Ribo-seq libs
## For multiple libs
res2 <- codon_usage_exp(df, outputLibs(df, type = "pshifted", output.mode = "list"),
                 min_counts_cds_filter = 10)
codon_usage_plot(res2)
}
\seealso{
Other codon: 
\code{\link{codon_usage_exp}()},
\code{\link{codon_usage}()}
}
\concept{codon}
