% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{codonSumsPerGroup}
\alias{codonSumsPerGroup}
\title{Get hits per codon}
\usage{
codonSumsPerGroup(h, indeces, L, N, reg_len, runLengths, countList)
}
\arguments{
\item{h}{indices per tuple}

\item{indeces}{whole list of indices}

\item{L}{Lengths}

\item{N}{hit sums}

\item{reg_len}{size of runs}

\item{runLengths}{integers per run}

\item{countList}{a Rle of count repetitions (000,1,00,1 etc)}
}
\value{
a list of codon sums
}
\description{
Helper for entropy function, normally not used directly
}
