% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{conditionNames}
\alias{conditionNames}
\title{Get condition name variants}
\usage{
conditionNames()
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: WT is main naming, but a variant is control
control will then be renamed to WT
}
\seealso{
Other experiment_naming: 
\code{\link{batchNames}()},
\code{\link{cellLineNames}()},
\code{\link{fractionNames}()},
\code{\link{inhibitorNames}()},
\code{\link{libNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{stageNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
