% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_directory_config.R
\name{config.exper}
\alias{config.exper}
\title{Set directories for experiment}
\usage{
config.exper(experiment, assembly, type, config = ORFik::config())
}
\arguments{
\item{experiment}{short name of experiment (must be valid as a folder name)}

\item{assembly}{name of organism and assembly (must be valid as a folder name)}

\item{type}{name of sequencing type, Ribo-seq, RNA-seq, CAGE.. Can be more than
one.}

\item{config}{a named character vector of length 3,
default: \code{ORFik::config()}}
}
\value{
named character vector of paths for experiment
}
\description{
Defines a folder for:
1. fastq files (raw_data)\cr
2. bam files (processed data)\cr
3. references (organism annotation and STAR index)\cr
4. Experiment (name of experiment)
}
\examples{
## Save to default config location
#config.exper("Alexaki_Human", "Homo_sapiens_GRCh38_101", c("Ribo-seq", "RNA-seq"))
}
