% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{convert_to_covRle}
\alias{convert_to_covRle}
\title{Convert libraries to coverage RLEs}
\usage{
convert_to_covRle(
  df,
  in_files = filepath(df, "pshifted"),
  split.by.strand = TRUE,
  seq_info = seqinfo(df),
  weight = "score"
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{in_files}{paths to input files, default pshifted files:
\code{filepath(df, "pshifted")} in ofst format}

\item{split.by.strand}{logical, default TRUE, split into forward and reverse
strand RleList inside covRle object.}

\item{weight}{integer, numeric or single length character. Default "score".
Use score column in loaded in_files.}

\item{seqinfo}{SeqInfo object, default \code{seqinfo(findFa(df))}}
}
\value{
invisible(NULL), files saved to disc
}
\description{
Saved in folder "cov_RLE" relative to default libraries of experiment
}
