% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertToOneBasedRanges}
\alias{convertToOneBasedRanges}
\title{Convert a GRanges Object to 1 width reads}
\usage{
convertToOneBasedRanges(gr, method = "5prime", addScoreColumn = FALSE)
}
\arguments{
\item{gr}{GRanges Object to reduce}

\item{method}{the method to reduce, see info. (5prime defualt)}

\item{addScoreColumn}{logical (FALSE), if TRUE, add a score column that
sums up the hits per position.}
}
\value{
Converted GRanges object
}
\description{
There are 3 ways of doing this
1. Take 5' ends, reduce away rest (5prime)
2. Tile and include all (tileAll)
3. Take middle point per GRanges (middle)
}
\seealso{
Other utils: \code{\link{bedToGR}}, \code{\link{findFa}},
  \code{\link{fread.bed}}, \code{\link{is.gr_or_grl}},
  \code{\link{is.grl}}, \code{\link{validGRL}}
}
\concept{utils}
