% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plots.R
\name{correlation.plots}
\alias{correlation.plots}
\title{Correlation plots between all samples}
\usage{
correlation.plots(df, output.dir, region = "mrna", type = "fpkm")
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{output.dir}{directory to save to, 2 files named: cor_plot.png and
cor_plot_log2.png}

\item{region}{a character (default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{type}{which value to use, "fpkm", alternative "counts".}
}
\value{
invisible(NULL)
}
\description{
Correlation plots between all samples
}
