% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{countTable}
\alias{countTable}
\title{Extract count table directly from experiment}
\usage{
countTable(df, region = "mrna", type = "count")
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{region}{a character vector (default: "mrna"), make raw count matrices
of whole mrnas or one of (leaders, cds, trailers). Can also be a}

\item{type}{default: "count" (raw counts matrix), alternative is "fpkm",
"log2fpkm" or "log10fpkm"}
}
\value{
a DEseq summerizedExperiment object (count)
 or matrix (if fpkm input)
}
\description{
Extracts by getting /QC_STATS directory, and searching for region
}
