% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{coverageByTranscriptC}
\alias{coverageByTranscriptC}
\title{coverageByTranscript with coverage input}
\usage{
coverageByTranscriptC(x, transcripts, ignore.strand = !strandMode(x))
}
\arguments{
\item{x}{a covRle (one RleList for each strand in object),
 must have defined and correct
seqlengths in its SeqInfo object.}

\item{transcripts}{\code{\link{GRangesList}}}

\item{ignore.strand}{a logical (default: length(x) == 1)}
}
\value{
Integer Rle of coverage, 1 per transcript
}
\description{
Extends the function with direct genome coverage input,
see \code{\link{coverageByTranscript}} for original function.
}
