% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{coverageByTranscriptW}
\alias{coverageByTranscriptW}
\title{Remake of coverageByTranscript}
\usage{
coverageByTranscriptW(x, transcripts, ignore.strand = FALSE, weight = 1L)
}
\arguments{
\item{x}{reads (\code{\link{GRanges}}, GAlignment)}

\item{transcripts}{\code{\link{GRangesList}}}

\item{ignore.strand}{a logical (default: FALSE)}

\item{weight}{a vector (default: 1L), if single number applies for all,
else equal size as x}
}
\value{
Integer Rle of coverage, 1 per transcript
}
\description{
Allows weights, see \code{\link{coverageByTranscript}}
}
