% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{coverageGroupings}
\alias{coverageGroupings}
\title{Get grouping for a coverage table in ORFik}
\usage{
coverageGroupings(logicals, grouping = "GF")
}
\arguments{
\item{logicals}{size 2 logical vector, the is.null checks for each column,}

\item{grouping}{which grouping to perform}
}
\value{
a quote of the grouping to pass to data.table
}
\description{
Either of two groupings:
GF: Gene, fraction
FGF: Fraction, position, feature
It finds which of these exists, and auto groups
}
\details{
Normally not used directly!
}
