% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{coverageHeatMap}
\alias{coverageHeatMap}
\title{Create a heatmap of coverage}
\usage{
coverageHeatMap(coverage, output = NULL, scoring = "zscore")
}
\arguments{
\item{coverage}{a data.table of coverage with columns position, score and
fraction}

\item{output}{character string (NULL), if set, saves the plot as pdf
to path given.}

\item{scoring}{character vector (zscore), either of zScore,
transcriptNormalized, sum, mean, which score was used to create coverage ?}
}
\value{
a ggplot object of the coverage plot, NULL if output is set,
then the plot will only be saved to location.
}
\description{
Coverage rows in heat map is fraction
Coverage column in heat map is score, default zscore of counts
}
\details{
See vignette for example
}
\seealso{
Other coveragePlot: \code{\link{pSitePlot}},
  \code{\link{savePlot}}, \code{\link{windowCoveragePlot}}
}
\concept{coveragePlot}
