% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\docType{class}
\name{covRle-class}
\alias{covRle-class}
\title{Coverage Rle for both strands or single}
\description{
Given a run of coverage(x) where x are reads,
this class combines the 2 strands into 1 object
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment.zf}()},
\code{\link{ORFik.template.experiment}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism,experiment-method}},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
