% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{covRle}
\alias{covRle}
\title{Coverage Rlelist for both strands}
\usage{
covRle(forward, reverse = RleList())
}
\description{
Coverage Rlelist for both strands
}
\examples{
covRle(RleList(), RleList())
}
