% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\docType{class}
\name{covRleList-class}
\alias{covRleList-class}
\title{List of covRle}
\value{
a covRleList object
}
\description{
Given a run of coverage(x) where x are reads,
this covRle combines the 2 strands into 1 object
This list can again combine these into 1 object, with accession functions
and generalizations.
}
\seealso{
Other covRLE: 
\code{\link{covRle}},
\code{\link{covRle-class}},
\code{\link{covRleFromGR}()},
\code{\link{covRleList}}
}
\concept{covRLE}
