% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{covRleList}
\alias{covRleList}
\title{Coverage Rlelist for both strands}
\usage{
covRleList(list, fraction = names(list))
}
\arguments{
\item{list}{a list or List of covRle objects of equal length and lengths}

\item{fraction}{character, default \code{names(list)}.
Names to elements of list, can be integers, as readlengths etc.}
}
\value{
a covRleList object
}
\description{
Coverage Rlelist for both strands
}
\examples{
covRleList(List(covRle()))
}
\seealso{
Other covRLE: 
\code{\link{covRle}},
\code{\link{covRle-class}},
\code{\link{covRleFromGR}()},
\code{\link{covRleList-class}}
}
\concept{covRLE}
