% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{create_time_txdb}
\alias{create_time_txdb}
\title{Load creation time of txdb safely}
\usage{
create_time_txdb(txdb_path)
}
\value{
a character string of time
}
\description{
Load creation time of txdb safely
}
\keyword{internal}
