% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defining_ORFs_helpers.R
\name{define_trailer}
\alias{define_trailer}
\title{Creates GRanges object as trailer in ORF.}
\usage{
define_trailer(ORFranges, transcriptRanges, lengthOftrailer = 10)
}
\arguments{
\item{ORFranges}{GRanges object of your Open Reading Frame.}

\item{transcriptRanges}{GRanges object of transtript.}

\item{lengthOftrailer}{Numeric. Default is 10.}
}
\value{
A GRanges object of trailer.
}
\description{
Creates GRanges object as trailer in ORF.
}
\examples{
#' #defineTrailer()
}

