% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{DEG_gorilla}
\alias{DEG_gorilla}
\title{GO analysis with GOrilla}
\usage{
DEG_gorilla(dt, output_dir)
}
\arguments{
\item{dt}{a data.table of DEG or DTEG results, must also have appended
gene symbols column called "external_gene_name"}

\item{output_dir}{path to save results}
}
\value{
a data.table with urls per contrast, this is also saved in
output_dir
}
\description{
Per contrast, split list into regulation groups and run GOrilla go analysis.
If you want to change LFC threshold or p-value cutoff, you reassign
Regulation column by LFC and/or p-value.
}
