% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{DEG_gorilla_local_load_data}
\alias{DEG_gorilla_local_load_data}
\title{Load all GOrilla xls files from study}
\usage{
DEG_gorilla_local_load_data(gorilla_local_dir)
}
\arguments{
\item{gorilla_local_dir}{path to directory of local GORilla html and xls data}
}
\value{
a data.table
}
\description{
Output as data.table with column 'analysis' describing DEG
contrast and column 'go_category' describing COMPONENT, FUNCTION or PROCESS.
}
