% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_analysis.R
\name{DEG_gorilla_plot}
\alias{DEG_gorilla_plot}
\title{Plot GOrilla xls results}
\usage{
DEG_gorilla_plot(dt, top_n = 20L, enrich_cutoff = 8)
}
\arguments{
\item{dt}{a data.table with results loaded using
\code{DEG_gorilla_local_load_data}}

\item{top_n}{maximum number of GO terms per analysis / component split. The
sorting is done as -GeneRatio, -p.adjust, so it extracts the top 20
highest GeneRatio terms by p.adjust values (decreasing sort).}

\item{enrich_cutoff}{numeric, default 8. Minimum enrichment, set it lower
to get more generic groups, higher to get more specific groups.}
}
\value{
a ggplot, it uses facet_wrap to split analysis / components.
}
\description{
Inspired by the enrichment plot from the package clusterProfiler.
}
