% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dif_expression_helpers.R
\name{DEG_model_results}
\alias{DEG_model_results}
\title{Get DESeq2 model results from model}
\usage{
DEG_model_results(ddsMat_rna, target.contrast, pairs, p.value = 0.05)
}
\arguments{
\item{ddsMat_rna}{a DESeqDataSet object with results stored as metadata columns.}

\item{target.contrast}{a character vector, default \code{design[1]}.
The column in the ORFik experiment that represent the comparison contrasts.
By default: the first design factor of the full experimental design.
This is the factor you will do the comparison on. DESeq will normalize
the counts based on the full design, but the log fold change values will
be based on this contrast only. It is usually the 'condition' column.}

\item{pairs}{list of character pairs, the experiment contrasts. Default:
\code{combn.pairs(unlist(df.rfp[, target.contrast])}}

\item{p.value}{a numeric, default 0.05 in interval (0,1)
or "" to not show.
What p-value used for the analysis? Will be shown as a caption.}
}
\value{
a data.table
}
\description{
Get DESeq2 model results from model
}
\examples{
## Simple example (use ORFik template, then use only RNA-seq)
df <- ORFik.template.experiment()
df.rna <- df[df$libtype == "RNA",]
design(df.rna) # The full experimental design
target.contrast <- design(df.rna)[1] # Default target contrast
#ddsMat_rna <- DEG_model(df.rna, target.contrast)
#pairs <- combn.pairs(unlist(df[, target.contrast]))
#dt <- DEG_model_results(ddsMat_rna, target.contrast, pairs)
}
