% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_drive}
\alias{detect_drive}
\title{Detects the mounted drive based on a mounted path}
\usage{
detect_drive(ref_path = path.expand(config()["ref"]))
}
\arguments{
\item{ref_path}{= path.expand(config()["ref"])}
}
\value{
character, name of FileSystem drive of mounted path,
NA_character_ if not found
}
\description{
Detects the mounted drive based on a mounted path
}
\examples{
detect_drive(tempdir())
}
