% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints.R
\name{detectRibosomeShifts}
\alias{detectRibosomeShifts}
\title{Shift ribosome footprints}
\usage{
detectRibosomeShifts(footprints, txdb, start = TRUE, stop = FALSE,
  offset_plots = FALSE, top_tx = 10)
}
\arguments{
\item{footprints}{a bam or bed file path, or a GAlignment object
of ribo-seq reads}

\item{txdb}{a txdb object from a gtf file}

\item{start}{logical (T), include starts}

\item{stop}{logical (F), include stops}

\item{offset_plots}{logical (F), plot shifting for validation}

\item{top_tx}{numeric(10), which top percentage of transcripts to use.}
}
\value{
a data.frame with shift lengths and offsets
}
\description{
Given a TxDb object and a bam file path, get the predicted shifts
of the footprints to the p-site.
}
\examples{
gtf <- system.file("extdata", "example.gtf",
        package = "ORFik") ## location of the gtf file
suppressWarnings(txdb <-
 GenomicFeatures::makeTxDbFromGFF(gtf, format = "gtf"))
bam <- system.file("extdata", "example.bam",
        package = "ORFik") ## location of the bam file
detectRibosomeShifts(bam, txdb, start=TRUE, stop=FALSE,
                     offset_plots=FALSE, top_tx=10)

}
