% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{diff_exp_codon}
\alias{diff_exp_codon}
\title{Differential codon dwell time analysis}
\usage{
diff_exp_codon(
  dt,
  min_ratio_change = 1.7,
  min_total_N_codons = 100,
  exclude_start_stop = FALSE,
  codon_score = "percentage",
  background = NULL
)
}
\arguments{
\item{dt}{}
}
\description{
Differential codon dwell time analysis
}
