% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{diff_exp_codon_plot}
\alias{diff_exp_codon_plot}
\title{Differential codon dwell time plot}
\usage{
diff_exp_codon_plot(
  dt,
  min_total_N_codons = ifelse(!is.null(attr(dt, "min_total_N_codons")), attr(dt,
    "min_total_N_codons"), 100L),
  only_significant_difexp = FALSE,
  add_plotly_tooltip = TRUE
)
}
\arguments{
\item{dt}{a data.table created from ORFik codon dwell time analysis}

\item{min_total_N_codons}{integer, default
ifelse(!is.null(attr(dt, "min_total_N_codons")), attr(dt, "min_total_N_codons"), 100L).
How many codons are minimum to not color it red (i.e. "the not enough codons" group).}

\item{only_significant_difexp}{logical, default FALSE. If TRUE, don't show non
significant codons.}

\item{add_plotly_tooltip}{logical, default TRUE.
If you code \code{plotly::ggplotly(plot, tooltip = "text")}, you will get a nice
tooltip by default.}
}
\value{
a ggplot object
}
\description{
Differential codon dwell time plot
}
