% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{disengagementScore}
\alias{disengagementScore}
\title{Disengagement score (DS)}
\usage{
disengagementScore(grl, RFP, GtfOrTx)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{GtfOrTx}{if Gtf: a TxDb object of a gtf file,
if tx: a GrangesList of transcripts, called from:
exonsBy(Gtf, by = "tx", use.names = T)}
}
\value{
a named vector of numeric values of scores
}
\description{
is defined as (RPFs over ORF)/(RPFs downstream to tx end).
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1242/dev.098344
}
\examples{
ORF <- GRanges(seqnames = "1",
ranges = IRanges(start = c(1, 10, 20),
                end = c(5, 15, 25)),
strand = "+")
grl <- GRangesList(tx1_1 = ORF)
tx <- GRangesList(tx1 = GRanges("1", IRanges(1, 50),"+"))
RFP <- GRanges("1",
ranges = IRanges(start = c(1, 25, 50),
                end = c(1, 25, 50)),
strand = "+")
disengagementScore(grl, RFP, tx)
}
\seealso{
Other features: \code{\link{ORFDistToCds}},
  \code{\link{ORFScores}}, \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{RibosomeStallingScore}},
  \code{\link{computeFeatures}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subsetCoverage}}, \code{\link{te}}
}
