% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_extend.R
\name{distanceToFollowing}
\alias{distanceToFollowing}
\title{Distance to following range group}
\usage{
distanceToFollowing(grl, grl2 = grl, ignore.strand = FALSE)
}
\arguments{
\item{grl}{a GRangesList}

\item{grl2}{a GRangesList, default 'grl'. The list that defines
restrictions on extension. Can also be another set, which is
used as 'roadblocks' for extension.}

\item{ignore.strand}{logical, default FALSE}
}
\value{
numeric vector of distance
}
\description{
Follow means downstream GRangesList element, not including itself.
}
