% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{distOrfToCds}
\alias{distOrfToCds}
\title{get distances between orf ends and starts of transcripts cds' belonging to orfs.
matching is done by transcript names, so they must match.}
\usage{
distOrfToCds(ORFs, cds, fiveUTRs)
}
\arguments{
\item{ORFs}{orfs as GRangesList}

\item{cds}{cds as GRangesList}

\item{fiveUTRs}{fiveUTRs as GRangesList}
}
\description{
get distances between orf ends and starts of transcripts cds' belonging to orfs.
matching is done by transcript names, so they must match.
}
