% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{distToCds}
\alias{distToCds}
\title{Get distances between orf ends and starts of transcripts cds' belonging to orfs.}
\usage{
distToCds(ORFs, fiveUTRs, cds = NULL, extension = NULL)
}
\arguments{
\item{ORFs}{orfs as \code{\link[GenomicRanges]{GRangesList}},
names of orfs must be transcript names}

\item{fiveUTRs}{fiveUTRs as \code{\link[GenomicRanges]{GRangesList}},
must be original unchanged fiveUTRs}

\item{cds}{cds' as \code{\link[GenomicRanges]{GRangesList}},
only add if you used cage extension}

\item{extension}{needs to be set! set to 0 if you did not use cage
if you used cage to change tss' when finding the orfs, standard cage
extension is 1000}
}
\value{
an integer vector, +1 means one base upstream of cds, -1 means
  2nd base in cds, 0 means orf stops at cds start.
}
\description{
Matching is done by transcript names.
fiveUTRs must be used to make transcript positions possible.
The cds start site, will be presumed to be on + 1 of end of fiveUTRs
See article:  10.1074/jbc.R116.733899
}
\examples{
grl <- GRangesList(tx1_1 = GRanges("1", IRanges(1,10), "+"))
fiveUTRs <- GRangesList(tx1 = GRanges("1", IRanges(1,20), "+"))
dist <- distToCds(grl, fiveUTRs, extension = 0)
## Now dist contains distances as vector
}
\seealso{
Other features: \code{\link{computeFeatures}},
  \code{\link{disengagementScore}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
