% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{download.SRA.metadata}
\alias{download.SRA.metadata}
\title{Downloads metadata from SRA}
\usage{
download.SRA.metadata(SRP, outdir, remove.invalid = TRUE, auto.detect = FALSE)
}
\arguments{
\item{SRP}{a string, a study ID as either the SRP, ERP, DRP, PRJ or GSE of the study,
examples would be "SRP226389" or "ERP116106". If GSE it will try to convert to the SRP
to find the files.}

\item{outdir}{directory to save file,
The file will be called "SraRunInfo_SRP.csv", where SRP is
the SRP argument.
The directory will be created if not existing.}

\item{remove.invalid}{logical, default TRUE. Remove Runs with 0 reads (spots)}

\item{auto.detect}{logical, default FALSE. If TRUE, ORFik will add additional columns:\cr
LIBRARYTYPE: (is this Ribo-seq or mRNA-seq, CAGE etc), \cr
REPLICATE: (is this replicate 1, 2 etc),\cr
STAGE: (Which time point, cell line or tissue is this, HEK293, TCP-1, 24hpf etc),\cr
CONDITION: (is this Wild type control or a mutant etc).\cr
These values are only qualified guesses from the metadata, so always double check!}
}
\value{
a data.table of the metadata, 1 row per sample,
 SRR run number defined in Run column.
}
\description{
Given a experiment identifier, query information from different locations of SRA
to get a complete metadata table of the experiment.
}
\examples{
## Originally on SRA
outdir <- tempdir() # Specify output directory
# download.SRA.metadata("SRP226389", outdir)
## Originally on ENA
# download.SRA.metadata("ERP116106", outdir)
## Originally on GEO (GSE)
# download.SRA.metadata("GSE61011", outdir)
}
\references{
doi: 10.1093/nar/gkq1019
}
\seealso{
Other sra: 
\code{\link{download.SRA}()},
\code{\link{download.ebi}()},
\code{\link{install.sratoolkit}()},
\code{\link{rename.SRA.files}()}
}
\concept{sra}
