% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_metadata.R
\name{download_gene_homologues}
\alias{download_gene_homologues}
\title{Download homologue information of a gene}
\usage{
download_gene_homologues(
  gene_id = "ENSG00000110092",
  organism = "Homo sapiens"
)
}
\arguments{
\item{organism, }{default NULL. Scientific name (e.g. Homo sapiens)}

\item{gene}{character, gene name (ensembl gene id, not symbol!)}
}
\value{
character, summary text for gene from the database.
}
\description{
Uses ncbi gene database for vertebrates
}
