% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_metadata.R
\name{download_gene_info}
\alias{download_gene_info}
\title{Download summary information of a gene}
\usage{
download_gene_info(gene = "CCND1", organism = "Homo sapiens")
}
\arguments{
\item{gene}{character, gene name (symbol)}

\item{organism, }{default NULL. Scientific name (e.g. Homo sapiens)}
}
\value{
character, summary text for gene from the database.
}
\description{
Uses ncbi gene database summary from RefSeq
}
