% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{downstreamN}
\alias{downstreamN}
\title{Restrict GRangesList}
\usage{
downstreamN(grl, firstN = 150L)
}
\arguments{
\item{grl}{(GRangesList)}

\item{firstN}{(integer) Allow only this many bp downstream, maximum.}
}
\value{
a GRangesList of reads restricted to firstN and tiled by 1
}
\description{
Will restrict GRangesList to `N` bp downstream from the first base.
}
\keyword{internal}
