% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{downstreamOfPerGroup}
\alias{downstreamOfPerGroup}
\title{per group get the part downstream of position defined in downstreamOf}
\usage{
downstreamOfPerGroup(tx, downstreamOf)
}
\arguments{
\item{tx}{a GRangesList, usually of Transcripts to be changed}

\item{downstreamOf}{a vector of integers, for each group in tx, where
is the new start point of first valid exon.}
}
\description{
per group get the part downstream of position defined in downstreamOf
}
