% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{end,GRanges-method}
\alias{end,GRanges-method}
\title{Get ends of GRanges}
\usage{
\S4method{end}{GRanges}(x)
}
\arguments{
\item{x}{a GRanges}
}
\value{
an integer (length equal to x)
}
\description{
Faster version than S4Vector generic caller
}
