% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{entropy}
\alias{entropy}
\title{Calucalte Entropy value of input reads overlapping}
\usage{
entropy(grl, reads)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} that the reads can map to}

\item{reads}{a GAlignment object, ig. from ribo seq, or rna seq,
can also be GRanges or GRangesList}
}
\value{
A numeric vector containing entropy per ORF
}
\description{
For example reads per orf.
The entropy interval per group is a real number in the interval (0:1)
Where 0 is no variance in reads over group.
}
\examples{
ORF <- GRanges("1",
ranges = IRanges(start = c(1, 12, 22),
                end = c(10, 20, 32)),
strand = "+")
ORF$names = "tx1_1"
names(ORF) <- rep("tx1", 3)
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(c(25, 35), c(25, 35)),"+")
# NB: grl must have same names as cds, + _1 etc
entropy(grl, RFP)
# or on cds
cdsORF <- GRanges("1", IRanges(35, 44),"+")
cdsORF$names = "tx1"
names(cdsORF) <- "tx1"
cds <-  GRangesList(tx1 = cdsORF)
entropy(cds, RFP)
}
\seealso{
Other features: \code{\link{computeFeaturesSpecial}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
