% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{envExp,experiment-method}
\alias{envExp,experiment-method}
\title{Get ORFik experiment environment}
\usage{
\S4method{envExp}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
environment pointer, name of environment: pointer
}
\description{
More correctly, get the pointer reference, default is .GlobalEnv
}
