% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{envExp<-,experiment-method}
\alias{envExp<-,experiment-method}
\title{Set ORFik experiment environment}
\usage{
\S4method{envExp}{experiment}(x) <- value
}
\arguments{
\item{x}{environment pointer, name of environment: pointer}
}
\value{
an ORFik \code{\link{experiment}}
}
\description{
More correctly, set the pointer reference, default is .GlobalEnv
}
