% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{envExp<-}
\alias{envExp<-}
\title{Set ORFik experiment environment}
\usage{
envExp(x) <- value
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}

\item{value}{environment pointer to assign to experiment}
}
\value{
an ORFik \code{\link{experiment}} with updated environment
}
\description{
More correctly, set the pointer reference, default is .GlobalEnv
}
