% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{envExp}
\alias{envExp}
\title{Get ORFik experiment environment}
\usage{
envExp(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
environment pointer, name of environment: pointer
}
\description{
More correctly, get the pointer reference, default is .GlobalEnv
}
