% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exists.ftp.file.fast}
\alias{exists.ftp.file.fast}
\title{A copy of biomartr ftp check}
\usage{
exists.ftp.file.fast(url, file.path)
}
\arguments{
\item{url}{character, full path directory of url}

\item{file.path}{character, full path url to file}
}
\value{
logical, TRUE if file exists
}
\description{
Will be removed when biomartr::exists.ftp.file.new
 is pushed to CRAN stable
}
