% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exports.R
\name{export.bed12}
\alias{export.bed12}
\title{Export grouped exon structures in BED12 format.}
\usage{
export.bed12(grl, file, rgb = 0)
}
\arguments{
\item{grl}{A `GRangesList`.}

\item{file}{Output file path.}

\item{rgb}{Integer vector, default `0`, maximum `255`. Either a single value
or a vector of length `length(grl)` specifying the display color per group.
In IGV/UCSC, `0` is black and `255` is blue.}
}
\value{
`invisible(NULL)`. The function writes the file and returns
  invisibly.
}
\description{
This is useful as a compact representation of grouped ORFs or other
transcript-like features, and the output can be loaded directly in UCSC
Genome Browser or IGV.
}
\details{
If `grl` has no names, groups will be named `1, 2, 3, ...`.
}
\examples{
grl <- GRangesList(GRanges("1", c(1,3,5), "+"))
}
\references{
\url{https://bedtools.readthedocs.io/en/latest/content/general-usage.html#bed-format}
}
\seealso{
Other utils: 
\code{\link{bedToGR}()},
\code{\link{convertToOneBasedRanges}()},
\code{\link{export.bigWig}()},
\code{\link{export.fstwig}()},
\code{\link{export.wiggle}()},
\code{\link{fimport}()},
\code{\link{findFa}()},
\code{\link{fread.bed}()},
\code{\link{optimizeReads}()},
\code{\link{readBam}()},
\code{\link{readBigWig}()},
\code{\link{readWig}()}
}
\concept{utils}
