% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_helpers.R
\name{extract_START_sites_from_CDSs}
\alias{extract_START_sites_from_CDSs}
\title{Creates GRanges object of Transcription Start Sites from CDS.}
\usage{
extract_START_sites_from_CDSs(cds)
}
\arguments{
\item{cds}{GRanges List object of your CDSs.}
}
\value{
A GRanges object of TSS sites.
}
\description{
It will return A postions of ATG transcription start sites.
}
\examples{
cds <- GRangesList('gene_plus_strand' = GRanges(Rle(c('1'), c(4)),
                                                IRanges(c(925942, 930155, 939040, 939275),
                                                        width=c(72, 182, 90, 17)),
                                                Rle(strand(c('+', '+', '+', '+')))),
                   'gene_minus_strand' = GRanges(Rle(c('1'), c(4)),
                                                 IRanges(c(245929870, 245927931,
                                                           245863799, 245858562),
                                                         width=c(32, 103, 88, 109)),
                                                 Rle(strand(c('-', '-', '-', '-')))))
extract_START_sites_from_CDSs(cds)

}

