% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imports.R
\name{file_ext_without_compression}
\alias{file_ext_without_compression}
\title{Get file extension of files without compressions}
\usage{
file_ext_without_compression(x, compressions = c("gzip", "gz", "bgz", "zip"))
}
\arguments{
\item{x}{character paths}

\item{compressions}{character vector, default: c("gzip", "gz", "bgz", "zip").
Expand if you have other formats}
}
\value{
character vector of file extensions of paths excluding suffix vector defined in compression.
}
\description{
Get file extension of files without compressions
}
\examples{
file_ext_without_compression(c("abc.bam.gz", "def.bam.zip"))
}
