% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{filterCage}
\alias{filterCage}
\title{Filter peak of cage-data by value}
\usage{
filterCage(cage, filterValue = 1, fiveUTRs = NULL)
}
\arguments{
\item{cage}{The raw cage-data, as GRanges. Must contain a score column,
with the count hits per position.}

\item{filterValue}{The integer of counts(score) to filter on
for a tss to pass as hit}

\item{fiveUTRs}{a GRangesList (NULL), if added will filter out cage reads by
these following rules:
all reads in region (-5:-1, 1:5) for each tss will be removed, removes noise.}
}
\value{
the filtered Granges object
}
\description{
Filter peak of cage-data by value
}
