% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_filters.R
\name{filterExtremePeakGenes}
\alias{filterExtremePeakGenes}
\title{Filter out transcript by a median filter}
\usage{
filterExtremePeakGenes(
  tx,
  reads,
  upstream = NULL,
  downstream = NULL,
  multiplier = "0.99",
  min_cutoff = "0.999",
  pre_filter_minimum = 0,
  average = "median"
)
}
\arguments{
\item{tx}{a GRangesList}

\item{reads}{a GAlignments or GRanges}

\item{multiplier}{a character or numeric, default "0.99",
either a quantile if input is string[0-1],
like "0.99", or numeric value if input is numeric.
How much bigger than median / mean counts per gene,
must a value be to be defined as extreme ?}

\item{min_cutoff}{a character or numeric, default "0.999",
either a quantile if input is string[0-1],
like "0.999", or numeric value if input is numeric. Lowest allowed value}

\item{pre_filter_minimum}{numeric, default 0. If value is x,
will remove all positions in all genes with coverage < x,
before median filter is applied. Set to 1 to remove all 0 positions.}

\item{average}{character, default "median". Alternative: "mean".
How to scale the multiplier argument, from median or mean of gene coverage.}
}
\value{
GRangesList (filtered)
}
\description{
For removing extreme peaks in coverage plots, use high quantiles, like
99. If you want to find general stall sites of Ribo-seq, you should use a
pre_filter minimum and
}
