% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterTranscripts}
\alias{filterTranscripts}
\title{Get the transcripts that have minimum lengths of leaders, cds and trailer.}
\usage{
filterTranscripts(txdb, minFiveUTR = 30L, minCDS = 150L,
  minThreeUTR = 30L, stopOnEmpty = TRUE)
}
\arguments{
\item{txdb}{a TxDb object from gtf}

\item{minFiveUTR}{(integer) minimum bp for 5' UTR during filtering for the
transcripts}

\item{minCDS}{(integer) minimum bp for CDS during filtering for the
transcripts}

\item{minThreeUTR}{(integer) minimum bp for 3' UTR during filtering for the
transcripts}

\item{stopOnEmpty}{logical TRUE, stop if no valid names are found ?}
}
\value{
a character vector of valid tramscript names
}
\description{
Filter transcripts to those who have 5' UTR, CDS, 3' UTR of some lengths,
pick the longest per gene.
}
\details{
If a transcript does not have a 3' UTR, then the length is 0,
so they will be filtered out. So only transcripts with leaders, cds and
3' UTRs will be returned. You can set the integer to 0, that will return all
within that group.
}
\examples{
gtf_file <- system.file("extdata", "annotations.gtf", package = "ORFik")
txdb <- GenomicFeatures::makeTxDbFromGFF(gtf_file, format = "gtf")
txNames <- filterTranscripts(txdb)

}
