% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{find_in_frame_ORFs}
\alias{find_in_frame_ORFs}
\title{Creates GRangesList of Open Reading Frames mapped to genomic coordinates
Input is a Grangeslist of regions to search, together with a DNAStringSet with
fastaSequence in same order as the grl.}
\usage{
find_in_frame_ORFs(grl, fastaSeqs, startCodon = "ATG",
  stopCodon = "TAA|TAG|TGA", longestORF = F, minimumLength = 0)
}
\arguments{
\item{grl}{GRangesList of sequences to search for orfs, in Genomic coordinates}

\item{fastaSeqs}{DNA sequences to search for Open Reading Frames, must be DNAStringSet.}

\item{startCodon}{string. Default is "ATG".}

\item{stopCodon}{string. Default is "TAA|TAG|TGA".}

\item{longestORF}{bolean. Default FALSE. Defines whether pick longest ORF only.
When FALSE will report all open reaidng frames, even overlapping small ones.}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs to be at least
 START + 8*3 [bp] + STOP.}
}
\value{
A GRangesList of ORFs.
}
\description{
Creates GRangesList of Open Reading Frames mapped to genomic coordinates
Input is a Grangeslist of regions to search, together with a DNAStringSet with
fastaSequence in same order as the grl.
}
