% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{find_ORFs}
\alias{find_ORFs}
\title{Creates IRanges with Open Reading Frames from input sequences.}
\usage{
find_ORFs(seq, seqname = "unknown", startCodon = NULL, strand = "*",
  numcode = 1, longestORF = T, minimumLength = 0)
}
\arguments{
\item{seq}{string. Input sequence.}

\item{seqname}{string. Name of the sequence, meybe name of the transcript or name of the chromosome.
Final GRanges will use this name as seqname.}

\item{startCodon}{string. Default is NULL. This is gonna overwrite START codons used in numcode. eg. "ATG|GTG".}

\item{strand}{string. One of "*" which is default, "+" or "-"}

\item{numcode}{numeric.
The \url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.}

\item{longestORF}{bolean. Default TRUE. Defines whether pick longest ORF only.
When FALSE will report all open reaidng frames, even overlapping, small ones.}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs to be at least START + 8*3 [bp] + STOP.}
}
\value{
A List of GRanges of ORFs mapped to fasta file. Each ORF includes START and STOP codon,
frame, seqname, and strand.
}
\description{
Creates IRanges with Open Reading Frames from input sequences.
}
\examples{
#find_ORFs()

}

