% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs_fa.R
\name{find_ORFs_fa}
\alias{find_ORFs_fa}
\title{Creates GRanges with Open Reading Frames from fasta files.}
\usage{
find_ORFs_fa(file, startCodon = c("ATG"), stopCodon = c("TAA", "TAG",
  "TGA"), longestORF = T, minimumLength = 0)
}
\arguments{
\item{file}{- Path to fasta file.}

\item{startCodon}{Default is c('ATG').}

\item{stopCodon}{Default is c('TAA', 'TAG', 'TGA').}

\item{longestORF}{bolean. Default TRUE. Defines whether pick longest ORF only.
When FALSE will report all open reaidng frames, even overlapping, small ones.}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs to be at least START + 8*3 [bp] + STOP.}
}
\value{
A List of GRanges of ORFs mapped to fasta file. Each ORF includes START and STOP codon,
frame, seqname, and strand.
}
\description{
Each fasta header is treated separately, and name of the sequence will be used as seqname in
returned GRanges object. Frame of the Open Reading Frame is also returned in
metadata column 'frame'.
}
\examples{
#find_ORFs_fa()
}

