% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{find_url_ebi}
\alias{find_url_ebi}
\title{Locates and check if fastq files exists in ebi}
\usage{
find_url_ebi(SRR, stop.on.error = FALSE)
}
\arguments{
\item{SRR}{character, SRR, ERR or DRR numbers.}

\item{stop.on.error}{logical FALSE, if true will stop
if all files are not found.}
}
\value{
full url to fastq files, same length as input
(2 urls for paired end data). Returns empty character() if all
files not found.
}
\description{
Look for files in ebi following url: ftp://ftp.sra.ebi.ac.uk/vol1/fastq
Paired end and single end fastq files
}
