% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFa}
\alias{findFa}
\title{Convenience wrapper for Rsamtools FaFile}
\usage{
findFa(faFile)
}
\arguments{
\item{faFile}{a character path or FaFile}
}
\value{
a FaFile or BSgenome
}
\description{
Convenience wrapper for Rsamtools FaFile
}
