% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{findLibrariesInFolder}
\alias{findLibrariesInFolder}
\title{Get all library files in folder}
\usage{
findLibrariesInFolder(dir, types)
}
\arguments{
\item{dir}{The directory to find bam, bed, wig files.}

\item{types}{All accepted types of bam, bed, wig files..}
}
\value{
(character vector) All files found from types parameter.
}
\description{
Get all library files in folder
}
