% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_filters.R
\name{findMaxPeakPerGenes}
\alias{findMaxPeakPerGenes}
\title{Find max peak per gene}
\usage{
findMaxPeakPerGenes(tx, reads, top_tx = 0.5, min_reads_per_tx = 10)
}
\arguments{
\item{tx}{a GRangesList}

\item{reads}{a GAlignments or GRanges}

\item{top_tx}{numeric, default 0.50 (50\% top tx by reads).}

\item{min_reads_per_tx}{numeric, default 10. Gene must have at least
10 reads}
}
\value{
a data.table
}
\description{
For finding the highest peak per gene. With some default filters.
}
