% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findNGSPairs}
\alias{findNGSPairs}
\title{Find pair of forward and reverse strand wig / bed files and
paired end bam files split in two}
\usage{
findNGSPairs(
  paths,
  f = c("forward", "fwd"),
  r = c("reverse", "rev"),
  format = "wig"
)
}
\arguments{
\item{paths}{a character path at least one .wig / .bed file}

\item{f}{Default (c("forward", "fwd")
a character vector for forward direction regex.}

\item{r}{Default (c("reverse", "rev")
a character vector for reverse direction regex.}

\item{format}{default "wig", for bed do "bed". Also searches
compressions of these variants.}
}
\value{
if not all are paired, return original list,
if they are all paired, return a data.table with matches as 2 columns
}
\description{
Find pair of forward and reverse strand wig / bed files and
paired end bam files split in two
}
\keyword{internal}
