% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findORFs}
\alias{findORFs}
\title{Creates an IRangesList of Open Reading Frames}
\usage{
findORFs(fastaSeqs, startCodon = "ATG", stopCodon = "TAA|TAG|TGA",
  longestORF = FALSE, minimumLength = 0)
}
\arguments{
\item{fastaSeqs}{DNA sequences to search for Open Reading Frames,
must be DNAStringSet.}

\item{startCodon}{string. Default is "ATG".}

\item{stopCodon}{string. Default is "TAA|TAG|TGA".}

\item{longestORF}{bolean. Default FALSE. When TRUE pick
longest ORF per group/chromosome.
When FALSE will report all open reaidng frames, even overlapping small ones.}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs to be at least
 START + 8*3 [bp] + STOP.}
}
\value{
An IRangesList of ORFs grouped by fasta sequence from input.
}
\description{
Input is a  DNAStringSet / character vector with fastaSequence.
If you want Genomic coordinates, use function \code{\link{findMapORFs}}
}
\examples{
seqs <- c("ATGGGTATTTATA") # the dna sequence
startCodons <- "ATG|TGG|GGG"
stopCodons <- "TAA|AAT|ATA"

## now run
findORFs(seqs,startCodons, stopCodons,
                   longestORF = FALSE, minimumLength = 0)
## You will find 1 orf (1-11)

}
