% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findORFsFasta}
\alias{findORFsFasta}
\title{Creates GRanges with Open Reading Frames from fasta files.}
\usage{
findORFsFasta(file, startCodon = "ATG", stopCodon = "TAA|TAG|TGA",
  longestORF = TRUE, minimumLength = 0, is.circular = FALSE)
}
\arguments{
\item{file}{- Path to fasta file.}

\item{startCodon}{(character) Possible START codons to search for. Check
\code{\link{startDefinition}} for helper function.}

\item{stopCodon}{(character) Possible STOP codons to search for. Check
\code{\link{stopDefinition}} for helper function.}

\item{longestORF}{(logical) Default FALSE. When TRUE will only report ORFs
that are longest, all smaller overlapping ORFs will be ignored.
When FALSE will report all possible ORFs in all three reading frames.}

\item{minimumLength}{(integer) Default is 0. Minimum length of ORF, without
counting 3bp for START and STOP codons. For example minimumLength = 8 will
result in size of ORFs to be at least START + 8*3 [bp] + STOP. Use this param
to restrict search.}

\item{is.circular}{a logical(F), should ORFs that start before end
and after start be searched for.}
}
\value{
A GRanges object of ORFs mapped from fasta file.
 Each ORF includes START and STOP codon, seqname, strand and frame.
}
\description{
Each fasta header is treated separately, and name of the sequence will
be used as seqname in returned GRanges object.
Frame of the Open Reading Frame is also
returned in metadata column 'frame'.
}
\examples{
filePath <- system.file("extdata", "orfFindingExample.fasta",
package = "ORFik") ## location of the fasta file

findORFsFasta(filePath)
## orfs are now returned as GRanges.

}
\seealso{
Other findORFs: \code{\link{findMapORFs}},
  \code{\link{findORFs}}, \code{\link{startDefinition}},
  \code{\link{stopDefinition}}
}
