% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_filters.R
\name{findPeaksPerGenes}
\alias{findPeaksPerGenes}
\title{Find peaks per gene}
\usage{
findPeaksPerGenes(tx, reads, top_tx = 0.5, min_reads_per_tx = 10, type = "max")
}
\arguments{
\item{tx}{a GRangesList}

\item{reads}{a GAlignments or GRanges, must be 1 width reads like p-shifts,
or other reads that is single positioned.}

\item{top_tx}{numeric, default 0.50 (50\% top tx by reads).}

\item{min_reads_per_tx}{numeric, default 10. Gene must have at least
10 reads}

\item{type}{character, default "max". Get only max peak per gene.
Alternatives: "all", all peaks passing the input filter will be returned.
"median", only peaks that is higher than the median of all peaks}
}
\value{
a data.table of gene_id, position, counts of the peak, zscore
and standard deviation of the peak compared to rest of gene area.
}
\description{
For finding the peaks per gene, with some default filters.
A peak is basically a position of very high coverage compared to
its surrounding area.
}
