% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findWigPairs}
\alias{findWigPairs}
\title{Find pair of forward and reverse strand wig files}
\usage{
findWigPairs(paths)
}
\arguments{
\item{paths}{a character path at least one .wig file}
}
\value{
if not all are paired, return original list,
if they are all paired, return a data.table with matches as 2 columns
}
\description{
Find pair of forward and reverse strand wig files
}
