% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{firstEndPerGroup}
\alias{firstEndPerGroup}
\title{get first end per granges group}
\usage{
firstEndPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a GRangesList}

\item{keep.names}{a boolean, keep names or not}
}
\description{
get first end per granges group
}
