% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{firstExonPerGroup}
\alias{firstExonPerGroup}
\title{Get first exon per GRangesList group}
\usage{
firstExonPerGroup(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}
}
\value{
a GRangesList of the first exon per group
}
\description{
grl must be sorted, call ORFik:::sortPerGroup if needed
}
