% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_helper.R
\name{fix_malformed_gff}
\alias{fix_malformed_gff}
\title{Fix a malformed gff file}
\usage{
fix_malformed_gff(gff)
}
\arguments{
\item{gff}{character, path to gtf, can not be gzipped!}
}
\value{
path of fixed gtf
}
\description{
Basically removes all info lines with character length > 32768 and
save that new file.
}
\examples{
# fix_malformed_gff("my_bad_gff.gff")

}
