% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{fixSeqnames}
\alias{fixSeqnames}
\title{seqnames cleanup
For many datasets, the fa file and the gtf file have different naming
This functions tries to fix the naming to the GRanges standard
chrX instead of X chr1 instead of 1 etc..}
\usage{
fixSeqnames(grl)
}
\arguments{
\item{grl}{a GRangesList}
}
\description{
seqnames cleanup
For many datasets, the fa file and the gtf file have different naming
This functions tries to fix the naming to the GRanges standard
chrX instead of X chr1 instead of 1 etc..
}
