% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_region.R
\name{flankPerGroup}
\alias{flankPerGroup}
\title{Get flanks per group}
\usage{
flankPerGroup(grl)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}
}
\value{
a GRangesList, 1 GRanges per group with:
 start as minimum start of group and end as maximum per group.
}
\description{
For a GRangesList, get start and end site, return back as GRangesList.
}
\examples{
grl <- GRangesList(tx1 = GRanges("1", IRanges(c(1,5), width = 2), "+"),
                   tx2 = GRanges("2", IRanges(c(10,15), width = 2), "+"))
flankPerGroup(grl)
}
