% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{footprints.analysis}
\alias{footprints.analysis}
\title{Pre shifting plot analysis}
\usage{
footprints.analysis(rw, heatmap, region = "start of CDS")
}
\arguments{
\item{rw}{a data.table of position, score and fraction (read length)
of either TIS or TES (translation end site, around 3' UTR)}

\item{heatmap}{a logical or character string, default FALSE.
If TRUE, will plot heatmap of
raw reads before p-shifting to console, to see if shifts given make sense.
You can also set a filepath to save the file there.}

\item{region}{a character string, default "start of CDS"}
}
\value{
invisible(NULL)
}
\description{
For internal use only!
}
