% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{fpkmRFP}
\alias{fpkmRFP}
\title{Get fpkm of ribo-seq data}
\usage{
fpkmRFP(grl, RFP, pseudoCount = 0)
}
\arguments{
\item{grl}{a GRangesList object with usually ORFs, but can also be
either leaders, cds', 3' utrs or  ORFs are a special case,
see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}
}
\value{
a numeric vector with the fpkm values
}
\description{
Get fpkm of ribo-seq data
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkmRNA}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
