% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{fpkmRNA}
\alias{fpkmRNA}
\title{Get fpkm of RNA-seq data}
\usage{
fpkmRNA(grl, RNA, tx_len, pseudoCount = 0)
}
\arguments{
\item{grl}{a GRangesList object with usually ORFs, but can also be
either leaders, cds', 3' utrs or  ORFs are a special case,
see argument tx_len}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}

\item{tx_len}{the transcript lengths of the transcripts
a named (tx names) vector of integers.
If you have the transcripts as GRangesList,
call ORFik:::widthPerGroup(tx, T)
If not:
NB!!! if you used cage data, then
the tss for the the leaders have changed,
therefor the tx lengths have changed,
if so, call:
te(grl, RNA, RFP, tx_len = TxLen(Gtf, fiveUTRs))
where fiveUTRs are the changed cageLeaders.}
}
\value{
a numeric vector with the fpkm values
}
\description{
Get fpkm of RNA-seq data
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkmRFP}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
