% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{fractionLength}
\alias{fractionLength}
\title{Fraction Length}
\usage{
fractionLength(grl, tx_len)
}
\arguments{
\item{grl}{a GRangesList object with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{tx_len}{the transcript lengths of the transcripts
a named (tx names) vector of integers.
If you have the transcripts as GRangesList,
call ORFik:::widthPerGroup(tx, T)
If not:
Normally you call argument as: tx_len = txLen(Gtf)
fractionLength(grl, tx_len = txLen(Gtf))
See ?ORFik:::txLen
NB!!! if you used cage data, then
the tss for the the leaders have changed,
therefor the tx lengths have changed,
if so, call:
te(grl, RNA, RFP, tx_len = TxLen(Gtf, fiveUTRs))
where fiveUTRs are the changed cageLeaders.}
}
\value{
a numeric vector of ratios
}
\description{
is defined as (length of grls)/(length of transcripts)
see article: 10.1242/dev.098343
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
