% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{fractionNames}
\alias{fractionNames}
\title{Get cell fraction name variants}
\usage{
fractionNames()
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: cytosolic, mitochondrial, specific gene knock down
}
\seealso{
Other experiment_naming: 
\code{\link{batchNames}()},
\code{\link{cellLineNames}()},
\code{\link{cellTypeNames}()},
\code{\link{conditionNames}()},
\code{\link{inhibitorNames}()},
\code{\link{libNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{stageNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
\keyword{internal}
