% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{geneToSymbol}
\alias{geneToSymbol}
\title{Get gene symbols from Ensembl gene ids}
\usage{
geneToSymbol(
  df,
  organism_name = organism(df),
  gene_ids = filterTranscripts(df, by = "gene"),
  org.dataset = paste0(tolower(substr(organism_name, 1, 1)), gsub(".* ", replacement =
    "", organism_name), "_gene_ensembl"),
  ensembl = useEnsembl("ensembl", dataset = org.dataset),
  attribute = c(`Homo sapiens` = "hgnc_symbol", `Mus musculus` = "mgi_symbol", `Rattus
    norvegicus` = "mgi_symbol")[organism_name]
)
}
\value{
data.table with 2 columns gene_id and gene_symbol named after
attribute, sorted in order of gene_ids input.
}
\description{
If your organism is not in this list of supported
organisms, manually assign the input arguments
}
\details{
Will check for already existing table of all genes, and use that instead
of redownloading every time.
}
