% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{get_5ends}
\alias{get_5ends}
\title{Get the start of each read}
\usage{
get_5ends(alignments, read_length)
}
\arguments{
\item{alignments}{the ribo-seq footprints}

\item{read_length}{the read length to search for in alignments}
}
\value{
a GRanges object with a hit column
}
\description{
Get the start of each read
}
