% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{get_cds150}
\alias{get_cds150}
\title{Get cds' for shoelaces periodicity}
\usage{
get_cds150(txdb, txNames)
}
\arguments{
\item{txdb}{a TxDb object from gtf}

\item{txNames}{a character vector of the longest
valid transcript names}
}
\value{
a GRangesList of tiled cds' to 150 length
}
\description{
Gives you the first 150 bases per cds
}
