% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{get_longest_tx}
\alias{get_longest_tx}
\title{Get the longest transcripts}
\usage{
get_longest_tx(txdb)
}
\arguments{
\item{txdb}{a TxDb object from gtf}
}
\value{
a character vector of valid tramscript names
}
\description{
These settings must be true:
utr5 & utr3 > 30 width
cds > 150 width
If these are accepted, it is returned
}
