% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_helper.R
\name{get_phix_genome}
\alias{get_phix_genome}
\title{Download genome (fasta), annotation (GTF) and contaminants}
\usage{
get_phix_genome(phix, output.dir, gunzip)
}
\arguments{
\item{phix}{logical, default FALSE, download phiX sequence to filter
 out Illumina control reads. ORFik defines Phix as a contaminant genome.
Phix is used in Illumina sequencers for sequencing quality control.
Genome is: refseq, Escherichia phage phiX174.
If sequencing facility created fastq files with the command \code{bcl2fastq},
then there should be very few phix reads left in the fastq files received.}

\item{output.dir}{directory to save downloaded data}

\item{gunzip}{logical, default TRUE, uncompress downloaded files
that are zipped when downloaded, should be TRUE!}
}
\value{
A named character vector with the paths to the downloaded genome,
annotation, and any contaminant references. If `merge_contaminants = TRUE`,
only the merged contaminant fasta is returned, not the individual files.
}
\description{
This function downloads the genome, annotation, and contaminant references
used for alignment if they are not already present. By default it uses the
Ensembl reference. When the run completes, it stores a file called
\code{file.path(output.dir, "outputs.rds")} with the output paths for the
downloaded files.
}
\details{
For most non-model non-vertebrate organisms, you need
my fork of biomartr for it to work:
remotes::install_github("Roleren/biomartr")
If a download fails or you used the wrong organism or assembly, delete the
incorrect files and run again.\cr
Set \code{remake = TRUE} to rebuild everything from scratch.\cr

Files created during the download process include:\cr
- A fasta index for the genome\cr
- A TxDb to speed up GTF/GFF reading\cr
- Separate or merged contaminant files\cr
Optional files that can also be created:\cr
- Gene symbols (hgnc, etc)\cr
- Uniprot ids (For name of protein structures)\cr
To use references already on disk, pass their paths as \code{genome} and
\code{GTF}. Supply \code{organism} and \code{output.dir} as well.
A completed run reuses the paths in \code{outputs.rds}; choose a separate
output directory or set \code{remake = TRUE} when changing references.
}
\examples{

## Get Saccharomyces cerevisiae genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel")
## Download and add pseudo 5' UTRs
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel",
#  pseudo_5UTRS_if_needed = 100)
## Get Danio rerio genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Danio rerio", tempdir())

output.dir <- "/Bio_data/references/zebrafish"
## Get Danio rerio and Phix contamints to deplete during alignment
#getGenomeAndAnnotation("Danio rerio", output.dir, phix = TRUE)

## Optimize for ORFik (speed up for large annotations like human or zebrafish)
#getGenomeAndAnnotation("Danio rerio", tempdir(), optimize = TRUE)

# Drosophila melanogaster (toplevel exists only)
#getGenomeAndAnnotation("drosophila melanogaster", output.dir = file.path(config["ref"],
# "Drosophila_melanogaster_BDGP6"), assembly_type = "toplevel")
## How to save malformed refseq gffs:
## First run function and let it crash:
#annotation <- getGenomeAndAnnotation(organism = "Arabidopsis thaliana",
#  output.dir = "~/Desktop/test_plant/",
#  assembly_type = "primary_assembly", db = "refseq")
## Then apply a fix (example for linux, too long rows):
# fixed_gff <- fix_malformed_gff("~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.gff")
## Then updated arguments:
# annotation <- c(fixed_gff, "~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.fna")
# names(annotation) <- c("gtf", "genome")
# Then make the txdb (for faster R use)
# makeTxdbFromGenome(annotation["gtf"], annotation["genome"], organism = "Arabidopsis thaliana")
}
\references{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4919035/
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.allsteps.multiQC}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.multiQC}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{install.fastp}()}
}
\keyword{internal}
