% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_helper.R
\name{get_silva_rRNA}
\alias{get_silva_rRNA}
\title{Download Silva SSU & LSU sequences}
\usage{
get_silva_rRNA(output.dir)
}
\arguments{
\item{output.dir}{directory to save downloaded data}
}
\value{
filepath to downloaded file
}
\description{
Version downloaded is 138.1. NR99_tax (non redundant)
}
\details{
If it fails from timeout, set higher timeout: options(timeout = 200)
}
\examples{
output.dir <- tempdir()
# get_silva_rRNA(output.dir)
}
