% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{get_start_stop}
\alias{get_start_stop}
\title{Get start and stops of cds windows}
\usage{
get_start_stop(txdb, txNames, start = TRUE, stop = TRUE)
}
\arguments{
\item{txdb}{a txdb object from gtf}

\item{txNames}{a character vector of the longest
valid transcript names}

\item{start}{logical (T), include starts}

\item{stop}{logical (F), include stops}
}
\value{
a list of starts and stops with cds windows
}
\description{
For each cds in longest_tx, get a window around start and stop
of each cds. +29 and -30 from cds start
}
