% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{get_top_tx}
\alias{get_top_tx}
\title{Get top percentage of transcript riboseq reads}
\usage{
get_top_tx(ribo, top_tx)
}
\arguments{
\item{ribo}{a list of lists with overlaps of transcripts and ribo-seq reads}

\item{top_tx}{numeric(10), which top percentage of transcripts to use.}
}
\value{
a list of lists, with accepted lists in ribo
}
\description{
Get top percentage of transcript riboseq reads
}
