% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{getFilteredCageData}
\alias{getFilteredCageData}
\title{Get CageData as granges}
\usage{
getFilteredCageData(dataName, filterValue = 1)
}
\arguments{
\item{dataName}{The location of the cage-file}

\item{filterValue}{The number of counts(score) to filter on for a tss to pass as hit}
}
\description{
Get CageData as granges
}
