% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{getGenomeAndFasta}
\alias{getGenomeAndFasta}
\title{Download genome (fasta), GTF and contaminants}
\usage{
getGenomeAndFasta(
  organism,
  output.dir,
  db = "ensembl",
  GTF = TRUE,
  genome = TRUE,
  phix = FALSE,
  ncRNA = "",
  tRNA = "",
  rRNA = "",
  gunzip = TRUE,
  remake = FALSE
)
}
\arguments{
\item{organism}{scientific name of organism}

\item{output.dir}{directory to save downloaded data}

\item{db}{database to use, default adviced: "ensembl"}

\item{GTF}{logical, default: TRUE, download gtf of organism}

\item{genome}{logical, default: TRUE, download genome of organism}

\item{phix}{logical, default FALSE, download phix sequence to filter
 out with.
Only use if illumina sequencing. Phix is used in Illumina sequencers for
sequencing quality control. Genome is: refseq, Escherichia virus phiX174}

\item{ncRNA}{chatacter, default "" (no download), a contaminant genome,
if not "" it must be a character vector
of species common name (not scientific) Homo sapiens is human, Rattus norwegicus is rat etc,
download ncRNA sequence to filter out with. From NONCODE online server, if you cant find
common name see: http://www.noncode.org/download.php/}

\item{tRNA}{chatacter, default "" (not used),
if not "" it must be a character vector  to valid path of mature
tRNAs fasta file to remove as contaminants. Find and download your wanted mtRNA at:
http://gtrnadb.ucsc.edu/, or run trna-scan on you genome.}

\item{rRNA}{chatacter, default "" (not used),
if not "" it must be a character vector to valid path of mature
rRNA fasta file to remove as contaminants. Find and download your wanted rRNA at:
https://www.arb-silva.de/}

\item{gunzip}{logical, default TRUE, uncompress downloaded files
that are zipped as default}

\item{remake}{logical, default: FALSE, if TRUE remake everything specified}
}
\value{
a character vector of genomeas and gtf downloaded
}
\description{
Will create a R transcript database (TxDb object) from the genome. \cr
It will also index the genome \cr
If you misspelled something or crashed, delete wrong files and
run again.\cr
Do remake = TRUE, to do it all over again.
}
\examples{
output.dir <- "/Bio_data/references/zebrafish"
#getGenomeAndFasta("Danio rerio", output.dir)
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{install.fastp}()}
}
\concept{STAR}
