% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGRanges}
\alias{getGRanges}
\title{Internal GRanges loader from fst data.frame}
\usage{
getGRanges(df)
}
\arguments{
\item{df}{a data.frame with columns minimum 4 columns:
seqnames, start, strand and width.\cr
Additional columns will be assigned as meta columns}
}
\value{
GRanges object
}
\description{
Internal GRanges loader from fst data.frame
}
\keyword{internal}
