% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{getNGenesCoverage}
\alias{getNGenesCoverage}
\title{Get number of genes per grouping}
\usage{
getNGenesCoverage(coverage)
}
\arguments{
\item{coverage}{a data.table with coverage}
}
\value{
number of genes in coverage
}
\description{
Get number of genes per grouping
}
