% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{getStartStopWindows}
\alias{getStartStopWindows}
\title{Get Start and Stop codon within specified windows over CDS.}
\usage{
getStartStopWindows(txdb, txNames, start = TRUE, stop = TRUE,
  window_size = 30L, cds = NULL)
}
\arguments{
\item{txdb}{a txdb object of annotations}

\item{txNames}{a character vector of the transcript names to use}

\item{start}{(logical) whether to include start codons}

\item{stop}{(logical) whether to incude stop codons}

\item{window_size}{(integer) size of the window to extract upstream of the
start/stop codon and downstream}

\item{cds}{a GRangesList with cds, a speedup if you already have them
loaded.}
}
\value{
a list with two slots "starts" and "stops", each contains a
GRangesList of windows around start and stop codons for the transcripts of
interest
}
\description{
For each cds in \code{txdb} object, filtered by \code{txNames},
get a window around start and stop codons within \code{window_size}
downstream and upstream of the codon.
}
\examples{
gtf_file <- system.file("extdata", "annotations.gtf", package = "ORFik")
txdb <- GenomicFeatures::makeTxDbFromGFF(gtf_file, format = "gtf")
txNames <- txNamesWithLeaders(txdb)
getStartStopWindows(txdb, txNames)

}
