% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_helpers.R
\name{getWeights}
\alias{getWeights}
\title{Get weights from a subject GenomicRanges object}
\usage{
getWeights(subject, weight = 1L)
}
\arguments{
\item{subject}{a GRanges, IRanges, GAlignment, GAlignmentPairs or
covRle object}

\item{weight}{a numeric/integer vector or metacolumn name.
(default: 1L, no differential weighting).
If weight is name of defined meta column in reads object,
it gives the number of times a read was found at that position.
GRanges("chr1", 1, "+", score = 5), would mean "score" column tells
that this alignment region was found 5 times.
if 1L it means each read is weighted equal as 1,
this is what among others countOverlaps() presumes,
if single number (!= 1), it repeats for all ranges,
if vector with length > 1, it must be equal size of the
reads object.}
}
\value{
a numeric vector of weights of equal size to subject
}
\description{
Get weights from a subject GenomicRanges object
}
\keyword{internal}
