% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{GrangesSplitByExonSkeleton}
\alias{GrangesSplitByExonSkeleton}
\title{Helperfunction for map_to_GRanges
split a GRangesList on the exon boundaries of the skeleton
if you have a grl without the exon splits, but a reference(skeleton),
then use this function}
\usage{
GrangesSplitByExonSkeleton(grl, skeleton)
}
\arguments{
\item{grl}{The GRangesList found so far by the orf-scanner}

\item{skeleton}{the reference skeleton to find exon splits}
}
\description{
Helperfunction for map_to_GRanges
split a GRangesList on the exon boundaries of the skeleton
if you have a grl without the exon splits, but a reference(skeleton),
then use this function
}
