% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{groupGRangesBy}
\alias{groupGRangesBy}
\title{make GRangesList from GRanges,}
\usage{
groupGRangesBy(gr, other = NULL)
}
\arguments{
\item{gr}{a GRanges object}

\item{other}{a vector of names to group, no 2 groups can have same name}
}
\value{
a GRangesList named after names(Granges) if other is NULL, else
names are from unique(other)
}
\description{
Grouped by names or another column(other)
 ig. if GRanges should be grouped by gene,
 give gene column as other
}
