% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{GroupGRangesByNames}
\alias{GroupGRangesByNames}
\title{make GRangesList from GRanges, grouped by names}
\usage{
GroupGRangesByNames(gr)
}
\arguments{
\item{gr}{a GRanges object}
}
\description{
make GRangesList from GRanges, grouped by names
}
