% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{GroupGRangesByOther}
\alias{GroupGRangesByOther}
\title{make GRangesList from GRanges, grouped by other}
\usage{
GroupGRangesByOther(gr, other)
}
\arguments{
\item{gr}{a GRanges object}

\item{other}{a vector of names to group by}
}
\description{
make GRangesList from GRanges, grouped by other
}
