% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{groupings}
\alias{groupings}
\title{Get number of ranges per group as an iterator}
\usage{
groupings(grl)
}
\arguments{
\item{grl}{GRangesList}
}
\value{
an integer vector
}
\description{
Get number of ranges per group as an iterator
}
\examples{
grl <- GRangesList(GRanges("1", c(1, 3, 5), "+"),
                   GRanges("1", c(19, 21, 23), "+"))
ORFik:::groupings(grl)

}
