% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{gSort}
\alias{gSort}
\title{A faster reimplementation of GenomicRanges::sort() for GRangesList}
\usage{
gSort(grl, decreasing = F)
}
\arguments{
\item{grl}{a GRangesList}

\item{decreasing}{should the first in each group have max(start(group))
->T or min-> default(F) ?}
}
\description{
A faster reimplementation of GenomicRanges::sort() for GRangesList
}
