% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{hasHits}
\alias{hasHits}
\title{Hits from reads}
\usage{
hasHits(grl, reads, keep.names = FALSE)
}
\arguments{
\item{grl}{a GRanges or GRangesList}

\item{reads}{a GAlignment or GRanges object with reads}

\item{keep.names}{logical (F), keep names or not}
}
\value{
a list of logicals, T == hit, F == no hit
}
\description{
Finding GRanges groups that have overlap hits with reads
}
