% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import.bedo}
\alias{import.bedo}
\title{Load GRanges object from .bedo}
\usage{
import.bedo(path)
}
\arguments{
\item{path}{a character, location on disc (full path)}
}
\value{
GRanges object
}
\description{
.bedo is .bed ORFik, an optimized bed format for coverage reads with read lengths
.bedo is a text based format with columns:
chromosome, start, stop, width, strand, (cigar # M's, match/mismatch total)
, duplicates of that read
}
\details{
export with export.bedo
}
