% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imports.R
\name{import.bedoc}
\alias{import.bedoc}
\title{Load GAlignments object from .bedoc}
\usage{
import.bedoc(path)
}
\arguments{
\item{path}{a character, location on disc (full path)}
}
\value{
GAlignments object
}
\description{
A much faster way to store, load and use bam files.\cr
.bedoc is .bed ORFik, an optimized bed format for coverage reads with
cigar and replicate number.\cr
.bedoc is a text based format with columns (5 maximum):\cr
1. chromosome\cr 2. cigar: (cigar # M's, match/mismatch total)
\cr 3. start (left most position) \cr 4. strand (+, -, *)\cr
5. score: duplicates of that read\cr
}
\details{
Positions are 1-based, not 0-based as .bed.
export with export.bedo
}
