% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_imports.R
\name{import.fstwig}
\alias{import.fstwig}
\title{Import region from fastwig}
\usage{
import.fstwig(gr, dir, id = "", readlengths = "all")
}
\arguments{
\item{gr}{a GRanges object of exons}

\item{dir}{prefix to filepath for file strand and chromosome will be added}

\item{id}{id to column type, not used currently!}

\item{readlengths}{integer / character vector, default "all". Or a subset
of readlengths.}
}
\value{
a data.table with columns specified by readlengths
}
\description{
Import region from fastwig
}
