% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{importGtfFromTxdb}
\alias{importGtfFromTxdb}
\title{Import the GTF / GFF that made the txdb}
\usage{
importGtfFromTxdb(txdb)
}
\arguments{
\item{txdb}{a TxDb, path to txdb / gff or ORFik experiment object}
}
\value{
data.frame, the gtf/gff object imported with rtracklayer::import
}
\description{
Import the GTF / GFF that made the txdb
}
