% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{inFrameWithCDS}
\alias{inFrameWithCDS}
\title{find frame for each orf relative to cds}
\usage{
inFrameWithCDS(dists)
}
\arguments{
\item{dists}{a vector of distances between ORF and cds
Input of this function, is the output of the function distOrfToCds}
}
\description{
find frame for each orf relative to cds
}
