% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{install.sratoolkit}
\alias{install.sratoolkit}
\title{Download sra toolkit}
\usage{
install.sratoolkit(folder = "~/bin", version = "2.11.3")
}
\arguments{
\item{folder}{installation folder, default `"~/bin"`}

\item{version}{toolkit version, default `"2.11.3"`}
}
\value{
Path to `fastq-dump` inside the installed sratoolkit
}
\description{
ORFik supports Linux and macOS. The Linux installer has been tested on
64-bit CentOS and Ubuntu. On other Linux distributions, the CentOS binaries
are used as a fallback.
}
\examples{
# install.sratoolkit()
## Custom folder and version (not advised)
folder <- "/I/WANT/IT/HERE/"
# install.sratoolkit(folder, version = "2.10.9")

}
\references{
https://ncbi.github.io/sra-tools/fastq-dump.html
}
\seealso{
Other sra: 
\code{\link{browseSRA}()},
\code{\link{download.SRA}()},
\code{\link{download.SRA.metadata}()},
\code{\link{download.ebi}()},
\code{\link{get_bioproject_candidates}()},
\code{\link{rename.SRA.files}()}
}
\concept{sra}
