% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.gr_or_grl}
\alias{is.gr_or_grl}
\title{Helper function to check for GRangesList or GRanges class}
\usage{
is.gr_or_grl(class)
}
\arguments{
\item{class}{the class you want to check if is GRL or GR,
either a character from class or the object itself.}
}
\value{
a boolean
}
\description{
Helper function to check for GRangesList or GRanges class
}
\seealso{
Other utils: \code{\link{bedToGR}},
  \code{\link{convertToOneBasedRanges}},
  \code{\link{findFa}}, \code{\link{fread.bed}},
  \code{\link{is.grl}}, \code{\link{validGRL}}
}
\concept{utils}
