% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{is.grl}
\alias{is.grl}
\title{Helper function to check for GRangesList}
\usage{
is.grl(class)
}
\arguments{
\item{class}{the class you want to check if is GRL,
either a character from class or the object itself.}
}
\value{
a boolean
}
\description{
Helper function to check for GRangesList
}
\seealso{
Other validity: 
\code{\link{checkRFP}()},
\code{\link{checkRNA}()},
\code{\link{is.ORF}()},
\code{\link{is.gr_or_grl}()},
\code{\link{is.range}()},
\code{\link{validGRL}()},
\code{\link{validSeqlevels}()}
}
\concept{validity}
