% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{isOverlappingCds}
\alias{isOverlappingCds}
\title{find for each orf if it overlaps cds range}
\usage{
isOverlappingCds(dists)
}
\arguments{
\item{dists}{a vector of distances between ORF and cds
Input of this function, is the output of the function distOrfToCds}
}
\description{
find for each orf if it overlaps cds range
}
