% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{kozacSequenceScore}
\alias{kozacSequenceScore}
\title{Make a score for each ORFs start region
The closer the sequence is to the kozak sequence
The higher the score, based on simplification of PWM
score system: 4 upstream, 5 downstream of start
CACCATGGC, 1+3+1+2, skip ATG, +2+1 = 10
CGCCATGGC, 1+!2+1+2, skip ATG, +2+1 = 9
Inspired by experimental bit values for each position}
\usage{
kozacSequenceScore(grl, faFile = NULL, species = "human")
}
\arguments{
\item{grl}{a GRangesList grouped by ORF}

\item{faFile}{a faFile from the fasta file}

\item{species}{which species to use, currently only support human}
}
\description{
Make a score for each ORFs start region
The closer the sequence is to the kozak sequence
The higher the score, based on simplification of PWM
score system: 4 upstream, 5 downstream of start
CACCATGGC, 1+3+1+2, skip ATG, +2+1 = 10
CGCCATGGC, 1+!2+1+2, skip ATG, +2+1 = 9
Inspired by experimental bit values for each position
}
