% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_plots.R
\name{kozak_IR_ranking}
\alias{kozak_IR_ranking}
\title{Rank kozak initiation sequences}
\usage{
kozak_IR_ranking(cds_k, mrna, dt.ir, faFile, group.min = 10, species = "human")
}
\arguments{
\item{cds_k}{cds ranges (GRangesList)}

\item{mrna}{mrna ranges (GRangesList)}

\item{dt.ir}{data.table with a column called IR, initiation rate}

\item{faFile}{\code{\link{FaFile}}, BSgenome, fasta/index file path or an
ORFik \code{\link{experiment}}. This file is usually used to find the
transcript sequences from some GRangesList.}

\item{group.min}{numeric, default 10. Minimum transcripts per initation
group to be included}

\item{species}{("human"), which species to use,
currently supports human (Homo sapiens), zebrafish (Danio rerio) and mouse (Mus musculus).
Both scientific or common name for these species will work.
You can also specify a pfm for your own species.
Syntax of pfm is an rectangular integer matrix,
where all columns must sum to the same value, normally 100.
See example for more information.
Rows are in order: c("A", "C", "G", "T")}
}
\value{
a ggplot grid object
}
\description{
Defined as region (-4, -1) relative to TIS
}
