% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{lastExonEndPerGroup}
\alias{lastExonEndPerGroup}
\title{get last end per granges group}
\usage{
lastExonEndPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
a Rle(keep.names = T), or integer vector(F)
}
\description{
get last end per granges group
}
