% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{lastExonPerGroup}
\alias{lastExonPerGroup}
\title{Get last exon per GRangesList group}
\usage{
lastExonPerGroup(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}
}
\value{
a GRangesList of the last exon per group
}
\description{
grl must be sorted, call ORFik:::sortPerGroup if needed
}
