% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{lastExonStartPerGroup}
\alias{lastExonStartPerGroup}
\title{Get last start per granges group}
\usage{
lastExonStartPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
a Rle(keep.names = T), or integer vector(F)
}
\description{
Get last start per granges group
}
