% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{length,covRle-method}
\alias{length,covRle-method}
\title{length covRle}
\usage{
\S4method{length}{covRle}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
an integer, number of chromosomes in covRle object
}
\description{
Number of chromosomes
}
