% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{length,covRleList-method}
\alias{length,covRleList-method}
\title{length covRleList}
\usage{
\S4method{length}{covRleList}(x)
}
\arguments{
\item{x}{a covRleList object}
}
\value{
an integer, number of covRle objects
}
\description{
Number of covRle objects
}
