% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{length,GRangesList-method}
\alias{length,GRangesList-method}
\title{Get length of GRangesList}
\usage{
\S4method{length}{GRangesList}(x)
}
\arguments{
\item{x}{a GRangesList}
}
\value{
an integer (length 1)
}
\description{
Faster version than S4Vector generic caller
}
