% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{lengths,covRle-method}
\alias{lengths,covRle-method}
\title{lengths covRle}
\usage{
\S4method{lengths}{covRle}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
a named integer vector of chromosome lengths
}
\description{
Lengths of each chromosome
}
