% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{lengths,covRleList-method}
\alias{lengths,covRleList-method}
\title{lengths covRleList}
\usage{
\S4method{lengths}{covRleList}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
a named integer vector of chromosome lengths
}
\description{
Lengths of each chromosome
}
