% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{libFolder,experiment-method}
\alias{libFolder,experiment-method}
\title{Get path to ORFik experiment library folder}
\usage{
\S4method{libFolder}{experiment}(x, mode = "first")
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}

\item{mode}{character, default "first". Alternatives: "unique", "all".}
}
\value{
a character path
}
\description{
Get path to ORFik experiment library folder
}
