% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{libNames}
\alias{libNames}
\title{Get library name variants}
\usage{
libNames()
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: RFP is main naming, but a variant is ribo-seq
ribo-seq will then be renamed to RFP
}
\seealso{
Other experiment_naming: 
\code{\link{cellLineNames}()},
\code{\link{conditionNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{stageNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
