% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{libraryTypes}
\alias{libraryTypes}
\title{Which type of experiments?}
\usage{
libraryTypes(df)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}
}
\value{
library types (character vector)
}
\description{
Which type of experiments?
}
\seealso{
Other ORFik_experiment: 
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
