% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{list.experiments}
\alias{list.experiments}
\title{List current experiment available}
\usage{
list.experiments(
  dir = "~/Bio_data/ORFik_experiments/",
  pattern = "*",
  libtypeExclusive = NULL
)
}
\arguments{
\item{dir}{directory for ORFik experiments: default:
"~/Bio_data/ORFik_experiments/"}

\item{pattern}{allowed patterns in experiment file name:
default ("*", all experiments)}

\item{libtypeExclusive}{search for experiments with exclusivly this
libtype, default (NULL, all)}
}
\description{
Will only search .csv extension, also exclude any experiment
with the word template.
}
\examples{
## Make your experiments
df <- ORFik.template.experiment(TRUE)
df2 <- df[1:6,] # Only first 2 libs
## Save them
# save.experiment(df, "~/Bio_data/ORFik_experiments/exp1.csv")
# save.experiment(df2, "~/Bio_data/ORFik_experiments/exp1_subset.csv")
## List all experiment you have:
## Path above is default path, so no dir argument needed
#list.experiments()
#list.experiments(pattern = "subset")
## For non default directory experiments
#list.experiments(dir = "MY/CUSTOM/PATH)
}
