% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_directory_config.R
\name{listGenomes}
\alias{listGenomes}
\title{List genomes created with ORFik}
\usage{
listGenomes(reference.folder = ORFik::config()["ref"])
}
\arguments{
\item{reference.folder}{character path, default:
\code{ORFik::config()["ref"]}.}
}
\value{
a data.table with 4 columns:\cr
 - character (name of folder)\cr
 - logical (does it have a gtf)
 - logical (does it have a fasta genome)
 - logical (does it have a STAR index)
 @export
 @examples
 listGenomes()
}
\description{
Given the reference.folder, list all valid references.
}
