% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadRegion}
\alias{loadRegion}
\title{Load transcript region}
\usage{
loadRegion(txdb, part = "tx")
}
\arguments{
\item{txdb}{a GRangesList or txdb object}

\item{part}{a character, one of: tx, leader, cds, trailer}
}
\value{
a GrangesList of region
}
\description{
Load if not already GRangesList
}
