% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadTxdb}
\alias{loadTxdb}
\title{General loader for txdb}
\usage{
loadTxdb(txdb)
}
\arguments{
\item{txdb}{a TxDb file or a path to one of:
(.gtf ,.gff, .gff2, .gff2, .db or .sqlite)}
}
\value{
a TxDb object
}
\description{
Useful to allow fast TxDb loader like .db
}
\examples{
library(GenomicFeatures)
# Get the gtf txdb file
txdbFile <- system.file("extdata", "hg19_knownGene_sample.sqlite",
                        package = "GenomicFeatures")
txdb <- loadDb(txdbFile)

}
