% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{mainNames}
\alias{mainNames}
\title{Get main name from variant name}
\usage{
mainNames(names, dt)
}
\arguments{
\item{names}{a character vector of names that must exist in dt$allNames}

\item{dt}{a data.table with 2 columns (mainName, allNames)}
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: RFP is main naming, but a variant is ribo-seq
ribo-seq will then be renamed to RFP
}
\seealso{
Other experiment_naming: 
\code{\link{cellLineNames}()},
\code{\link{conditionNames}()},
\code{\link{libNames}()},
\code{\link{repNames}()},
\code{\link{stageNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
