% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{makeBiotype}
\alias{makeBiotype}
\title{Make Gene biotype from gtf.}
\usage{
makeBiotype(
  gtf,
  biotype_file_out_path = file.path(dirname(gtf), "biotype_per_gene.fst")
)
}
\arguments{
\item{gtf}{path to gtf file}

\item{biotype_file_out_path}{path to save, default \code{file.path(dirname(getGtfPathFromTxdb(txdb, stop.error = TRUE)), "gene_symbol_tx_table.fst")}}
}
\value{
the data.table of gene_id and biotype
}
\description{
Make Gene biotype from gtf.
}
