% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{makeExonRanks}
\alias{makeExonRanks}
\title{Make grouping by exons ranks}
\usage{
makeExonRanks(grl, byTranscript = FALSE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}

\item{byTranscript}{logical (default: FALSE), groups orfs by transcript
name or ORF name, if ORfs are by transcript, check duplicates.}
}
\value{
an integer vector of indices for exon ranks
}
\description{
There are two ways to make vector of exon ranking:
1. Iterate per exon in ORF, byTranscript = FALSE
2. Iterate per ORF in transcript, byTranscript = TRUE.
}
\details{
Either by transcript or by original groupings.
Must be ordered, so that same transcripts are ordered together.
}
