% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeGAlignmentsFromDataFrame}
\alias{makeGAlignmentsFromDataFrame}
\title{Convert data.frame to GAlignment object}
\usage{
makeGAlignmentsFromDataFrame(x)
}
\arguments{
\item{x}{a data.frame / data.table with at least 4 columns:
seqnames, pos, cigar and strand, can also have a 5th column score.}
}
\value{
a GAlignments object
}
\description{
Convert data.frame to GAlignment object
}
