% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{makeGRangesListFromCharacter}
\alias{makeGRangesListFromCharacter}
\title{Convert a character vector to GRangesList}
\usage{
makeGRangesListFromCharacter(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a GRangesList
}
\description{
Convert a character vector to GRangesList
}
\examples{
vec <- c("1:14598834-14598914:+", "1:15210514-15210562:+;1:15214895-15215025:+")
makeGRangesListFromCharacter(vec)
}
