% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{makeORFNames}
\alias{makeORFNames}
\title{Make ORF names per orf, grl must be grouped by transcript}
\usage{
makeORFNames(grl)
}
\arguments{
\item{grl}{a GRangesList}
}
\description{
Make ORF names per orf, grl must be grouped by transcript
}
