% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{makeSummarizedExperimentFromBam}
\alias{makeSummarizedExperimentFromBam}
\title{Make a count matrix from a libraries}
\usage{
makeSummarizedExperimentFromBam(
  df,
  saveName = NULL,
  longestPerGene = TRUE,
  geneOrTxNames = "gene",
  region = "mrna",
  type = "count"
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{saveName}{a character (default NULL),
if set save experiment to path given. Always saved as .rds.,
it is optional to add .rds, it will be added for you if not present.}

\item{longestPerGene}{a logical (default TRUE), if FALSE all transcript
isoforms per gene.}

\item{geneOrTxNames}{a character (default gene), if tx use with tx names}

\item{region}{(default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{type}{default: "count" (raw counts matrix), alternative is "fpkm",
"log2fpkm" or "log10fpkm"}
}
\value{
a SummarizedExperiment object
}
\description{
Make a summerizedExperiment object from bam files
}
\details{
If txdb or gtf path is added, it is a rangedSummerizedExperiment
}
\examples{
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)

# makeSummarizedExperimentFromBam(df)
# Only cds (coding sequences):
# makeSummarizedExperimentFromBam(df, region = "cds")
# FPKM instead of raw counts
# makeSummarizedExperimentFromBam(df, type = "fpkm")
}
