% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defining_ORFs_helpers.R
\name{map_to_GRanges}
\alias{map_to_GRanges}
\title{Creates list of IRanges with Open Reading Frames.}
\usage{
map_to_GRanges(ORFdef, grangesObj)
}
\arguments{
\item{ORFdef}{- list of IRanges objects representing found ORFs.}

\item{grangesObj}{- GRanges object to map coordinates back from each IRanges ORF.}
}
\value{
A List of GRanges objects of ORFs mapped to grangesObj.
}
\description{
Creates list of IRanges with Open Reading Frames.
}
\examples{
#map_to_GRanges()
}

