% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{mapORFOrdering}
\alias{mapORFOrdering}
\title{Get map ordering from unique ORFs}
\usage{
mapORFOrdering(grl)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}
}
\value{
an integer vector
}
\description{
If you use function uniqueORFs, you loose ordering
from original set.
This function is used to remember the original mapping.
Usually used to cut down on number of calculations for features, etc.
}
\examples{
gr1 <- GRanges("1", IRanges(1,10), "+")
gr2 <- GRanges("1", IRanges(20, 30), "+")
# make a grl with duplicated ORFs (gr1 twice)
grl <- GRangesList(tx1_1 = gr1, tx2_1 = gr2, tx3_1 = gr1)
reOrdering <- mapORFOrdering(grl) # remember ordering
uniqueORFs <- uniqueORFs(grl)
# now map back
reMappedGrl <- uniqueORFs[reOrdering]
}
