% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{matchColors}
\alias{matchColors}
\title{Match coloring of coverage plot}
\usage{
matchColors(coverage, colors)
}
\arguments{
\item{coverage}{a data.table with coverage}

\item{colors}{a character vector of colors}
}
\value{
number of genes in coverage
}
\description{
Match coloring of coverage plot
}
