% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_or_grl_helpers.R
\name{matchNaming}
\alias{matchNaming}
\title{Match naming of GRangesList}
\usage{
matchNaming(gr, reference)
}
\arguments{
\item{gr}{a \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}}
or GRanges object}

\item{reference}{a GRangesList of a reference}
}
\value{
a GRangesList
}
\description{
Given a GRangesList and a reference, make the naming convention and
the number of metacolumns equal to reference
}
