% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchSeqStyle}
\alias{matchSeqStyle}
\title{A wrapper for seqlevelsStyle}
\usage{
matchSeqStyle(range, chrStyle = NULL)
}
\arguments{
\item{range}{a ranged object, (GRanges, GAlignment etc)}

\item{chrStyle}{a GRanges object, or a character style (Default: NULL) to
get seqlevelsStyle from. Is chromosome 1 called chr1 or 1, is mitocondrial
chromosome called MT or chrM etc. Will use 1st seqlevel-
style if more are present. Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a GAlignment/GRanges object depending on input.
}
\description{
To make sure chromosome naming is correct (chr1 vs 1 vs I etc)
}
