% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots_helpers.R
\name{metaPosition}
\alias{metaPosition}
\title{Calculate metaplot coverage of reads around input GRanges object.}
\usage{
metaPosition(cdsRanges, AllRiboReadsRangesResized)
}
\arguments{
\item{cdsRanges}{GRanges object of your CDSs start or stop postions.}

\item{AllRiboReadsRangesResized}{GRanges object of your reads.
You should resize them beforehand to width of 1 to focus on 5' ends of footprints.}
}
\value{
A data.frame with frequencies (Freq) of reads mapped to
positions (Position) specified in cdsRanges along with frame (Frame).
}
\description{
It should create window of window_size around input ranges eg.
for granges starting at 100-100: and window_size = 3
97-103
}
\examples{
#metaPosition()

}
