% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{name}
\alias{name}
\title{Get name of ORFik experiment}
\usage{
name(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
character, name of experiment
}
\description{
Get name of ORFik experiment
}
