% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{names,GRangesList-method}
\alias{names,GRangesList-method}
\title{Get names of GRangesList}
\usage{
\S4method{names}{GRangesList}(x)
}
\arguments{
\item{x}{a GRangesList}
}
\value{
a character vector
}
\description{
Faster version than S4Vector generic caller
}
