% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{names<-,GRangesList-method}
\alias{names<-,GRangesList-method}
\title{Get names of GRangesList}
\usage{
\S4method{names}{GRangesList}(x) <- value
}
\arguments{
\item{x}{a GRangesList}

\item{value}{character vector of names}
}
\value{
a GRangesList with updated names
}
\description{
Faster version than S4Vector generic caller
}
