% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{optimizeTranscriptRegions}
\alias{optimizeTranscriptRegions}
\title{Make optimized GRangesList objects saved to disc}
\usage{
optimizeTranscriptRegions(
  txdb,
  base_path = optimized_txdb_path(txdb, create.dir = TRUE),
  regions = c("tx", "mrna", "leaders", "cds", "trailers", "ncRNA")
)
}
\arguments{
\item{txdb}{a TxDb object, ORFik experiment object or a path to one of:
(.gtf ,.gff, .gff2, .gff2, .db or .sqlite),
Only in the loadRegion function: if it is a GRangesList, it will return it self.}

\item{base_path}{Directy and file prefix for files, will append "_region.rds", where region is
specific region.}

\item{regions}{character, default: c("tx", "mrna", "leaders", "cds", "trailers", "ncRNA").
Valid options specified by loadRegion.}
}
\value{
invisible(NULL)
}
\description{
Much faster to load
}
