% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{orfID}
\alias{orfID}
\title{Get id's for orf}
\usage{
orfID(grl, with.tx = FALSE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}}}

\item{with.tx}{a boolean, include transcript names,
if you want unique orfs, so that they dont have multiple
versions on different isoforms, set it to FALSE.}
}
\value{
a character vector of ids, 1 per orf
}
\description{
These id's can be uniqued by isoform etc,
this is not supported by GenomicRanges.
}
