% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFik.R
\docType{package}
\name{ORFik-package}
\alias{ORFik}
\alias{ORFik-package}
\title{ORFik for analysis of open reading frames.}
\description{
Main goals:
\enumerate{
\item Find open reading frames
\item Compatible with GRanges
\item Make it fast enough for big data
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JokingHero/ORFik}
  \item Report bugs at \url{https://github.com/JokingHero/ORFik/issues}
}

}
