% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{ORFik.template.experiment.zf}
\alias{ORFik.template.experiment.zf}
\title{An ORFik experiment to see how it looks}
\usage{
ORFik.template.experiment.zf(as.temp = FALSE)
}
\arguments{
\item{as.temp}{logical, default FALSE, load as ORFik experiment.
If TRUE, loads as data.frame template of the experiment.}
}
\value{
an ORFik \code{\link{experiment}}
}
\description{
Toy-data created to resemble Zebrafish genes:\cr
Number of genes: 150\cr
Ribo-seq: 1 library
}
\examples{
ORFik.template.experiment.zf()
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism,experiment-method}},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
