% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{ORFikReport}
\alias{ORFikReport}
\title{Full ORFik report}
\usage{
ORFikReport(experiment, outputDir = ".")
}
\arguments{
\item{experiment}{an ORFik experiment data.frame}

\item{outputDir}{directory to save output, default (.)}
}
\value{
NULL (all output stored to disc and .GlobalEnv)
}
\description{
Full ORFik report
}
