% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{OrfRankOrder}
\alias{OrfRankOrder}
\title{Get the orf number in transcripts, ig. second orf _1 -> 2}
\usage{
OrfRankOrder(grl)
}
\arguments{
\item{grl}{a GRangesList object with ORFs}
}
\description{
Get the orf number in transcripts, ig. second orf _1 -> 2
}
