% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{orfScore}
\alias{orfScore}
\title{Get ORF score for a GRangesList of ORFs}
\usage{
orfScore(grl, RFP)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object with ORFs}

\item{RFP}{ribozomal footprints, given as Galignment object,
Granges or GRangesList}
}
\value{
a matrix with 4 columns, the orfscore and score of
each of the 3 tiles
}
\description{
That is, of the 3 possible frame in an ORF
Is the first one most important, by how much ?
NB! Only support + and - strand, not *
See article: 10.1002/embj.201488411
}
\examples{
ORF <- GRanges(seqnames = "1",
ranges = IRanges(start = c(1, 10, 20),
                end = c(5, 15, 25)),
strand = "+")
names(ORF) <- c("tx1", "tx1", "tx1")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25),"+")
orfScore(grl, RFP)

}
\seealso{
Other features: \code{\link{computeFeaturesSpecial}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
