% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{ORFStartSites}
\alias{ORFStartSites}
\title{get the start sites from a GRangesList of orfs grouped by orfs
In ATGTTTTGG, get the position of the A.}
\usage{
ORFStartSites(grl, asGR = F, keep.names = F, is.sorted = F)
}
\arguments{
\item{grl}{a GRangesList object}

\item{asGR}{a boolean, return as GRanges object}

\item{is.sorted}{a speedup, if you know the ranges are sorted}
}
\value{
if asGR is False, a vector, if True a GRanges object
}
\description{
get the start sites from a GRangesList of orfs grouped by orfs
In ATGTTTTGG, get the position of the A.
}
