% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{OrfToTxNames}
\alias{OrfToTxNames}
\title{Get transcript names from orf names
names must either be a column called names, or the names of the
grl object must be orf names}
\usage{
OrfToTxNames(grl)
}
\arguments{
\item{grl}{a GRangeslist grouped by ORF}
}
\description{
Get transcript names from orf names
names must either be a column called names, or the names of the
grl object must be orf names
}
