% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{outputLibs}
\alias{outputLibs}
\title{Output bam/bed/wig files to R as variables}
\usage{
outputLibs(df, chrStyle = NULL, envir = .GlobalEnv)
}
\arguments{
\item{df}{an ORFik experiment data.frame}

\item{chrStyle}{the sequencelevels style (GRanges object or chr)}

\item{envir}{environment to save to, default (.GlobalEnv)}
}
\description{
Variable names defined by df
}
