% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{parse_cigar}
\alias{parse_cigar}
\title{Shift ribo-seq reads using cigar string}
\usage{
parse_cigar(cigar, shift, is_plus_strand)
}
\arguments{
\item{cigar}{the cigar of the reads}

\item{shift}{the shift as numeric}

\item{is_plus_strand}{logical}
}
\value{
the shifted read
}
\description{
Shift ribo-seq reads using cigar string
}
