% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{parseCigar}
\alias{parseCigar}
\title{Shift ribo-seq reads using cigar string}
\usage{
parseCigar(cigar, shift, is_plus_strand)
}
\arguments{
\item{cigar}{the cigar of the reads}

\item{shift}{the shift as integer}

\item{is_plus_strand}{logical}
}
\value{
the shifted read
}
\description{
Example if you want to change a read of length 20, by +12.
You need to account for gaps etc, this is done using the
cigar string of the read.
}
