% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{pasteDir}
\alias{pasteDir}
\title{A paste function for directories
Makes sure slashes are corrected, and not doubled.}
\usage{
pasteDir(...)
}
\arguments{
\item{...}{any amount of arguments that are possible to convert
to characters}
}
\value{
the pasted string
}
\description{
A paste function for directories
Makes sure slashes are corrected, and not doubled.
}
\keyword{internal}
