% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{percentage_to_ratio}
\alias{percentage_to_ratio}
\title{Convert percentage to ratio of 1}
\usage{
percentage_to_ratio(top_tx, cds, minimum.cds = 1000)
}
\arguments{
\item{top_tx}{numeric}

\item{cds}{GRangesList object}

\item{minimum.cds}{numeric, default 1000}
}
\value{
numeric, as ratio of 1
}
\description{
50 -> 0.5 etc, if length cds > minimum.cds
}
\keyword{internal}
