% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{periodicity}
\alias{periodicity}
\title{Find if there are periodicity in sample}
\usage{
periodicity(cds150_over, ends_uniq, top_tx)
}
\arguments{
\item{cds150_over}{a GRanges of tiled cds' to 150 length}

\item{ends_uniq}{a GRanges object with 5' ends of reads}

\item{top_tx}{numeric(10), which top percentage of transcripts to use.}
}
\value{
a logical, if it is periodic.
}
\description{
Find if there are periodicity in sample
}
