% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_plots.R
\name{plot.te_rna}
\alias{plot.te_rna}
\title{Translational efficiency plots}
\usage{
\method{plot}{te_rna}(
  dt,
  output.dir = NULL,
  filter.rfp = 1,
  filter.rna = 1,
  plot.title = "",
  width = 6,
  height = "auto",
  dot.size = 0.4
)
}
\arguments{
\item{dt}{a data.table with the results from \code{\link{te.table}}}

\item{output.dir}{a character path, default NULL(no save), or a directory
to save to a file will be called "TE_within.png"}

\item{filter.rfp}{numeric, default 1. What is the minimum fpkm value?}

\item{filter.rna}{numeric, default 1. What is the minimum fpkm value?}

\item{plot.title}{title for plots, usually name of experiment etc}

\item{width}{numeric, default 6 (in inches)}

\item{height}{a numeric, width of plot in inches. Default "auto".}

\item{dot.size}{numeric, default 0.4, size of point dots in plot.}
}
\value{
a ggplot object
}
\description{
Create TE plot of:\cr
- Within sample (TE log2 vs mRNA fpkm) ("default")\cr
}
\examples{
#df.rfp <- read.experiment("Riboseq")
#df.rna <- read.experiment("RNAseq")
#dt <- te.table(df.rfp, df.rna)
#plot.te_rna(dt)
}
