% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots_helpers.R
\name{plot_periodic_lengths_start}
\alias{plot_periodic_lengths_start}
\title{Plot the periods of reads around cds start codon}
\usage{
plot_periodic_lengths_start(start_df)
}
\arguments{
\item{start_df}{a data.frame with the periods}
}
\description{
Plot the periods of reads around cds start codon
}
