% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots_helpers.R
\name{plot_periodic_lengths_stop}
\alias{plot_periodic_lengths_stop}
\title{Plot the periods of reads around cds stop codon}
\usage{
plot_periodic_lengths_stop(stop_df)
}
\arguments{
\item{stop_df}{a data.frame with the periods}
}
\description{
Plot the periods of reads around cds stop codon
}
