% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_plots.R
\name{plotHelper}
\alias{plotHelper}
\title{Helper function for coverage plots}
\usage{
plotHelper(
  coverage,
  df,
  outdir,
  scores,
  returnCoverage = FALSE,
  title = "coverage metaplot",
  colors = c("skyblue4", "orange"),
  plotFunction = "windowCoveragePlot"
)
}
\arguments{
\item{coverage}{a data.table containing at least columns (count, position),
it is possible to have additionals: (genes, fraction, feature)}

\item{df}{an ORFik \code{\link{experiment}}}

\item{outdir}{directory to save to}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}

\item{returnCoverage}{(defualt: FALSE), return the ggplot object (TRUE)
or NULL (FALSE).}

\item{title}{Title to give plot}

\item{colors}{Which colors to use, default (skyblue4)}

\item{plotFunction}{Which plot function, default: windowCoveragePlot}
}
\value{
NULL (or ggplot object if returnCoverage is TRUE)
}
\description{
Should only be used internally
}
