% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{prettyScoring}
\alias{prettyScoring}
\title{Prettify scoring name}
\usage{
prettyScoring(scoring)
}
\arguments{
\item{scoring}{a character (the scoring)}
}
\value{
a new scoring name or the same if pretty
}
\description{
Prettify scoring name
}
\keyword{internal}
