% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudo.transform}
\alias{pseudo.transform}
\title{Transform object}
\usage{
pseudo.transform(x, scale = log2, by.reference = FALSE)
}
\arguments{
\item{x}{a numeric vector or data.frame/data.table of numeric columns}

\item{scale}{a function, default log2, which function to transform with.}

\item{by.reference}{logical, FALSE. if TRUE, update object by reference
if it is data.table.}
}
\value{
same object class as x, with transformed values
}
\description{
Similar to normal transform like log2 or log10.
But keep 0 values as 0, to avoid Inf values and negtive values
are made as -scale(abs(x)), to avoid NaN values.
}
