% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{pSitePlot}
\alias{pSitePlot}
\title{Plot area around TIS for p-shifted reads}
\usage{
pSitePlot(hitMap, length = 29, region = "start", output = NULL)
}
\arguments{
\item{hitMap}{a data.frame/data.table, given from metaWindow
(must have columns: position, score and frame)}

\item{length}{an integer (29), which length is this for?}

\item{region}{a character (start), either "start or "stop"}

\item{output}{character string (NULL), if set, saves the plot as pdf or png
to path given. If no format is given, is save as pdf.}
}
\value{
a ggplot object of the coverage plot, NULL if output is set,
then the plot will only be saved to location.
}
\description{
Usefull to validate p-shifting is correct
Can be used for any coverage of region around a point, like TIS, TSS,
stop site etc.
}
\details{
See vignette for example
}
\seealso{
Other coveragePlot: \code{\link{coverageHeatMap}},
  \code{\link{savePlot}}, \code{\link{windowCoveragePlot}}
}
\concept{coveragePlot}
