% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{QCfolder}
\alias{QCfolder}
\title{Get path to ORFik experiment QC folder}
\usage{
QCfolder(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get path to ORFik experiment QC folder
}
