% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCplots}
\alias{QCplots}
\title{Correlation and coverage plots for ORFikQC}
\usage{
QCplots(
  df,
  region = "mrna",
  stats_folder = paste0(dirname(df$filepath[1]), "/QC_STATS/")
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{region}{a character (default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{stats_folder}{directory to save, default:
paste0(dirname(df$filepath[1]), "/QC_STATS/")}
}
\value{
invisible(NULL) (objects stored to disc)
}
\description{
Correlation plots default to mRNA covering reads.
Meta plots defaults to leader, cds, trailer.\cr
Output will be stored in same folder as the
libraries in df.\cr
Correlation plots will be raw count correlation and
log2(count + 1) correlation between samples.
}
\details{
Is part of \code{\link{QCreport}}
}
\seealso{
Other QC report: 
\code{\link{QCreport}()},
\code{\link{QCstats}()}
}
\concept{QC report}
