% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCreport}
\alias{QCreport}
\title{Load QC Statistics report}
\usage{
QCreport(df, path = paste0(dirname(df$filepath[1]), "/QC_STATS/STATS.csv"))
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{path}{path to QC statistics report, default:
paste0(dirname(df$filepath[1]), "/QC_STATS/STATS.csv")}
}
\value{
data.table of QC report or NULL if not exists
}
\description{
Load QC Statistics report
}
\examples{
# df <- read.experiment("experiment/path")
# report <- QCreport(df)
}
\seealso{
Other ORFikQC: 
\code{\link{ORFikQC}()},
\code{\link{QCplots}()}
}
\concept{ORFikQC}
