% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCstats}
\alias{QCstats}
\title{Load QC Statistics report}
\usage{
QCstats(df, path = paste0(dirname(df$filepath[1]), "/QC_STATS/STATS.csv"))
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{path}{path to QC statistics report, default:
paste0(dirname(df$filepath[1]), "/QC_STATS/STATS.csv")}
}
\value{
data.table of QC report or NULL if not exists
}
\description{
Load QC Statistics report
}
\examples{
# df <- read.experiment("experiment/path")

## First make QC report
# QCreport(df)
# stats <- QCstats(df)
}
\seealso{
Other QC report: 
\code{\link{QCplots}()},
\code{\link{QCreport}()}
}
\concept{QC report}
