% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{r,covRle-method}
\alias{r,covRle-method}
\title{strandMode covRle}
\usage{
\S4method{r}{covRle}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
the forward RleList
}
\description{
strandMode covRle
}
