% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{rankOrder}
\alias{rankOrder}
\title{Get the orf rank in transcripts}
\usage{
rankOrder(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object with ORFs}
}
\value{
a numeric vector of integers
}
\description{
ig. second orf _2 -> 2
}
\references{
doi: 10.1074/jbc.R116.733899
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
