% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_IO.R
\name{read.experiment}
\alias{read.experiment}
\title{Read an ORFik experiment definition into an `experiment` object.}
\usage{
read.experiment(
  file,
  in.dir = ORFik::config()["exp"],
  validate = TRUE,
  output.env = .GlobalEnv
)
}
\arguments{
\item{file}{Relative path to an ORFik experiment, or a template data frame
from \code{\link{create.experiment}}. If a full path is supplied,
`in.dir` is ignored.}

\item{in.dir}{Directory containing experiment `.csv` files. Defaults to
\code{ORFik::config()["exp"]}, which by default is
`"~/Bio_data/ORFik_experiments/"`. Ignored when `file` is not a path or is
already a full path.}

\item{validate}{Logical, default `TRUE`. Validate that referenced library
files exist and that the experiment layout is consistent.}

\item{output.env}{Environment to associate with the experiment, default
`.GlobalEnv`. This can be changed later with `envExp(df) <- new.env()`.}
}
\value{
An ORFik \code{\link{experiment}} object.
}
\description{
The input must either be a valid ORFik experiment `.csv` file or a template
data frame produced by \code{\link{create.experiment}}.
}
\examples{
# From file
\dontrun{
# Read from file
df <- read.experiment(filepath) # <- valid ORFik .csv file
}
## Read from (create.experiment() template)
df <- ORFik.template.experiment()

## To save it, do:
# save.experiment(df, file = "path/to/save/experiment")
## You can then do:
# read.experiment("path/to/save/experiment")
# or (identical):
# read.experiment("experiment", in.dir = "path/to/save/")
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment}()},
\code{\link{ORFik.template.experiment.zf}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism,experiment-method}},
\code{\link{outputLibs}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
