% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_RDSQS}
\alias{read_RDSQS}
\title{Read RDS or QS format file}
\usage{
read_RDSQS(file, nthread = 5)
}
\arguments{
\item{file}{path to file with "rds" or "qs" file extension}

\item{nthread}{numeric, number of threads for qs::qread}
}
\value{
R object loaded from file
}
\description{
Read RDS or QS format file
}
\examples{
df <- ORFik::ORFik.template.experiment()
path <- ORFik:::countTablePath(df)
read_RDSQS(path)
}
