% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readBam}
\alias{readBam}
\title{Custome bam reader}
\usage{
readBam(path, chrStyle = NULL)
}
\arguments{
\item{path}{a character path to .bam file}

\item{chrStyle}{a GRanges object, or a characterPath (Default: NULL) to
get seqlevelsStyle from. Is chromosome 1 called chr1 or 1, is mitocondrial
chromosome called MT or chrM etc. If GRanges object, will use 1st seqlevel-
style if more are present. Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a GAlignment object of bam file
}
\description{
Safer version that handles the most important error done.
In the future will use a faster .bam loader for big .bam files in R.
}
