% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{readFootprints}
\alias{readFootprints}
\title{Get only unique mapping reads}
\usage{
readFootprints(file, filterNH1 = TRUE)
}
\arguments{
\item{file}{a bam or bed file path or GAlignment object}

\item{filterNH1}{logical(T), only keep unique reads, require a
meta column called NH}
}
\value{
a GAlignments object
}
\description{
Helper for detectRibosomeShifts
}
\examples{
bam <- system.file("extdata", "example.bam",
       package = "ORFik") ## location of the bam file
readFootprints(bam)

}
