% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{reduce_Keep_Attr}
\alias{reduce_Keep_Attr}
\title{Reduce GRanges / GRangesList}
\usage{
reduce_Keep_Attr(grl, keep.names = FALSE, drop.empty.ranges = FALSE,
  min.gapwidth = 1L, with.revmap = FALSE, with.inframe.attrib = FALSE,
  ignore.strand = FALSE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} or GRanges object}

\item{keep.names}{(FALSE) keep the names and meta columns of the GRangesList}

\item{drop.empty.ranges}{(FALSE) if a group is empty (width 0), delete it.}

\item{min.gapwidth}{(1L) how long gap can it be to say they belong together}

\item{with.revmap}{(FALSE) return info on which mapped to which}

\item{with.inframe.attrib}{(FALSE) For internal use.}

\item{ignore.strand}{(FALSE), can different strands be reduced together.}
}
\value{
A reduced GRangesList
}
\description{
Extends function \code{\link[GenomicRanges]{reduce}}
by trying to keep names and meta columns if it is a
GRangesList. If keep.names == F, it's just the normal
GenomicRanges::reduce.
}
\details{
Only tested for orfik, might not work for other naming conventions.
}
