% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{reduce_keep_naming}
\alias{reduce_keep_naming}
\title{Reduce a GRangesList}
\usage{
reduce_keep_naming(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}
}
\value{
A reduced GRangesList
}
\description{
Extends function \code{\link[GenomicRanges]{reduce}}
by trying to keep names and meta columns
}
