% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{refFolder}
\alias{refFolder}
\title{Get path to ORFik experiment genome reference folder}
\usage{
refFolder(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get path to ORFik experiment genome reference folder
}
