% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{regionPerReadLength}
\alias{regionPerReadLength}
\title{Find proportion of reads per position per read length in region}
\usage{
regionPerReadLength(
  grl,
  reads,
  acceptedLengths = NULL,
  withFrames = TRUE,
  weight = "score",
  BPPARAM = bpparam()
)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually either leaders, cds', 3' utrs or ORFs}

\item{reads}{a \code{\link{GAlignments}} or \code{\link{GRanges}} object of
RiboSeq, RnaSeq etc. Weigths for scoring is default the 'score'
column in 'reads'}

\item{acceptedLengths}{an integer vector (NULL), the read lengths accepted.
Default NULL, means all lengths accepted.}

\item{withFrames}{logical TRUE, add ORF frame (frame 0, 1, 2), starting
on first position of every grl.}

\item{weight}{(default: 'score'), if defined a character name
of valid meta column in subject. GRanges("chr1", 1, "+", score = 5),
would mean score column tells that this alignment region was found 5 times.
ORFik .bedo files, contains a score column like this.
As do CAGEr CAGE files and many other package formats.
You can also assign a score column manually.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam()}
}
\value{
a data.table with lengths by coverage.
}
\description{
This is defined as:
Given some transcript region (like CDS), get coverage per position.
}
\examples{
cds <- GRangesList(tx1 = GRanges("1", 100:129, "+"))
reads <- GRanges("1", seq(79,129, 3), "+")
dt <- regionPerReadLength(cds, reads, withFrames = TRUE)
dt
## Sum up reads in each frame per read length
dt[,.(sumCount = sum(count)),by = .(length, frame)]
}
\seealso{
Other coverage: 
\code{\link{coverageScorings}()},
\code{\link{metaWindow}()},
\code{\link{scaledWindowPositions}()},
\code{\link{windowPerReadLength}()}
}
\concept{coverage}
