% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{regroupRleList}
\alias{regroupRleList}
\title{Regroup rle from GRangesList}
\usage{
regroupRleList(rle, supergroups)
}
\arguments{
\item{rle}{A RleList to reduce groups on.}

\item{supergroups}{A GRangesList to group by}
}
\value{
A regrouped RleList
}
\description{
Almost direct copy of IRanges:::regroupBySupergroup.
But only works on rle and GRangesList.
This function will be removed if
IRanges:::regroupBySupergroup is exported.
}
