% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{remakeTxdbExonIds}
\alias{remakeTxdbExonIds}
\title{Get new exon ids after update of txdb}
\usage{
remakeTxdbExonIds(txList)
}
\arguments{
\item{txList}{a list, call of as.list(txdb)}
}
\value{
a new valid ordered list of exon ids (integer)
}
\description{
Get new exon ids after update of txdb
}
