% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{remove.experiments}
\alias{remove.experiments}
\title{Remove bam/bed/wig files load in R as variables}
\usage{
remove.experiments(df, envir = .GlobalEnv)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{envir}{environment to save to, default (.GlobalEnv)}
}
\value{
NULL (objects removed from envir specified)
}
\description{
Variable names defined by df, in envir defined
}
\examples{
df <- ORFik.template.experiment()
# Output to .GlobalEnv with:
# outputLibs(df)
# Then remove them with:
# remove.experiments(df)
}
