% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove.file_ext}
\alias{remove.file_ext}
\title{Remove file extension of path}
\usage{
remove.file_ext(path, basename = FALSE)
}
\arguments{
\item{path}{character path (allows multiple paths)}

\item{basename}{relative path (TRUE) or full path (FALSE)? (default: FALSE)}
}
\value{
character path without file extension
}
\description{
Allows removal of compression
}
\keyword{internal}
