% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{removeMetaCols}
\alias{removeMetaCols}
\title{Removes meta columns}
\usage{
removeMetaCols(grl)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} or GRanges object}
}
\value{
same type and structure as input without meta columns
}
\description{
Removes meta columns
}
\keyword{internal}
