% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uORF_helpers.R
\name{removeORFsWithinCDS}
\alias{removeORFsWithinCDS}
\title{Remove ORFs that are within cds}
\usage{
removeORFsWithinCDS(grl, cds)
}
\arguments{
\item{grl}{(GRangesList), the ORFs to filter}

\item{cds}{(GRangesList), the coding sequences (main ORFs on transcripts),
to filter against.}
}
\value{
(GRangesList) of filtered uORFs
}
\description{
Remove ORFs that are within cds
}
\seealso{
Other uorfs: 
\code{\link{addCdsOnLeaderEnds}()},
\code{\link{filterUORFs}()},
\code{\link{removeORFsWithSameStartAsCDS}()},
\code{\link{removeORFsWithSameStopAsCDS}()},
\code{\link{removeORFsWithStartInsideCDS}()},
\code{\link{uORFSearchSpace}()}
}
\concept{uorfs}
