% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uORF_helpers.R
\name{removeORFsWithStartInsideCDS}
\alias{removeORFsWithStartInsideCDS}
\title{Remove ORFs that have start site within the CDS}
\usage{
removeORFsWithStartInsideCDS(grl, cds)
}
\arguments{
\item{grl}{(GRangesList), the ORFs to filter}

\item{cds}{(GRangesList), the coding sequences (main ORFs on transcripts),
to filter against.}
}
\value{
(GRangesList) of filtered uORFs
}
\description{
Remove ORFs that have start site within the CDS
}
\seealso{
Other uorfs: 
\code{\link{addCdsOnLeaderEnds}()},
\code{\link{filterUORFs}()},
\code{\link{removeORFsWithSameStartAsCDS}()},
\code{\link{removeORFsWithSameStopAsCDS}()},
\code{\link{removeORFsWithinCDS}()},
\code{\link{uORFSearchSpace}()}
}
\concept{uorfs}
\keyword{internal}
