% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{removeTxdbExons}
\alias{removeTxdbExons}
\title{Remove exons in txList that are not in fiveUTRs}
\usage{
removeTxdbExons(txList, fiveUTRs)
}
\arguments{
\item{txList}{a list, call of as.list(txdb)}

\item{fiveUTRs}{a GRangesList of 5' leaders}
}
\value{
a list, modified call of as.list(txdb)
}
\description{
Remove exons in txList that are not in fiveUTRs
}
