% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{removeTxdbTranscripts}
\alias{removeTxdbTranscripts}
\title{Remove specific transcripts in txdb List}
\usage{
removeTxdbTranscripts(txList, fiveUTRs)
}
\arguments{
\item{txList}{a list, call of as.list(txdb)}

\item{fiveUTRs}{a GRangesList of 5' leaders}
}
\value{
a txList
}
\description{
Remove all transcripts, except the ones in fiveUTRs.
}
\keyword{internal}
