% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{rename.SRA.files}
\alias{rename.SRA.files}
\title{Rename SRA files from metadata}
\usage{
rename.SRA.files(files, new_names)
}
\arguments{
\item{files}{a character vector, all the files}

\item{new_names}{a character vector of new names or
a data.table with metadata to use to rename}
}
\value{
a character vector of new file names
}
\description{
Rename SRA files from metadata
}
