% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{resFolder,experiment-method}
\alias{resFolder,experiment-method}
\title{Get ORFik experiment main output directory}
\usage{
\S4method{resFolder}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get ORFik experiment main output directory
}
