% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{resFolder}
\alias{resFolder}
\title{Get ORFik experiment main output directory}
\usage{
resFolder(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get ORFik experiment main output directory
}
