% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defining_ORFs_helpers.R
\name{resize_ORF}
\alias{resize_ORF}
\title{Resizes down ORF to the desired length, removing inside. Preserves exons.}
\usage{
resize_ORF(grangesObj, orf_goal_length)
}
\description{
@export
 @param grangesObj A GRanges object of ORF.
 @param orf_goal_length numeric. Desired length of ORF.
 @return GRanges object of resized ORF
}

