% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{resizeORF}
\alias{resizeORF}
\title{Resizes down ORF to the desired length, removing inside. Preserves exons.}
\usage{
resizeORF(grangesObj, orf_goal_length)
}
\arguments{
\item{grangesObj}{A GRanges object of ORF.}

\item{orf_goal_length}{numeric. Desired length of ORF.}
}
\value{
GRanges object of resized ORF
}
\description{
Resizes down ORF to the desired length, removing inside. Preserves exons.
}
