% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{restrictTSSByUpstreamLeader}
\alias{restrictTSSByUpstreamLeader}
\title{Restrict extension of 5' UTRs to closest upstream TSS}
\usage{
restrictTSSByUpstreamLeader(fiveUTRs)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}
}
\value{
GRangesList object of restricted fiveUTRs
}
\description{
That is minimum(CAGE extension, upstream TSS)
}
