% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{revElementsF}
\alias{revElementsF}
\title{Reverse elements within list}
\usage{
revElementsF(x)
}
\arguments{
\item{x}{RleList}
}
\value{
a RleList (reversed inside list elements)
}
\description{
A faster version of S4Vectors::revElements
}
