% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboORFs_detect_and_classify.R
\name{riboORFs}
\alias{riboORFs}
\title{Load Predicted translons}
\usage{
riboORFs(df, type = "table", folder = riboORFsFolder(df))
}
\arguments{
\item{df}{ORFik experiment}

\item{type}{default "table", alternatives: c("table", "ranges_candidates",
"ranges_predictions", "predictions")}

\item{folder}{base folder to check for computed results, default:
riboORFsFolder(df)}
}
\value{
a data.table, GRangesList or list of logical vector depending on input
}
\description{
Load Predicted translons
}
\examples{
df <- ORFik.template.experiment()
df <- df[df$libtype == "RFP",][c(1,2),]
# riboORFs(df) # Works when you have run prediction
}
