% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plots.R
\name{RiboQC.plot}
\alias{RiboQC.plot}
\title{Quality control for pshifted Ribo-seq data}
\usage{
RiboQC.plot(
  df,
  output.dir = file.path(dirname(df$filepath[1]), "QC_STATS/"),
  width = 6.6,
  height = 4.5,
  type = "pshifted",
  weight = "score",
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE)
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{output.dir}{directory to save plot,
default: file.path(dirname(df$filepath[1]), "QC_STATS/"). If NULL will not save.}

\item{width}{width of plot, default 6.6 (in inches)}

\item{height}{height of plot, default 4.5 (in inches)}

\item{type}{type of library loaded, default pshifted,
warning if not pshifted might crash if too many read lengths!}

\item{weight}{which column in reads describe duplicates, default "score".}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}}
}
\value{
ggplot object as a grid
}
\description{
Quality control for pshifted Ribo-seq data
}
\examples{
df <- ORFik.template.experiment()
df <- df[3,] #lets only p-shift RFP sample at index 3
#shiftFootprintsByExperiment(df)
#RiboQC.plot(df)
}
