% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{riboSeqReadWidths}
\alias{riboSeqReadWidths}
\title{Get Ribo-seq widths}
\usage{
riboSeqReadWidths(reads)
}
\arguments{
\item{reads}{a GRanges or GAlignment object.}
}
\value{
an integer vector of widths
}
\description{
Input a ribo-seq object and get width of reads,
 if input is p-shifted and GRanges, the "$score column" must
 exist, and contain the original read widths.
}
