% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{ribosomeReleaseScore}
\alias{ribosomeReleaseScore}
\title{Ribosome Release Score (RRS)}
\usage{
ribosomeReleaseScore(grl, RFP, GtfOrThreeUtrs, RNA = NULL)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{GtfOrThreeUtrs}{if Gtf: a TxDb object of a gtf file,
if ThreeUtrs: a GrangesList of transcripts, called from:
threeUTRsByTranscript(Gtf, use.names = T)}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}
}
\value{
a named vector of numeric values of scores, NA means that
 no 3' utr was found for that transcript.
}
\description{
Is defined as (RPFs over ORF)/(RPFs over 3' utrs),
normalized by lengths
, if RNA is added as argument, normalize by RNA counts over areas too
to justify location of 3' utrs
It can be seen as a ribosome stalling feature.
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1016/j.cell.2013.06.009
}
\examples{
ORF <- GRanges(seqnames = "1", ranges = IRanges(
               start = c(1, 10, 20),
               end = c(5, 15, 25)),
               strand = "+")
grl <- GRangesList(tx1_1 = ORF)
threeUTRs <- GRangesList(tx1 = GRanges("1", IRanges(40, 50), "+"))
RFP <- GRanges("1", IRanges(25, 25),"+")
RNA <- GRanges("1", IRanges(1, 50),"+")
ribosomeReleaseScore(grl, RFP, threeUTRs, RNA)
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
