% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{RibosomeStallingScore}
\alias{RibosomeStallingScore}
\title{Ribosome Stalling Score (RSS)}
\usage{
RibosomeStallingScore(grl, RFP)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}
}
\value{
a named vector of numeric values of scores
}
\description{
Is defined as (RPFs over ORF stop sites)/(RPFs over ORFs),
normalized by lengths
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1016/j.cels.2017.08.004
}
\details{
For a more accurate analysis see article:  10.1016/j.cels.2017.08.004.
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
