% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{runIDs}
\alias{runIDs}
\title{Get ORFik experiment library folder}
\usage{
runIDs(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character vector of runIDs, "" if not existing.
}
\description{
Get ORFik experiment library folder
}
