% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{save.experiment}
\alias{save.experiment}
\title{Save experiment to disc}
\usage{
save.experiment(df, file)
}
\arguments{
\item{df}{an ORFik experiment data.frame}

\item{file}{name of file to save df as}
}
\description{
Save experiment to disc
}
