% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{scaledWindowPositions}
\alias{scaledWindowPositions}
\title{Scale windows to a meta window of size}
\usage{
scaledWindowPositions(grl, reads, scaleTo = 100, scoring = "mean")
}
\arguments{
\item{grl}{GRangesList or GRanges of your ranges}

\item{reads}{GRanges object of your reads.}

\item{scaleTo}{an integer (100), if windows have different size,
a meta window can not directly be created, since a meta window must
have equal size for all windows. Rescale all windows to scaleTo.
i.e c(1,2,3) -> size 2 -> c(1, mean(2,3)) etc.}

\item{scoring}{a character, one of (mean, median, sum)}
}
\value{
A data.frame or data.table with average counts (avg_counts) of
reads mapped to positions (position) specified in windows along with
frame (frame).
}
\description{
For example scale a coverage plot of a all human CDS to width 100
}
\details{
Nice for making metaplots
}
