% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{seqinfo,covRle-method}
\alias{seqinfo,covRle-method}
\title{Seqinfo covRle
Extracted from forward RleList}
\usage{
\S4method{seqinfo}{covRle}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
integer vector with names
}
\description{
Seqinfo covRle
Extracted from forward RleList
}
