% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{seqinfo,experiment-method}
\alias{seqinfo,experiment-method}
\title{Seqinfo ORFik experiment
Extracted from fasta genome index}
\usage{
\S4method{seqinfo}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
integer vector with names
}
\description{
Seqinfo ORFik experiment
Extracted from fasta genome index
}
