% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{seqlevels,covRle-method}
\alias{seqlevels,covRle-method}
\title{Seqlevels covRle
Extracted from forward RleList}
\usage{
\S4method{seqlevels}{covRle}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
integer vector with names
}
\description{
Seqlevels covRle
Extracted from forward RleList
}
