% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{seqlevels,covRleList-method}
\alias{seqlevels,covRleList-method}
\title{Seqlevels covRleList
Extracted from forward RleList}
\usage{
\S4method{seqlevels}{covRleList}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
integer vector with names
}
\description{
Seqlevels covRleList
Extracted from forward RleList
}
