% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{seqlevels,experiment-method}
\alias{seqlevels,experiment-method}
\title{Seqlevels ORFik experiment
Extracted from fasta genome index}
\usage{
\S4method{seqlevels}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
integer vector with names
}
\description{
Seqlevels ORFik experiment
Extracted from fasta genome index
}
