% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{seqnames,experiment-method}
\alias{seqnames,experiment-method}
\title{Seqnames ORFik experiment
Extracted from fasta genome index}
\usage{
\S4method{seqnames}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
integer vector with names
}
\description{
Seqnames ORFik experiment
Extracted from fasta genome index
}
