% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints.R
\name{shiftFootprintsByExperiment}
\alias{shiftFootprintsByExperiment}
\title{Shift footprints of each file in experiment}
\usage{
shiftFootprintsByExperiment(
  df,
  out.dir = pasteDir(dirname(df$filepath[1]), "/pshifted/"),
  start = TRUE,
  stop = FALSE,
  top_tx = 10L,
  minFiveUTR = 30L,
  minCDS = 150L,
  minThreeUTR = 30L,
  firstN = 150L,
  min_reads = 1000,
  accepted.lengths = 26:34,
  output_format = c("ofst", "wig"),
  BPPARAM = bpparam(),
  tx = NULL,
  log = TRUE,
  heatmap = FALSE,
  must.be.periodic = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{output directory for files,
default: dirname(df$filepath[1]), making a /pshifted
folder at that location}

\item{start}{(logical) Whether to include predictions based on the start
codons. Default TRUE.}

\item{stop}{(logical) Whether to include predictions based on the stop
codons. Default FASLE. Only use if there exists 3' UTRs for the annotation.
If peridicity around stop codon is stronger than at the start codon, use
stop instead of start region for p-shifting.}

\item{top_tx}{(integer), default 10. Specify which \% of the top covered by RiboSeq
reads transcripts to use for estimation of the shifts. By default we take top 10%
top covered transcripts as they represent less noisy dataset. This is only
applicable when there are more than 1000 transcripts.}

\item{minFiveUTR}{(integer) minimum bp for 5' UTR during filtering for the
transcripts. Set to NULL if no 5' UTRs exists for annotation.}

\item{minCDS}{(integer) minimum bp for CDS during filtering for the
transcripts}

\item{minThreeUTR}{(integer) minimum bp for 3' UTR during filtering for the
transcripts. Set to NULL if no 3' UTRs exists for annotation.}

\item{firstN}{(integer) Represents how many bases of the transcripts
downstream of start codons to use for initial estimation of the
periodicity.}

\item{min_reads}{default (1000), how many reads must a read-length have to
be considered for periodicity.}

\item{accepted.lengths}{accepted readlengths, default 26:34, usually ribo-seq
is strongest between 27:32.}

\item{output_format}{default c("ofst", "wig"), use export.ofst or
wiggle format (wig) using \code{\link{export.wiggle}} ? Default is both.
The wig format version can be used in IGV, the score column is counts of that
read with that read length, the cigar reference width is lost,
ofst is much faster to save and load in R, and retain cigar reference width,
but can not be used in IGV. \cr You can also do bedoc format, bed format
keeping cigar: \code{\link{export.bedoc}}. bedoc is usually not used for
p-shifting.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam()}

\item{tx}{a GRangesList, if you do not have 5' UTRs in annotation, send
your own version. Example: extendLeaders(tx, 30)
Where 30 bases will be new "leaders". Since each original transcript was
either only CDS or non-coding (filtered out).}

\item{log}{logical, default (TRUE), output a log file with parameters used and
a .rds file with all shifts per library
(can be loaded with \code{\link{shifts.load}})}

\item{heatmap}{a logical or character string, default FALSE.
If TRUE, will plot heatmap of
raw reads before p-shifting to console, to see if shifts given make sense.
You can also set a filepath to save the file there.}

\item{must.be.periodic}{logical TRUE, if FALSE will not filter on
periodic read lengths. (The Fourier transform filter will be skipped).
This is useful if you are not going to do periodicity analysis, that is:
for you more coverage depth (more read lengths)
is more important than only keeping the high quality periodic read lengths.}

\item{verbose}{logical FALSE, if TRUE report details of
read length filtering (periodogram) and
change point analysis per read length.}
}
\value{
NULL (Objects are saved to out.dir/pshited/"name_pshifted.ofst",
wig, bedo or .bedo)
}
\description{
A function that combines the steps of periodic read length detection,
p-site shift detection and p-shifting into 1 function.
For more details, see: \code{\link{detectRibosomeShifts}}\cr
Saves files to a specified location as .ofst and .wig,
The .ofst file will include a score column containing read width. \cr
The .wig files, will be saved in pairs of +/- strand, and score column
will be replicates of reads starting at that position,
score = 5 means 5 reads.\cr
Remember that different species might have different default Ribosome
read lengths, for human, mouse etc, normally around 27:30.
}
\examples{
df <- ORFik.template.experiment()
df <- df[3,] #lets only p-shift RFP sample at index 3
# If you want to check it in IGV do:
shiftFootprintsByExperiment(df)
# Then use the .wig files that are created, which are readable in IGV.
# If you only need in R, do: (then you get no .wig files)
#shiftFootprintsByExperiment(df, output_format = "ofst")
}
\references{
https://bmcgenomics.biomedcentral.com/articles/10.1186/s12864-018-4912-6
}
\seealso{
Other pshifting: 
\code{\link{changePointAnalysis}()},
\code{\link{detectRibosomeShifts}()},
\code{\link{shiftFootprints}()}
}
\concept{pshifting}
