% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{shifts.load}
\alias{shifts.load}
\title{Load the shifts from experiment}
\usage{
shifts.load(
  df,
  path = pasteDir(dirname(df$filepath[1]), "/pshifted/shifting_table.rds")
)
}
\arguments{
\item{path}{path to .rds file containing the shifts as a list,
one list element per shifted bam file.}
}
\value{
a list of the shifts, one list element per shifted bam file.
}
\description{
Defaults to pshifts
}
\examples{
df <- ORFik.template.experiment()
# subset on Ribo-seq
df <- df[df$libtype == "RFP",]
#shiftFootprintsByExperiment(df)
#shifts.load(df)
}
