% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{show,covRle-method}
\alias{show,covRle-method}
\title{covRle show definition}
\usage{
\S4method{show}{covRle}(object)
}
\arguments{
\item{object}{a\code{\link{covRle}}}
}
\value{
print state of covRle
}
\description{
Show a simplified version of the covRle
}
