% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{show,covRleList-method}
\alias{show,covRleList-method}
\title{covRleList show definition}
\usage{
\S4method{show}{covRleList}(object)
}
\arguments{
\item{object}{a\code{\link{covRleList}}}
}
\value{
print state of covRleList
}
\description{
Show a simplified version of the covRleList.
}
