% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{show,experiment-method}
\alias{show,experiment-method}
\title{experiment show definition}
\usage{
\S4method{show}{experiment}(object)
}
\arguments{
\item{object}{an ORFik \code{\link{experiment}}}
}
\value{
print state of experiment
}
\description{
Show a simplified version of experiment.
The show function simplifies the view so that any
column of data (like replicate or stage) is not shown, if all
values are identical in that column. Filepath is also never shown.
}
