% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{simpleLibs}
\alias{simpleLibs}
\title{Will make a simplified version of NGS libraries}
\usage{
simpleLibs(
  df,
  out.dir = paste0(dirname(df$filepath[1]), "/bedo/"),
  addScoreColumn = TRUE,
  addSizeColumn = TRUE,
  must.overlap = NULL,
  method = "None"
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default:
paste0(dirname(df$filepath[1]), "/bedo/"),
if it is NULL, it will just reassign R objects to simplified libraries.}

\item{addScoreColumn}{logical (FALSE), if TRUE, add a score column that
sums up the hits per unique range This will make each read unique, so
that each read is 1 time, and score column gives the number of hits.
A useful compression. If addSizeColumn is FALSE, it will not differentiate
between reads with same start and stop, but different length.}

\item{addSizeColumn}{logical (FALSE), if TRUE, add a size column that
for each read, that gives original width of read. Useful if you need
original read lengths. This takes care of soft clips etc.}

\item{must.overlap}{default (NULL), else a GRanges / GRangesList object, so
only reads that overlap (must.overlap) are kept. This is useful when you
only need the reads over transcript annotation or subset etc.}

\item{method}{character, default "None", the method to reduce ranges,
for more info see \code{\link{convertToOneBasedRanges}}}
}
\value{
NULL (saves files to disc or R .GlobalEnv)
}
\description{
Export files as .bedo files. It is a bed file with 2 score columns.
Gives a massive speedup when cigar strings are not needed.
}
\details{
See \code{\link{export.bedo}} for information on file format
}
