% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_helpers.R
\name{sortPerGroup}
\alias{sortPerGroup}
\title{Sort a GRangesList}
\usage{
sortPerGroup(grl, ignore.strand = FALSE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{ignore.strand}{a boolean, should minus strands be sorted from highest
to lowest(T)}
}
\value{
an equally named GRangesList, where each group is
 sorted within group.
}
\description{
A faster, more versatile reimplementation of GenomicRanges::sort()
for GRangesList, which works poorly for > 10k groups
It sorts each group, where + strands increasing by start
and - strands decreasing by ends.
}
\details{
If you want unusual sorts, like reversed ordering,
use \code{\link{gSort}} directly.
}
