% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{stageNames}
\alias{stageNames}
\title{Get stage name variants}
\usage{
stageNames()
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Example: 64Cell stage is same as 2 hours post fertilization,
so all 2hpf will be converted to 64Cell etc.
}
\seealso{
Other experiment_naming: 
\code{\link{cellLineNames}()},
\code{\link{conditionNames}()},
\code{\link{libNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
