% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.index}
\alias{STAR.index}
\title{Create STAR genome index}
\usage{
STAR.index(
  arguments,
  output.dir = paste0(dirname(arguments[1]), "/STAR_index/"),
  star.path = STAR.install(),
  max.cpus = min(90, detectCores() - 1),
  wait = TRUE,
  script = system.file("STAR_Aligner", "STAR_MAKE_INDEX.sh", package = "ORFik")
)
}
\arguments{
\item{arguments}{a named character vector containing paths wanted to
use for index creation. They must be named correctly:
names must be a subset of:
c("gtf", "genome", "phix", "rRNA", "tRNA","ncRNA")}

\item{output.dir}{directory to save indices, default:
paste0(dirname(arguments[1]), "/STAR_index/"), where arguments is the
arguments input for this function.}

\item{star.path}{path to STAR, default: STAR.install(),
if you don't have STAR installed at default location, it will install it there,
set path to a runnable star if you already have it.}

\item{max.cpus}{integer, default: min(90, detectCores() - 1),
number of threads to use. Default is minimum of 90 and maximum cores - 1}

\item{wait}{a logical (not \code{NA}) indicating whether the \R
    interpreter should wait for the command to finish, or run it
    asynchronously.  This will be ignored (and the interpreter will
    always wait) if \code{intern = TRUE}. When running the command
    asynchronously, no output will be displayed on the \code{Rgui}
    console in Windows (it will be dropped, instead).}

\item{script}{location of STAR index script,
default internal ORFik file. You can change it and give your own if you
need special alignments.}
}
\value{
output.dir, can be used as as input for STAR.align..
}
\description{
Used as reference when aligning data \cr
Get genome and gtf by running getGenomeAndFasta()
}
\details{
Can only run on unix systems (Linux and Mac), and requires
minimum 30GB memory on genomes like human, rat, zebrafish etc.
}
\examples{
# In this argument specify the one you want
#arguments <- c(path.GTF, path.genome, path.phix, path.rrna, path.trna, path.ncrna)
#names(arguments) <- c("gtf", "genome", "phix", "rRNA", "tRNA","ncRNA")
#STAR.index(arguments, "output.dir")
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.install}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{getGenomeAndFasta}()},
\code{\link{install.fastp}()}
}
\concept{STAR}
