% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.install}
\alias{STAR.install}
\title{Download and prepare STAR}
\usage{
STAR.install(folder = "~/bin", version = "2.7.4a")
}
\arguments{
\item{folder}{path to folder for download, fille will be named
"STAR-version", where version is version wanted.}

\item{version}{default "2.7.4a"}
}
\value{
path to runnable STAR
}
\description{
Will not run "make", only use precompiled STAR file.\cr
Only works for Linux and Mac (unix systems)
}
\examples{
#STAR.install("~/bin", version = "2.7.4a")
}
\references{
https://www.ncbi.nlm.nih.gov/pubmed/23104886
}
