% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{start,GRanges-method}
\alias{start,GRanges-method}
\title{Get starts of GRanges}
\usage{
\S4method{start}{GRanges}(x)
}
\arguments{
\item{x}{a GRanges}
}
\value{
an integer (length equal to x)
}
\description{
Faster version than S4Vector generic caller
}
