% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{start_definition}
\alias{start_definition}
\title{Returns start definition according to
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.}
\usage{
start_definition(transl_table)
}
\arguments{
\item{transl_table}{numeric.  NCBI genetic code number for translation.}
}
\value{
A string of START sies separatd with "|".
}
\description{
Returns start definition according to
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.
}
