% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq_features.R
\name{startCodonCoverage}
\alias{startCodonCoverage}
\title{Start codon coverage}
\usage{
startCodonCoverage(grl, RFP, is.sorted = TRUE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually either leaders, cds', 3' utrs or ORFs}

\item{RFP}{ribo seq reads as GAlignment, GRanges or GRangesList object}

\item{is.sorted}{logical (F), is grl sorted.}
}
\value{
an numeric vector of counts
}
\description{
Get the number of reads in start codon for each ORF
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{distToTSS}},
  \code{\link{entropy}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{initiationScore}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{startRegionCoverage}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
\concept{features}
