% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{startDefinition}
\alias{startDefinition}
\title{Returns start codon definitions}
\usage{
startDefinition(transl_table)
}
\arguments{
\item{transl_table}{numeric.  NCBI genetic code number for translation.}
}
\value{
A string of START sites separatd with "|".
}
\description{
According to:
<http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/
index.cgi?chapter=tgencodes#SG1>
ncbi genetic code number for translation. This version is a cleaned up
version, unknown indices removed.
}
\examples{
startDefinition
startDefinition(1)

}
\seealso{
Other findORFs: 
\code{\link{findMapORFs}()},
\code{\link{findORFsFasta}()},
\code{\link{findORFs}()},
\code{\link{findUORFs}()},
\code{\link{stopDefinition}()}
}
\concept{findORFs}
