% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{startRegion}
\alias{startRegion}
\title{Start region as GRangesList}
\usage{
startRegion(grl, tx = NULL, is.sorted = TRUE, upstream = 2L, downstream = 2L)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually either leaders, cds', 3' utrs or ORFs}

\item{tx}{default NULL, a GRangesList of transcripts or (container region),
names of tx must contain all grl names. The names of grl can also be the
ORFik orf names. that is "txName_id"}

\item{is.sorted}{logical (TRUE), is grl sorted.}

\item{upstream}{an integer (2), relative region to get upstream from.}

\item{downstream}{an integer (2), relative region to get downstream from}
}
\value{
a GRanges, or GRangesList object if any group had > 1 exon.
}
\description{
Get the start region of each ORF. If you want the start codon only,
set upstream = 0 or just use \code{\link{startCodons}}.
Standard is 2 upstream and 2 downstream, a width 5 window centered at
start site. since p-shifting is not 100% accurate, this window is
usually the reads from the start site.
}
\details{
If tx is null, then upstream will be forced to 0 and downstream to
a maximum of grl width (3' UTR end for mRNAs).
Since there is no reference for splicing.
}
\seealso{
Other features: 
\code{\link{computeFeaturesCage}()},
\code{\link{computeFeatures}()},
\code{\link{countOverlapsW}()},
\code{\link{disengagementScore}()},
\code{\link{distToCds}()},
\code{\link{distToTSS}()},
\code{\link{entropy}()},
\code{\link{floss}()},
\code{\link{fpkm_calc}()},
\code{\link{fpkm}()},
\code{\link{fractionLength}()},
\code{\link{initiationScore}()},
\code{\link{insideOutsideORF}()},
\code{\link{isInFrame}()},
\code{\link{isOverlapping}()},
\code{\link{kozakSequenceScore}()},
\code{\link{orfScore}()},
\code{\link{rankOrder}()},
\code{\link{ribosomeReleaseScore}()},
\code{\link{ribosomeStallingScore}()},
\code{\link{startRegionCoverage}()},
\code{\link{stopRegion}()},
\code{\link{subsetCoverage}()},
\code{\link{translationalEff}()}
}
\concept{features}
