% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq_features.R
\name{startRegionCoverage}
\alias{startRegionCoverage}
\title{Start region coverage}
\usage{
startRegionCoverage(grl, RFP, is.sorted = TRUE, window = 10L)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually either leaders, cds', 3' utrs or ORFs}

\item{RFP}{ribo seq reads as GAlignment, GRanges or GRangesList object}

\item{is.sorted}{logical (T), is grl sorted.}

\item{window}{integer (10), how many bases downstream to use.}
}
\value{
an numeric vector of counts
}
\description{
Get the number of reads in the start region of each ORF.
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{distToTSS}},
  \code{\link{entropy}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{initiationScore}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{startCodonCoverage}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
\concept{features}
