% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{startRegionString}
\alias{startRegionString}
\title{Get start region as DNA-strings per GRanges group}
\usage{
startRegionString(grl, tx, faFile, upstream = 20, downstream = 20)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} of ranges to find regions in.}

\item{tx}{a GRangesList of transcripts or (container region)}

\item{faFile}{a FaFile from the fasta file, see ?FaFile.
Can also be path to fastaFile with fai file in same dir.}

\item{upstream}{an integer vector, relative region to get upstream from.}

\item{downstream}{an integer, relative region to get downstream from}
}
\value{
a character vector of start regions
}
\description{
One window per start site, if upstream and downstream are both 0, then
only the startsite is returned.
}
