% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{stop_ribo}
\alias{stop_ribo}
\title{Get number of ribo reads per tile position of cds stop window}
\usage{
stop_ribo(stop, ends_uniq, top_tx)
}
\arguments{
\item{stop}{a GRangesList or logical, of tiled cds
stop windows of 60 length}

\item{ends_uniq}{a GRanges object with 5' ends of reads}

\item{top_tx}{numeric(10), which top percentage of transcripts to use.}
}
\value{
a list of reads per position
}
\description{
Get number of ribo reads per tile position of cds stop window
}
