% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{stopDefinition}
\alias{stopDefinition}
\title{Returns stop definition according to
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.}
\usage{
stopDefinition(transl_table)
}
\arguments{
\item{transl_table}{numeric.  NCBI genetic code number for translation.}
}
\value{
A string of STOP sies separatd with "|".
}
\description{
Returns stop definition according to
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.
}
