% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{stopSites}
\alias{stopSites}
\title{get the Stop sites from a GRangesList of orfs grouped by orfs}
\usage{
stopSites(grl, asGR = FALSE, keep.names = FALSE, is.sorted = FALSE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object}

\item{asGR}{a boolean, return as GRanges object}

\item{keep.names}{if asGR is False, do you still want
to keep a named vector}

\item{is.sorted}{a speedup, if you know the ranges are sorted}
}
\value{
if asGR is False, a vector, if True a GRanges object
}
\description{
In ATGTTTTGC, get the position of the C.
}
