% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{strandMode,covRleList-method}
\alias{strandMode,covRleList-method}
\title{strandMode covRle}
\usage{
\S4method{strandMode}{covRleList}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
integer vector with names
}
\description{
strandMode covRle
}
