% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{subchar}
\alias{subchar}
\title{Replaces string at pos with N.}
\usage{
subchar(string, pos)
}
\arguments{
\item{string}{string.}

\item{pos}{numeric vector.}
}
\value{
string with swapped N at pos
}
\description{
Replaces string at pos with N.
}

