% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{subset_coverage}
\alias{subset_coverage}
\title{Subset GRanges to get coverage.}
\usage{
subset_coverage(cov, y)
}
\arguments{
\item{cov}{A coverage object from coverage()}

\item{y}{GRanges object for which coverage should be extracted}
}
\value{
numeric vector of coverage of input GRanges object
}
\description{
GRanges object should be beforehand
 tiled to size of 1. This subsetting takes account for strand.
}
\examples{
#subset_coverage(x)
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}}, \code{\link{te}}
}
