% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{subset_to_frame}
\alias{subset_to_frame}
\title{Subset GRanges to get desired frame. GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.}
\usage{
subset_to_frame(x, frame)
}
\arguments{
\item{x}{A tiled to size of 1 GRanges object}

\item{frame}{A numeric indicating which frame to extract}
}
\value{
GRanges object reduced to only first frame
}
\description{
Subset GRanges to get desired frame. GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.
}
\examples{
#subset_to_frame(x, 1)

}
