% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{subset_to_stop}
\alias{subset_to_stop}
\title{Subset GRanges to get stop codons. GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.}
\usage{
subset_to_stop(x)
}
\arguments{
\item{x}{A tiled to size of 1 GRanges object}
}
\value{
GRanges object reduced to only stop codon
}
\description{
Subset GRanges to get stop codons. GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.
}
\examples{
#subset_to_stop(x)

}
