% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{subsetCoverage}
\alias{subsetCoverage}
\title{Subset GRanges to get coverage.}
\usage{
subsetCoverage(cov, y)
}
\arguments{
\item{cov}{A coverage object from coverage()}

\item{y}{GRanges object for which coverage should be extracted}
}
\value{
numeric vector of coverage of input GRanges object
}
\description{
GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{distToTSS}},
  \code{\link{entropy}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{initiationScore}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{startCodonCoverage}},
  \code{\link{startRegionCoverage}},
  \code{\link{translationalEff}}
}
\concept{features}
