% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{sum,covRle-method}
\alias{sum,covRle-method}
\title{sum covRle}
\usage{
\S4method{sum}{covRle}(x)
}
\arguments{
\item{x}{a covRle object}
}
\value{
an integer, sum of coverage per chromosomes in covRle object
}
\description{
Sum coverage per chromosome
}
