% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{tissueNames}
\alias{tissueNames}
\title{Get tissue name variants}
\usage{
tissueNames()
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: testis is main naming, but a variant is testicles.
testicles will then be renamed to testis.
}
\seealso{
Other experiment_naming: 
\code{\link{cellLineNames}()},
\code{\link{conditionNames}()},
\code{\link{libNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{stageNames}()}
}
\concept{experiment_naming}
