% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{toGR}
\alias{toGR}
\title{Converts different type of files to Granges
Only Accepts bed files for now}
\usage{
toGR(x, bed6 = TRUE)
}
\arguments{
\item{x}{A file containing column-based counts}

\item{bed6}{If bed6, no meta column is added}
}
\description{
Converts different type of files to Granges
Only Accepts bed files for now
}
