% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{TxLen}
\alias{TxLen}
\title{get transcript lengths
a helper function for easy length retrieval}
\usage{
TxLen(Gtf, changedFiveUTRs = NULL)
}
\arguments{
\item{Gtf}{a TxDb object of a gtf file}

\item{changedFiveUTRs}{a GRangesList object of leaders.
NB! only add this if you used cage data or other things to change the
leaders, therefor we need it to update transcript lengths.}
}
\description{
get transcript lengths
a helper function for easy length retrieval
}
