% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{txNamesWithLeaders}
\alias{txNamesWithLeaders}
\title{Get the transcripts that have minimum lengths of leaders and cds.}
\usage{
txNamesWithLeaders(txdb, minFiveUTR = 30L, minCDS = 150L,
  minThreeUTR = 30L)
}
\arguments{
\item{txdb}{a TxDb object from gtf}

\item{minFiveUTR}{(integer) minimum bp for 5' UTR during filtering for the
transcripts}

\item{minCDS}{(integer) minimum bp for CDS during filtering for the
transcripts}

\item{minThreeUTR}{(integer) minimum bp for 3' UTR during filtering for the
transcripts}
}
\value{
a character vector of valid tramscript names
}
\description{
Filter transcripts to those whohave 5' UTR, CDS, 3' UTR of
some lengths, pick the longest per gene.
}
\examples{
gtf_file <- system.file("extdata", "annotations.gtf", package = "ORFik")
txdb <- GenomicFeatures::makeTxDbFromGFF(gtf_file, format = "gtf")
txNames <- txNamesWithLeaders(txdb)

}
