% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{txSeqsFromFa}
\alias{txSeqsFromFa}
\title{Get transcript sequence from a GrangesList and a faFile or BSgenome}
\usage{
txSeqsFromFa(grl, faFile, is.sorted = FALSE)
}
\arguments{
\item{grl}{a GRangesList object}

\item{faFile}{FaFile or BSgenome used to find the transcripts,}

\item{is.sorted}{a speedup, if you know the ranges are sorted}
}
\value{
a DNAStringSet of the transcript sequences
}
\description{
A small safety wrapper around \code{\link{extractTranscriptSeqs}}
}
\seealso{
Other ExtendGenomicRanges: \code{\link{asTX}},
  \code{\link{coveragePerTiling}},
  \code{\link{overlapsToCoverage}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{windowPerGroup}}
}
\concept{ExtendGenomicRanges}
