% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{uniqueMappers,experiment-method}
\alias{uniqueMappers,experiment-method}
\title{Get ORFik uniqueMappers status}
\usage{
\S4method{uniqueMappers}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a logical (length 1)
}
\description{
Do you want to load/save libraries with unique mappers only,
for bam it subsets from file, for other formats it presumes a
directory './unique_mappers' relative to bam directory.
}
