% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{uniqueMappers<-}
\alias{uniqueMappers<-}
\title{Set ORFik uniqueMappers status}
\usage{
uniqueMappers(x) <- value
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}

\item{value}{a logical (length 1) (NA values not allowed)}
}
\value{
an ORFik \code{\link{experiment}} with updated uniqueMappers
}
\description{
Do you want to load/save libraries with unique mappers only,
for bam it subsets from file, for other formats it presumes a
directory './unique_mappers' relative to bam directory.
}
