% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{uniqueORFs}
\alias{uniqueORFs}
\title{Get the unique set of orfs}
\usage{
uniqueORFs(grl)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}
}
\value{
a GRangesList of unique orfs
}
\description{
Some orfs might be found several times, from different isoforms
If you want to have the unique orfs, not seperated by which
isoform it came from, use this function.
}
\details{
You will lose the transcript and name information, since
they no longer map to a transcript, but are now general.
}
\examples{
gr1 <- GRanges("1", IRanges(1,10), "+")
gr2 <- GRanges("1", IRanges(20, 30), "+")
# make a grl with duplicated ORFs (gr1 twice)
grl <- GRangesList(tx1_1 = gr1, tx2_1 = gr2, tx3_1 = gr1)
uniqueORFs <- uniqueORFs(grl)
length(grl)
length(uniqueORFs)
}
