% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_region.R
\name{unlistToExtremities}
\alias{unlistToExtremities}
\title{Get flanks as GRanges}
\usage{
unlistToExtremities(grl)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}
}
\value{
a GRangs object with meta column "group", which gives
}
\description{
For a GRangesList, get start and end site, return back as GRanges.
}
