% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{updateTxdbRanks}
\alias{updateTxdbRanks}
\title{Update exon ranks of exon data.frame inside txdb object}
\usage{
updateTxdbRanks(exons)
}
\arguments{
\item{exons}{a data.frame, call of as.list(txdb)$splicings}
}
\value{
a data.frame, modified call of as.list(txdb)
}
\description{
Update exon ranks of exon data.frame inside txdb object
}
\keyword{internal}
