% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{updateTxdbStartSites}
\alias{updateTxdbStartSites}
\title{Update start sites of leaders}
\usage{
updateTxdbStartSites(txList, fiveUTRs, removeUnused)
}
\arguments{
\item{txList}{a list, call of as.list(txdb)}

\item{fiveUTRs}{a GRangesList of 5' leaders}

\item{removeUnused}{logical (FALSE), remove leaders that did not have any
cage support. (standard is to set them to original annotation)}
}
\value{
a list, modified call of as.list(txdb)
}
\description{
Update start sites of leaders
}
