% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{validateExperiments}
\alias{validateExperiments}
\title{Validate ORFik \code{\link{experiment}}}
\usage{
validateExperiments(df, library.names = bamVarName(df), validate_libs = TRUE)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{library.names}{character vector, names of libraries, default:
name_decider(df, naming)}

\item{validate_libs}{logical, default TRUE. If FALSE, don't check that default
files exists (i.e. bam files), useful if you are using pshifted ofst etc
and don't have the bams anymore.}
}
\value{
invisible(NULL) (Stops if failed)
}
\description{
Check for valid existing, non-empty and all unique.
A good way to see if your experiment is valid.
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment}()},
\code{\link{ORFik.template.experiment.zf}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism,experiment-method}},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()}
}
\concept{ORFik_experiment}
\keyword{internal}
