% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{validExtension}
\alias{validExtension}
\title{Helper function to check valid combinations of extension and cageFiveUTRs}
\usage{
validExtension(extension, cageFiveUTRs)
}
\arguments{
\item{extension}{a numeric/integer to reassign 5' utrs.}

\item{cageFiveUTRs}{a GRangesList, if you used cage-data to extend 5' utrs,}
}
\description{
Helper function to check valid combinations of extension and cageFiveUTRs
}
