% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{validGRL}
\alias{validGRL}
\title{Helper Function to check valid GRangesList input}
\usage{
validGRL(class, type = "grl", checkNULL = FALSE)
}
\arguments{
\item{class}{as character vector the given class of
supposed GRangesList object}

\item{type}{a character vector, is it gtf, cds, 5', 3', for messages.}

\item{checkNULL}{should NULL classes be checked and return indeces of these?}
}
\value{
either NULL or indices (checkNULL == TRUE)
}
\description{
Helper Function to check valid GRangesList input
}
\seealso{
Other validity: 
\code{\link{checkRFP}()},
\code{\link{checkRNA}()},
\code{\link{is.ORF}()},
\code{\link{is.gr_or_grl}()},
\code{\link{is.grl}()},
\code{\link{validSeqlevels}()}
}
\concept{validity}
