% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{validSeqlevels}
\alias{validSeqlevels}
\title{Helper function to find overlaping seqlevels}
\usage{
validSeqlevels(grl, reads)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} or GRanges object}

\item{reads}{a GRanges, GAlignment or GAlignmentPairs object}
}
\value{
a character vector of valid seqlevels
}
\description{
Keep only seqnames in reads that are in grl
Useful to avoid seqname warnings in bioC
}
\seealso{
Other validity: 
\code{\link{checkRFP}()},
\code{\link{checkRNA}()},
\code{\link{is.ORF}()},
\code{\link{is.gr_or_grl}()},
\code{\link{is.grl}()},
\code{\link{is.range}()},
\code{\link{validGRL}()}
}
\concept{validity}
