% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{width,GRanges-method}
\alias{width,GRanges-method}
\title{Get width of GRanges}
\usage{
\S4method{width}{GRanges}(x)
}
\arguments{
\item{x}{a GRanges}
}
\value{
an integer (length 1)
}
\description{
Faster version than S4Vector generic caller
}
