% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{widthPerGroup}
\alias{widthPerGroup}
\title{Get list of widths per granges group}
\usage{
widthPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
an integer vector (named/unnamed) of widths
}
\description{
Get list of widths per granges group
}
\examples{
gr_plus <- GRanges(seqnames = c("chr1", "chr1"),
                   ranges = IRanges(c(7, 14), width = 3),
                   strand = c("+", "+"))
gr_minus <- GRanges(seqnames = c("chr2", "chr2"),
                    ranges = IRanges(c(4, 1), c(9, 3)),
                    strand = c("-", "-"))
grl <- GRangesList(tx1 = gr_plus, tx2 = gr_minus)
widthPerGroup(grl)
}
