% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplots_helpers.R
\name{window_resize}
\alias{window_resize}
\title{Creates window around GRanged object.}
\usage{
window_resize(GRanges_obj, window_size = 30)
}
\arguments{
\item{GRanges_obj}{GRanges object of your CDSs start or stop postions.}

\item{window_size}{Numeric. Default 30. What size of the window to consider.}
}
\value{
A GRanges object of resized by window_size sites.
}
\description{
It creates window of window_size around input ranges eg.
for GRanges starting at 100-100 and window_size of 3 will give
97-103
}
\examples{
window_resize(GRanges(Rle(c('1'), c(4)),
                      IRanges(c(100, 200, 200, 100), width=c(1, 1, 1, 1)),
                      Rle(strand(c('+', '+', '-', '-')))),
             window_size = 50)

}

