% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{windowPerGroup}
\alias{windowPerGroup}
\title{Get window region of tx around point of gr}
\usage{
windowPerGroup(gr, tx, downstream = 0L, upstream = 0L)
}
\arguments{
\item{gr}{a GRanges object (startSites and others, must be single point)}

\item{tx}{a GRangesList of transcripts or (container region)}

\item{downstream}{an integer, relative region to get downstream from}

\item{upstream}{an integer vector, relative region to get upstream from.}
}
\value{
a GRanges/GRangesList object if exon/introns
}
\description{
If downstreamFrom is 20, it means the window will start -20 downstream of
gr start site. It will keep exon structure of tx
}
\seealso{
Other ExtendGenomicRanges: \code{\link{asTX}},
  \code{\link{coveragePerTiling}},
  \code{\link{overlapsToCoverage}},
  \code{\link{pmapFromTranscriptF}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{txSeqsFromFa}}
}
\concept{ExtendGenomicRanges}
