% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{windowPerTranscript}
\alias{windowPerTranscript}
\title{Get coverage window per transcript}
\usage{
windowPerTranscript(
  txdb,
  reads,
  splitIn3 = TRUE,
  windowSize = 100,
  fraction = "1"
)
}
\arguments{
\item{txdb}{a TxDb object or a path to gtf/gff/db file.}

\item{reads}{GRanges or GAlignment of reads}

\item{splitIn3}{a logical(TRUE), split window in 3 (leader, cds, trailer)}

\item{windowSize}{an integer (100), size of windows}

\item{fraction}{a character (1), info on reads (which read length,
or which type (RNA seq))}
}
\value{
a data.table with columns position, score
}
\description{
NOTE: All ranges with smaller width than windowSize, will of course be
removed. What is the 100 position on a 1 width size ?
}
