% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{xAxisScaler}
\alias{xAxisScaler}
\title{Scale x axis correctly}
\usage{
xAxisScaler(covPos)
}
\arguments{
\item{covPos}{a numeric vector of positions in coverage}
}
\value{
a numeric vector from the seq() function, aligned to 0.
}
\description{
Works for all coverage plots, that need 0 position aligning
}
\details{
It basicly bins the x axis on floor(length of x axis / 20) or
1 if x < 20
}
