% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{addNewTSSOnLeaders}
\alias{addNewTSSOnLeaders}
\title{add cage max peaks as new transcript start sites for each 5' leader}
\usage{
addNewTSSOnLeaders(fiveUTRs, maxPeakPosition)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}

\item{maxPeakPosition}{The max peak for each 5' leader found by cage}
}
\description{
add cage max peaks as new transcript start sites for each 5' leader
}
