% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{alignmentFeatureStatistics}
\alias{alignmentFeatureStatistics}
\title{Create alignment feature statistcs}
\usage{
alignmentFeatureStatistics(df, type = "ofst", BPPARAM = bpparam())
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "ofst", "bedo", "bedoc" or "pshifted".
These are made with ORFik:::convertLibs() or shiftFootprintsByExperiment().
Can also be custom user made folders inside the experiments bam folder.
It acts in a recursive manner with priority: If you state "pshifted",
but it does not exist, it checks "ofst". If no .ofst files, it uses
"default", which always exists.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}.
You can also add a time remaining bar, for a more detailed pipeline.}
}
\value{
a data.table of the statistcs
}
\description{
Among others how much reads are in mRNA, introns, intergenic,
and check of reads from rRNA and other ncRNAs.
The better the annotation / gtf used, the more results you get.
}
\keyword{internal}
