% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{allFeatures}
\alias{allFeatures}
\title{Get all possible feature in ORFik}
\usage{
allFeatures(grl, orfFeatures = T, RFP, RNA = NULL, Gtf = NULL,
  tx = NULL, fiveUTRs = NULL, cds = NULL, threeUTRs = NULL,
  faFile = NULL, riboStart = 26, riboStop = 34, extension = NULL)
}
\arguments{
\item{grl}{a GRangesList object with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object
@param RNA rna seq reads as GAlignment, GRanges
or GRangesList object}

\item{Gtf}{a TxDb object of a gtf file,}

\item{tx}{a GrangesList of transcripts,
normally called from: exonsBy(Gtf, by = "tx", use.names = T)
only add this if you are not including Gtf file}

\item{fiveUTRs}{fiveUTRs as GRangesList, must be original unchanged fiveUTRs}

\item{cds}{a GRangesList of coding sequences}

\item{threeUTRs}{a GrangesList of transcript 3' utrs,
normally called from: threeUTRsByTranscript(Gtf, use.names = T)}

\item{faFile}{a FaFile from the fasta file, see ?FaFile}

\item{riboStart}{usually 26, the start of the floss interval, see ?floss}

\item{riboStop}{usually 34, the end of the floss interval}

\item{extension}{needs to be set! set to 0 if you did not use cage
if you used cage to change tss' when finding the orfs, standard cage
extension is 1000}
}
\value{
a data.table with scores, each column is one score type, name of
 columns are the names of the scores, i.g floss or fpkmRFP.
}
\description{
If you want to get all the features easily, run this function.
There are a lot of arguments, so make sure to input correct,
that is: use grl = "yourgrl", RFP = "yourRFP" instead of
"yourgrl", "yourRFP" etc.
The easiest method to run is to include as this:
allFeatures(grl = grl, orfFeatures =  T, RFP = RFP, RNA = RNA, Gtf = Gtf,
 faFile = faFile, extension = 0)
The other arguments will then be found by the Gtf file,
and extension = 0 means
you did not use cage to extend 5' utrs
Each feature have a link to an article describing feature,
try ?floss
}
