% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Features.R
\name{allFeaturesHelper}
\alias{allFeaturesHelper}
\title{Calculate the features in computeFeatures}
\usage{
allFeaturesHelper(
  grl,
  RFP,
  RNA,
  tx,
  fiveUTRs,
  cds,
  threeUTRs,
  faFile,
  riboStart,
  riboStop,
  orfFeatures,
  includeNonVarying,
  grl.is.sorted
)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually ORFs, but can also be either leaders, cds', 3' utrs, etc.}

\item{RFP}{RiboSeq reads as GAlignment, GRanges or GRangesList object}

\item{RNA}{RnaSeq reads as GAlignment, GRanges or GRangesList object}

\item{tx}{a GrangesList of transcripts,
normally called from: exonsBy(Gtf, by = "tx", use.names = T)
only add this if you are not including Gtf file
You do not need to reassign these to the cage peaks, it will do it for you.}

\item{fiveUTRs}{fiveUTRs as GRangesList, if you used cage-data to
extend 5' utrs, remember to input CAGE assigned version and not original!}

\item{cds}{a GRangesList of coding sequences}

\item{threeUTRs}{a GrangesList of transcript 3' utrs,
normally called from: threeUTRsByTranscript(Gtf, use.names = T)}

\item{faFile}{a FaFile or BSgenome from the fasta file, see ?FaFile}

\item{riboStart}{usually 26, the start of the floss interval, see ?floss}

\item{riboStop}{usually 34, the end of the floss interval}

\item{orfFeatures}{a logical, is the grl a list of orfs?}

\item{includeNonVarying}{a logical, if TRUE, include all features not
dependent on RiboSeq data and RNASeq data, that is: Kozak,
fractionLengths, distORFCDS, isInFrame, isOverlapping and rankInTx}

\item{grl.is.sorted}{logical (F), a speed up if you know argument grl
is sorted, set this to TRUE.}
}
\value{
a data.table with features
}
\description{
Not used directly, calculates all features.
}
