% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{assignFirstExons}
\alias{assignFirstExons}
\title{After all transcript start sites have been updated from cage,
 put grangeslist back together}
\usage{
assignFirstExons(firstExons, fiveUTRs)
}
\arguments{
\item{firstExons}{The first exon of every transcript from 5' leaders}

\item{fiveUTRs}{The 5' leader sequences as GRangesList}
}
\description{
After all transcript start sites have been updated from cage,
 put grangeslist back together
}
