% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{assignFirstExonsStartSite}
\alias{assignFirstExonsStartSite}
\title{Reassign the start positions of the first exons per group in grl}
\usage{
assignFirstExonsStartSite(grl, newStarts)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object}

\item{newStarts}{an integer vector of same length as grl, with new start values}
}
\description{
make sure your grl is sorted, since start of "-" strand
 objects should be the
 max end in group, use ORFik:::sortPerGroup(grl) to get sorted grl.
}
