% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{assignLastExonsStopSite}
\alias{assignLastExonsStopSite}
\title{Reassign the stop positions of the last exons per group in grl
make sure your grl is sorted, since stop of "-" strand objects should be the
min start in group, use ORFik:::sortPerGroup(grl) to get sorted grl.}
\usage{
assignLastExonsStopSite(grl, newStops)
}
\arguments{
\item{grl}{a GRangesList object}

\item{newStarts}{an integer vector of same length as grl, with new start values}
}
\description{
Reassign the stop positions of the last exons per group in grl
make sure your grl is sorted, since stop of "-" strand objects should be the
min start in group, use ORFik:::sortPerGroup(grl) to get sorted grl.
}
