% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{asTX}
\alias{asTX}
\title{Map genomic to transcript coordinates by reference}
\usage{
asTX(grl, reference)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} of ranges within
the reference, grl must have column called names that gives
grouping for result}

\item{reference}{a GrangesList of ranges that include and are bigger or
equal to grl ig. cds is grl and gene can be reference}
}
\value{
a GRangesList in transcript coordinates
}
\description{
Map genomic to transcript coordinates by reference
}
\seealso{
Other ExtendGenomicRanges: \code{\link{coveragePerTiling}},
  \code{\link{overlapsToCoverage}},
  \code{\link{pmapFromTranscriptF}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{txSeqsFromFa}}, \code{\link{windowPerGroup}}
}
\concept{ExtendGenomicRanges}
