% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{asTX}
\alias{asTX}
\title{map genomic to transcript coordinates by reference}
\usage{
asTX(grl, reference)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} of ranges within
the reference, grl must have column called names that gives
grouping for result}

\item{reference}{a GrangesList of ranges
that include and are bigger or equal to grl
ig. cds is grl and gene can be reference
@export}
}
\description{
map genomic to transcript coordinates by reference
}
