% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{asTX}
\alias{asTX}
\title{Map genomic to transcript coordinates by reference}
\usage{
asTX(grl, reference)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} of ranges within
the reference, grl must have column called names that gives
grouping for result}

\item{reference}{a GrangesList of ranges that include and are bigger or
equal to grl ig. cds is grl and gene can be reference}
}
\value{
a GRangesList in transcript coordinates
}
\description{
Map range coordinates between features in the genome and
transcriptome (reference) space.
}
\details{
Similar to GenomicFeatures' pmapToTranscripts, but in this version the
grl ranges are compared to reference ranges with same name, not
by index. And it has a security fix.
}
\seealso{
Other ExtendGenomicRanges: \code{\link{coveragePerTiling}},
  \code{\link{overlapsToCoverage}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{txSeqsFromFa}}, \code{\link{windowPerGroup}}
}
\concept{ExtendGenomicRanges}
