% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{bedToGR}
\alias{bedToGR}
\title{Converts different type of files to Granges
Only Accepts bed files for now, standard format from Fantom5}
\usage{
bedToGR(x, bed6 = TRUE)
}
\arguments{
\item{x}{An imported bed-file, to convert to GRanges}

\item{bed6}{If bed6, no meta column is added}
}
\description{
Converts different type of files to Granges
Only Accepts bed files for now, standard format from Fantom5
}
