% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bedToGR}
\alias{bedToGR}
\title{Converts bed style data.frame to Granges}
\usage{
bedToGR(x, skip.name = TRUE)
}
\arguments{
\item{x}{A \code{\link{data.frame}} from imported bed-file,
to convert to GRanges}

\item{skip.name}{default (TRUE), skip name column (column 4)}
}
\value{
a \code{\link{GRanges}} object from bed
}
\description{
For info on columns, see:
https://www.ensembl.org/info/website/upload/bed.html
}
\seealso{
Other utils: 
\code{\link{convertToOneBasedRanges}()},
\code{\link{export.bed12}()},
\code{\link{export.bigWig}()},
\code{\link{export.wiggle}()},
\code{\link{fimport}()},
\code{\link{findFa}()},
\code{\link{fread.bed}()},
\code{\link{optimizeReads}()},
\code{\link{readBam}()},
\code{\link{readWig}()}
}
\concept{utils}
\keyword{internal}
