% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_metadata.R
\name{browseSRA}
\alias{browseSRA}
\title{Open SRA in browser for specific bioproject}
\usage{
browseSRA(x)
}
\arguments{
\item{x}{character, bioproject ID.}
}
\description{
Open SRA in browser for specific bioproject
}
\examples{
browseSRA("PRJNA336542")

}
\seealso{
Other sra: 
\code{\link{download.SRA.metadata}()},
\code{\link{download.SRA}()},
\code{\link{download.ebi}()},
\code{\link{get_bioproject_candidates}()},
\code{\link{install.sratoolkit}()},
\code{\link{rename.SRA.files}()}
}
\concept{sra}
