% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{cageFromFile}
\alias{cageFromFile}
\title{Get Cage-Data from a file-path}
\usage{
cageFromFile(filePath)
}
\arguments{
\item{filePath}{The location of the cage file}
}
\value{
a GRanges object
}
\description{
Get Cage-Data from a file-path
}
