% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{cellLineNames}
\alias{cellLineNames}
\title{Get cell-line name variants}
\usage{
cellLineNames(convertToTissue = FALSE)
}
\arguments{
\item{convertToTissue}{logical, FALSE. If TRUE, return tissue type. NONE is
returned for general non-differentiated cell lines like 3T3.}
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: THP1 is main naming, but a variant is THP-1
THP-1 will then be renamed to THP1 (variables in R,
can not have - in them)
}
\seealso{
Other experiment_naming: 
\code{\link{batchNames}()},
\code{\link{conditionNames}()},
\code{\link{fractionNames}()},
\code{\link{inhibitorNames}()},
\code{\link{libNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{stageNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
\keyword{internal}
