% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{changePointAnalysis}
\alias{changePointAnalysis}
\title{Get the offset for specific RiboSeq read width}
\usage{
changePointAnalysis(x, feature = "start")
}
\arguments{
\item{x}{a vector with count per position to analyse, assumes the zero
is in the middle + 1 (position 0)}

\item{feature}{(character) either "start" or "stop"}
}
\value{
a single numeric offset
}
\description{
12 would mean p-site is 12 bases upstream
}
\seealso{
Other pshifting: 
\code{\link{detectRibosomeShifts}()},
\code{\link{shiftFootprintsByExperiment}()},
\code{\link{shiftFootprints}()}
}
\concept{pshifting}
