% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{checkRNA}
\alias{checkRNA}
\title{Helper Function to check valid RNA input}
\usage{
checkRNA(class)
}
\arguments{
\item{class, }{the given class of RNA object}
}
\value{
NULL, stop if unvalid object
}
\description{
Helper Function to check valid RNA input
}
\seealso{
Other validity: 
\code{\link{checkRFP}()},
\code{\link{is.ORF}()},
\code{\link{is.gr_or_grl}()},
\code{\link{is.grl}()},
\code{\link{validGRL}()},
\code{\link{validSeqlevels}()}
}
\concept{validity}
