% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{codonSumsPerGroup}
\alias{codonSumsPerGroup}
\title{Get read hits per codon}
\usage{
codonSumsPerGroup(grl, reads)
}
\arguments{
\item{grl}{a GRangesList}

\item{reads}{a GRanges or GAlignment}
}
\value{
a data.table with codon sums
}
\description{
Helper for entropy function, normally not used directly
Seperate each group into tuples (abstract codons)
Gives sum for each tuple within each group
}
\details{
Example: counts c(1,0,0,1), with reg_len = 2, gives
c(1,0) and c(0,1), these are summed and returned as data.table
10 bases, will give 3 codons, 1 base codons does not exist.
}
