% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{codonSumsPerGroup}
\alias{codonSumsPerGroup}
\title{Get hits per codon}
\usage{
codonSumsPerGroup(countList, reg_len, runLengths)
}
\arguments{
\item{countList}{a Rle of count repetitions (000,1,00,1 etc)}

\item{reg_len}{integer vector, size of runs}

\item{runLengths}{integer vector, duplications per run}
}
\value{
a list of codon sums
}
\description{
Helper for entropy function, normally not used directly
Seperate each group into tuplets (codons)
Gives sum for each tuplet within each group
Example: c(1,0,0,1), with reg_len = 2, gives
c(1,0) and c(0,1), these are summed and returned as list
}
