% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseDuplicatedReads,GAlignments-method}
\alias{collapseDuplicatedReads,GAlignments-method}
\title{Collapse duplicated reads}
\usage{
\S4method{collapseDuplicatedReads}{GAlignments}(x, addScoreColumn = TRUE)
}
\arguments{
\item{x}{a GAlignments object}

\item{addScoreColumn}{= TRUE, if FALSE,
only collapse and not add score column.}
}
\value{
a GAlignments object
}
\description{
For every GAlignments read, with the same:
seqname, start, cigar and strand, collapse and give a new
meta column called "score", which contains the number of duplicates
of that read.
}
