% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseDuplicatedReads}
\alias{collapseDuplicatedReads}
\title{Collapse duplicated reads}
\usage{
collapseDuplicatedReads(x, ...)
}
\arguments{
\item{x}{a GRanges, GAlignments or GAlignmentPairs object}

\item{...}{alternative arguments. addScoreColumn = TRUE, if FALSE,
only collapse and not add score column.}
}
\value{
a GRanges, GAlignments or GAlignmentPairs object, same as input
}
\description{
For every GAlignments read, with the same:
seqname, start, cigar and strand, collapse and give a new
meta column called "score", which contains the number of duplicates
of that read. If score column already exists, will return input object!
}
