% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combn.pairs}
\alias{combn.pairs}
\title{Create all unique combinations pairs possible}
\usage{
combn.pairs(x)
}
\arguments{
\item{x}{a character vector, will unique elements for you.}
}
\value{
a list of character vector pairs
}
\description{
Given a character vector, get all unique combinations of 2.
}
\keyword{internal}
