% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{computeFeatures}
\alias{computeFeatures}
\title{Get all possible features in ORFik}
\usage{
computeFeatures(grl, RFP, RNA = NULL, Gtf = NULL, faFile = NULL,
  riboStart = 26, riboStop = 34, orfFeatures = TRUE,
  includeNonVarying = TRUE, grl.is.sorted = FALSE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually ORFs, but can also be either leaders, cds', 3' utrs, etc.}

\item{RFP}{RiboSeq reads as GAlignment, GRanges or GRangesList object}

\item{RNA}{RnaSeq reads as GAlignment, GRanges or GRangesList object}

\item{Gtf}{a TxDb object of a gtf file,}

\item{faFile}{a FaFile or BSgenome from the fasta file, see ?FaFile}

\item{riboStart}{usually 26, the start of the floss interval, see ?floss}

\item{riboStop}{usually 34, the end of the floss interval}

\item{orfFeatures}{a logical, is the grl a list of orfs?}

\item{includeNonVarying}{a logical, if TRUE, include all features not
dependent on RiboSeq data and RNASeq data, that is: Kozak,
fractionLengths, distORFCDS, isInFrame, isOverlapping and rankInTx}

\item{grl.is.sorted}{logical (F), a speed up if you know argument grl
is sorted, set this to TRUE.}
}
\value{
a data.table with scores, each column is one score type, name of
columns are the names of the scores, i.g [floss()]
or [fpkm()]
}
\description{
If you want to get all the features easily, you can use this function.
Each feature have a link to an article describing its creation and idea
behind it. Look at the functions in the feature family to see all of them.
}
\details{
If you used CageSeq to reannotate your leaders your txDB object, must
contain the reassigned leaders. In the future release reasignment will
create txdb objects for you, but currently this is not supported,
therefore be carefull.
}
\examples{
# Usually the ORFs are found in orfik, which makes names for you etc.
# Here we make an example from scratch
gtf <- system.file("extdata", "annotations.gtf",
package = "ORFik") ## location of the gtf file
suppressWarnings(txdb <-
                  GenomicFeatures::makeTxDbFromGFF(gtf, format = "gtf"))
# use cds' as ORFs for this example
ORFs <- GenomicFeatures::cdsBy(txdb, by = "tx", use.names = TRUE)
ORFs <- makeORFNames(ORFs) # need ORF names
# make Ribo-seq data,
RFP <- unlistGrl(firstExonPerGroup(ORFs))
suppressWarnings(computeFeatures(ORFs, RFP, Gtf = txdb))
# For more details see vignettes.

}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{distToTSS}},
  \code{\link{entropy}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
\concept{features}
