% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{computeFeatures}
\alias{computeFeatures}
\title{Get all possible features in ORFik}
\usage{
computeFeatures(grl, RFP, RNA = NULL, Gtf = NULL, faFile = NULL,
  riboStart = 26, riboStop = 34, orfFeatures = TRUE,
  includeNonVarying = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
with usually ORFs, but can also be either leaders, cds', 3' utrs, etc.}

\item{RFP}{RiboSeq reads as GAlignment, GRanges or GRangesList object}

\item{RNA}{RnaSeq reads as GAlignment, GRanges or GRangesList object}

\item{Gtf}{a TxDb object of a gtf file,}

\item{faFile}{a FaFile or BSgenome from the fasta file, see ?FaFile}

\item{riboStart}{usually 26, the start of the floss interval, see ?floss}

\item{riboStop}{usually 34, the end of the floss interval}

\item{orfFeatures}{a logical, is the grl a list of orfs?}

\item{includeNonVarying}{a logical, if TRUE, include all features not
dependent on RiboSeq data and RNASeq data, that is: Kozak,
fractionLengths, distORFCDS, isInFrame, isOverlapping and rankInTx}
}
\value{
a data.table with scores, each column is one score type, name of
columns are the names of the scores, i.g \code{\link{floss}}
or \code{\link{fpkm}}
}
\description{
If you want to get all the features easily, you can use this function.
Each feature have a link to an article describing its creation and idea
behind it. Look at the functions in the feature family to see all of them.
}
\details{
If you used CageSeq to reannotate your leaders your txDB object, must contain
the reassigned leaders. In the future release reasignment will create txdb
objects for you, but currently this is not supported, therefore be carefull.
}
\examples{
# FIX or remove the @export

}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
