% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{computeFeaturesCage}
\alias{computeFeaturesCage}
\title{Get all possible features in ORFik}
\usage{
computeFeaturesCage(grl, RFP, RNA = NULL, Gtf = NULL, tx = NULL,
  fiveUTRs = NULL, cds = NULL, threeUTRs = NULL, faFile = NULL,
  riboStart = 26, riboStop = 34, extension = NULL, orfFeatures = TRUE,
  cageFiveUTRs = NULL, includeNonVarying = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
with usually ORFs, but can also be
either leaders, cds', 3' utrs or  ORFs are a special case,
see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}

\item{Gtf}{a TxDb object of a gtf file,}

\item{tx}{a GrangesList of transcripts,
normally called from: exonsBy(Gtf, by = "tx", use.names = T)
only add this if you are not including Gtf file
You do not need to reassign these to the cage peaks, it will do it for you.}

\item{fiveUTRs}{fiveUTRs as GRangesList, must be original unchanged fiveUTRs}

\item{cds}{a GRangesList of coding sequences}

\item{threeUTRs}{a GrangesList of transcript 3' utrs,
normally called from: threeUTRsByTranscript(Gtf, use.names = T)}

\item{faFile}{a FaFile or BSgenome from the fasta file, see ?FaFile}

\item{riboStart}{usually 26, the start of the floss interval, see ?floss}

\item{riboStop}{usually 34, the end of the floss interval}

\item{extension}{a numeric/integer needs to be set! set to 0 if you did not
use cage, if you used cage to change tss' when finding the orfs,
standard cage extension is 1000}

\item{orfFeatures}{a logical,  is the grl a list of orfs? Must be assigned.}

\item{cageFiveUTRs}{a GRangesList, if you used cage-data to extend 5' utrs,
include this, also extension must match with the extension used for these.}

\item{includeNonVarying}{a logical T, if TRUE, include all features not dependent on
Ribo-seq data and RNA-seq data, that is: Kozak, fractionLengths, distORFCDS,
isInFrame, isOverlapping and rankInTx}
}
\value{
a data.table with scores, each column is one score type, name of
 columns are the names of the scores, i.g \code{\link{floss}}
 or \code{\link{fpkm}}
}
\description{
Normally dont use this function, but instead use:
\code{\link{computeFeatures}}
}
\details{
A specialized version if you used Cage data, and don't have
a new txdb with reassigned leaders, transcripts and gene starts.
If you do have a txdb with cage reassignments, use computeFeatures
instead.
Each feature have a link to an article describing feature,
try ?floss
}
\examples{
 # a small example without cage-seq data:
 # we will find ORFs in the 5' utrs
 # and then calculate features on them
 \dontrun{
 if (requireNamespace("BSgenome.Hsapiens.UCSC.hg19")) {
  library(GenomicFeatures)
  # Get the gtf txdb file
  txdbFile <- system.file("extdata", "hg19_knownGene_sample.sqlite",
  package = "GenomicFeatures")
  txdb <- loadDb(txdbFile)

  # Extract sequences of fiveUTRs.
  fiveUTRs <- fiveUTRsByTranscript(txdb, use.names = TRUE)[1:10]
  faFile <- BSgenome.Hsapiens.UCSC.hg19::Hsapiens
  # need to suppress warning because of bug in GenomicFeatures, will
  # be fixed soon.
  tx_seqs <- suppressWarnings(extractTranscriptSeqs(faFile, fiveUTRs))

  # Find all ORFs on those transcripts and get their genomic coordinates
  fiveUTR_ORFs <- findMapORFs(fiveUTRs, tx_seqs)
  unlistedORFs <- unlist(fiveUTR_ORFs, use.names = TRUE)
  # group GRanges by ORFs instead of Transcripts
  fiveUTR_ORFs <- groupGRangesBy(unlistedORFs, unlistedORFs$names)

  # make some toy ribo seq and rna seq data
  starts <- unlist(ORFik:::firstExonPerGroup(fiveUTR_ORFs), use.names = FALSE)
  RFP <- promoters(starts, upstream = 0, downstream = 1)
  score(RFP) <- rep(29, length(RFP)) # the original read widths

  # set RNA seq to duplicate transcripts
  RNA <- unlist(exonsBy(txdb, by = "tx", use.names = TRUE), use.names = TRUE)

  cageNotUsed <- 0 # used to inform that no cage was used

  computeFeaturesCage(grl = fiveUTR_ORFs, orfFeatures =  TRUE, RFP = RFP, RNA = RNA,
   Gtf = txdb, faFile = faFile, extension = cageNotUsed)

}
# See vignettes for more examples
}

}
\seealso{
Other features: \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
