% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_directory_config.R
\name{config}
\alias{config}
\title{Read directory config for ORFik experiments}
\usage{
config(file = "~/Bio_data/ORFik_config.csv")
}
\arguments{
\item{file}{file of config for ORFik, default: "~/Bio_data/ORFik_config.csv"}
}
\value{
a named character vector of length 3
}
\description{
Defines a folder for:
1. fastq files (raw_data)\cr
2. bam files (processed data)\cr
3. references (organism annotation and STAR index)
}
\details{
Update or use another config using \code{config.save()} function.
}
