% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_directory_config.R
\name{config.save}
\alias{config.save}
\title{Save/update directory config for ORFik experiments}
\usage{
config.save(
  file = "~/Bio_data/ORFik_config.csv",
  fastq.dir,
  bam.dir,
  reference.dir,
  exp.dir = "~/Bio_data/ORFik_experiments/"
)
}
\arguments{
\item{file}{file of config for ORFik, default: "~/Bio_data/ORFik_config.csv"}

\item{fastq.dir}{directory where ORFik puts fastq file directories,
default: config()["fastq"]}

\item{bam.dir}{directory where ORFik puts bam file directories,
default: config()["bam"]}

\item{reference.dir}{directory where ORFik puts reference file directories,
default: config()["ref"]}

\item{exp.dir}{directory where ORFik puts experiment csv files,
default: "~/Bio_data/ORFik_experiments/", which is retrieved with
\code{config()["exp"]}}
}
\value{
invisible(NULL), file saved to disc
}
\description{
Defines a folder for fastq files (raw_data), bam files (processed data) and
references (organism annotation and STAR index)
}
\examples{
## Save at another config location
config_location <- "/media/Bio_data/ORFik_config.csv"
#config.save(config_location, "/media/Bio_data/raw_data/",
# "/media/Bio_data/processed_data", /media/Bio_data/references/)
}
