% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_directory_config.R
\name{config.save}
\alias{config.save}
\title{Save/update directory config for ORFik experiments}
\usage{
config.save(
  file = "~/Bio_data/ORFik_config.csv",
  fastq.dirs,
  bam.dirs,
  reference.dirs
)
}
\arguments{
\item{file}{file of config for ORFik, default: "~/Bio_data/ORFik_config.csv"}

\item{bam.dirs}{directory where ORFik puts bam file directories,
default: config()["bam"]}

\item{reference.dirs}{directory where ORFik puts reference file directories,
default: config()["ref"]}

\item{fastq.dir}{directory where ORFik puts fastq file directories,
default: config()["fastq"]}
}
\value{
invisible(NULL), file saved to disc
}
\description{
Defines a folder for fastq files (raw_data), bam files (processed data) and
references (organism annotation and STAR index)
}
