% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_conversion.R
\name{convert_to_bigWig}
\alias{convert_to_bigWig}
\title{Convert to BigWig}
\usage{
convert_to_bigWig(
  df,
  in_files = filepath(df, "pshifted"),
  out_dir = file.path(libFolder(df), "bigwig"),
  split.by.strand = TRUE,
  split.by.readlength = FALSE,
  seq_info = seqinfo(df),
  weight = "score",
  is_pre_collapsed = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{in_files}{paths to input files, default pshifted files:
\code{filepath(df, "pshifted")} in ofst format}

\item{split.by.strand}{logical, default TRUE, split into forward and reverse
strand RleList inside covRle object.}

\item{weight}{integer, numeric or single length character. Default "score".
Use score column in loaded in_files.}

\item{verbose}{logical, default TRUE, message about library output status.}
}
\value{
invisible(NULL), files saved to disc
}
\description{
Convert to BigWig
}
