% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_format_conversion.R
\name{convert_to_covRleList}
\alias{convert_to_covRleList}
\title{Convert libraries to covRleList objects}
\usage{
convert_to_covRleList(
  df,
  in_files = filepath(df, "pshifted"),
  out_dir = file.path(libFolder(df), "cov_RLE_List"),
  out_dir_merged = file.path(libFolder(df), "cov_RLE"),
  split.by.strand = TRUE,
  seq_info = seqinfo(df),
  weight = "score",
  verbose = TRUE
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{in_files}{paths to input files, default pshifted files:
\code{filepath(df, "pshifted")} in ofst format}

\item{out_dir_merged}{character vector of paths, default:
\code{file.path(libFolder(df), "cov_RLE")}.
Paths to merged output files, Set to NULL to skip making merged covRle.}

\item{split.by.strand}{logical, default TRUE, split into forward and reverse
strand RleList inside covRle object.}

\item{weight}{integer, numeric or single length character. Default "score".
Use score column in loaded in_files.}

\item{verbose}{logical, default TRUE, message about library output status.}
}
\value{
invisible(NULL), files saved to disc
}
\description{
Useful to store reads separated by readlength, for much faster
coverage calculation.
Saved by default in folder "cov_RLE_List" relative to default
libraries of experiment
}
