% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{convertLibs}
\alias{convertLibs}
\title{Converted format of NGS libraries}
\usage{
convertLibs(
  df,
  out.dir = libFolder(df),
  addScoreColumn = TRUE,
  addSizeColumn = TRUE,
  must.overlap = NULL,
  method = "None",
  type = "ofst",
  reassign.when.saving = FALSE,
  envir = envExp(df),
  BPPARAM = bpparam()
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default: libFolder(df),
if it is NULL, it will just reassign R objects to simplified libraries.
Will then create a final folder specfied as: paste0(out.dir, "/", type, "/").
Here the files will be saved in format given by the type argument.}

\item{addScoreColumn}{logical, default TRUE, if FALSE will not add
replicate numbers as score column, see ORFik::convertToOneBasedRanges.}

\item{addSizeColumn}{logical, default TRUE, if FALSE will not add
size (width) as size column, see ORFik::convertToOneBasedRanges.
Does not apply for (GAlignment version of.ofst) or .bedoc. Since they
contain the original cigar.}

\item{must.overlap}{default (NULL), else a GRanges / GRangesList object, so
only reads that overlap (must.overlap) are kept. This is useful when you
only need the reads over transcript annotation or subset etc.}

\item{method}{character, default "None", the method to reduce ranges,
for more info see \code{\link{convertToOneBasedRanges}}}

\item{type}{a character of format, default "ofst".
Alternatives: "ofst", "bigWig", "wig","bedo" or "bedoc". Which format you want.
Will make a folder within out.dir with this name containing the files.}

\item{reassign.when.saving}{logical, default FALSE. If TRUE, will reassign
library to converted form after saving. Ignored when out.dir = NULL.}

\item{envir}{environment to save to, default
\code{envExp(df)}, which defaults to .GlobalEnv, but can be set with
\code{envExp(df) <- new.env()} etc.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}.
You can also add a time remaining bar, for a more detailed pipeline.}
}
\value{
NULL (saves files to disc or R .GlobalEnv)
}
\description{
Export as either .ofst, .wig, .bigWig,.bedo (legacy format) or .bedoc (legacy format) files:\cr
Export files as .ofst for fastest load speed into R.\cr
Export files as .wig / bigWig for use in IGV or other genome browsers.\cr
The input files are checked if they exist from: \code{envExp(df)}.\cr
}
\details{
See \code{\link{export.ofst}}, \code{\link{export.wiggle}},
\code{\link{export.bedo}} and \code{\link{export.bedoc}}
for information on file formats.\cr
If libraries of the experiment are
already loaded into environment (default: .globalEnv) is will export
using those files as templates. If they are not in environment the
.ofst files from the bam files are loaded (unless you are converting
to .ofst then the .bam files are loaded).
}
\examples{
df <- ORFik.template.experiment()
#convertLibs(df)
# Keep only 5' ends of reads
#convertLibs(df, method = "5prime")
}
