% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plots.R
\name{correlation.plots}
\alias{correlation.plots}
\title{Correlation plots between all samples}
\usage{
correlation.plots(
  df,
  output.dir,
  region = "mrna",
  type = "fpkm",
  height = 400,
  width = 400,
  size = 0.15
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{output.dir}{directory to save to, 2 files named: cor_plot.pdf and
cor_plot_log2.pdf}

\item{region}{a character (default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{type}{which value to use, "fpkm", alternative "counts".}

\item{height}{numeric, default 400 (in mm)}

\item{width}{numeric, default 400 (in mm)}

\item{size}{numeric, size of dots, default 0.15.}
}
\value{
invisible(NULL)
}
\description{
Get 2 correlation plots of raw counts and log2(count + 1) over
selected region in: c("mrna", "leaders", "cds", "trailers")
}
