% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{coverageByTranscriptW}
\alias{coverageByTranscriptW}
\title{Remake of coverageByTranscript}
\usage{
coverageByTranscriptW(x, transcripts, ignore.strand = FALSE, weight = 1L)
}
\arguments{
\item{x}{reads (\code{\link{GRanges}}, \code{\link{GAlignments}})}

\item{transcripts}{\code{\link{GRangesList}}}

\item{ignore.strand}{a logical (default: FALSE)}

\item{weight}{a vector (default: 1L), if single number applies for all,
else it must be the string name of a defined meta column in "x",
that gives number of times a read was found.
GRanges("chr1", 1, "+", score = 5), would mean score column tells
that this alignment was found 5 times.}
}
\value{
Integer Rle of coverage, 1 per transcript
}
\description{
Allows weights, see \code{\link{coverageByTranscript}}
}
