% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{coverageHeatMap}
\alias{coverageHeatMap}
\title{Create a heatmap of coverage}
\usage{
coverageHeatMap(coverage, output = NULL, scoring = "zscore")
}
\arguments{
\item{coverage}{a data.table, e.g. output of scaledWindowCoverage}

\item{output}{character string (NULL), if set, saves the plot as pdf or png
to path given. If no format is given, is save as pdf.}

\item{scoring}{character vector (zscore), either of zScore,
transcriptNormalized, sum, mean, median, NULL. Set NULL if already scored.}
}
\value{
a ggplot object of the coverage plot, NULL if output is set,
then the plot will only be saved to location.
}
\description{
Coverage rows in heat map is fraction
Coverage column in heat map is score, default zscore of counts
}
\details{
See vignette for example

Remember if you want to change anything like colors, just return the
ggplot object, and reassign like: obj + scale_color_brewer() etc.
}
\examples{
# An ORF
grl <- GRangesList(tx1 = GRanges("1", IRanges(1, 6), "+"))
# Ribo-seq reads
range <- IRanges(c(rep(1, 3), 2, 3, rep(4, 2), 5, 6), width = 1 )
reads <- GRanges("1", range, "+")
reads$size <- c(rep(28, 5), rep(29, 4)) # read size
coverage <- ORFik:::windowPerReadLength(grl, reads = reads, upstream = 0,
                                        downstream = 5)

ORFik:::coverageHeatMap(coverage)

# See vignette for more examples

}
\seealso{
Other coveragePlot: \code{\link{pSitePlot}},
  \code{\link{savePlot}}, \code{\link{windowCoveragePlot}}
}
\concept{coveragePlot}
