% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{coveragePerTiling}
\alias{coveragePerTiling}
\title{Get coverage per group}
\usage{
coveragePerTiling(grl, reads, is.sorted = FALSE, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}
of 5' utrs or transcripts.}

\item{reads}{a GAlignment or GRanges object of RiboSeq, RnaSeq etc.}

\item{is.sorted}{logical (F), is grl sorted.}

\item{keep.names}{logical (T), keep names or not.}
}
\value{
a RleList, one integer-Rle per group with # of hits per position.
}
\description{
It tiles each GRangesList group, and finds hits per position
}
\details{
This is a safer speedup of coverageByTranscript from GenomicFeatures
}
\examples{
ORF <- GRanges(seqnames = "1",
               ranges = IRanges(start = c(1, 10, 20),
                                end = c(5, 15, 25)),
               strand = "+")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25), "+")
coveragePerTiling(grl, RFP)

}
\seealso{
Other ExtendGenomicRanges: \code{\link{asTX}},
  \code{\link{overlapsToCoverage}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{txSeqsFromFa}}, \code{\link{windowPerGroup}}
}
\concept{ExtendGenomicRanges}
