% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\docType{class}
\name{covRle-class}
\alias{covRle-class}
\title{Coverage Rle for both strands or single}
\value{
a covRLE object
}
\description{
Given a run of coverage(x) where x are reads,
this class combines the 2 strands into 1 object
}
\seealso{
Other covRLE: 
\code{\link{covRle}},
\code{\link{covRleFromGR}()},
\code{\link{covRleList}},
\code{\link{covRleList-class}}
}
\concept{covRLE}
