% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRLE_class.R
\name{covRle}
\alias{covRle}
\title{Coverage Rlelist for both strands}
\usage{
covRle(forward = RleList(), reverse = RleList())
}
\arguments{
\item{forward}{a RleList with defined seqinfo for forward strand counts}

\item{reverse}{a RleList with defined seqinfo for reverse strand counts}
}
\value{
a covRle object
}
\description{
Coverage Rlelist for both strands
}
\examples{
covRle()
covRle(RleList(), RleList())
chr_rle <- RleList(chr1 = Rle(c(1,2,3), c(1,2,3)))
covRle(chr_rle, chr_rle)
}
\seealso{
Other covRLE: 
\code{\link{covRle-class}},
\code{\link{covRleFromGR}()},
\code{\link{covRleList}},
\code{\link{covRleList-class}}
}
\concept{covRLE}
