% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create.experiment}
\alias{create.experiment}
\title{Create a template for new ORFik experiment}
\usage{
create.experiment(dir, exper, saveDir = NULL, types = c("bam", "bed",
  "wig"), txdb = "", fa = "", viewTemplate = TRUE)
}
\arguments{
\item{dir}{Which directory to create experiment from}

\item{exper}{Short name of experiment, max 5 characters long}

\item{saveDir}{Directory to save experiment csv file (NULL)}

\item{types}{Default (bam, bed, wig), which types of libraries to allow}

\item{txdb}{A path to gff/gtf file used for libraries}

\item{fa}{A path to fasta genome/sequences used for libraries}

\item{viewTemplate}{run View() on template when finished, default (TRUE)}
}
\value{
a data.frame, NOTE: this is not a ORFik experiment,
 only a template for it!
}
\description{
By using files in a folder. It will try to make an experiment table
with information per sample. You will have to fill in the details
that were not autodetected. Easiest to do in csv editor like libre Office
or excel. Remember that each row (sample) must have a unique combination
of values.
}
\examples{
template <- create.experiment(dir = system.file("extdata", "", package = "ORFik"),
                              exper = "ORFik", txdb = system.file("extdata",
                                    "annotations.gtf",
                                    package = "ORFik"),
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
}
