% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create.experiment}
\alias{create.experiment}
\title{Create a template for new ORFik experiment}
\usage{
create.experiment(dir, exper, saveDir = NULL, types = c("bam", "bed",
  "wig"), txdb = "", fa = "")
}
\arguments{
\item{dir}{Which directory to create experiment from}

\item{exper}{Short name of experiment, max 5 characters long}

\item{saveDir}{Directory to save experiment csv file (NULL)}

\item{types}{Default (bam, bed, wig), which types of libraries to allow}

\item{txdb}{A path to gff/gtf file used for libraries}

\item{fa}{A path to fasta genome/sequences used for libraries}
}
\value{
a data.frame, NOTE: this is not a ORFik experiment,
 only a template for it!
}
\description{
By using files in a folder. You will have to fill in the details
that were not autodetected.
}
