% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{diff_exp_codon}
\alias{diff_exp_codon}
\title{Differential codon dwell time analysis}
\usage{
diff_exp_codon(
  dt,
  min_ratio_change = 1.7,
  min_total_N_codons = 100,
  exclude_start_stop = FALSE,
  codon_score = "percentage",
  background = NULL
)
}
\arguments{
\item{dt}{a data.table created from ORFik codon dwell time analysis}

\item{min_ratio_change}{Fold change minimum, default: 1.7 (bi-directional)}

\item{min_total_N_codons}{= 100}

\item{exclude_start_stop}{logical, default FALSE. If TRUE, remove start codons (i.e. # and *)}

\item{codon_score}{character, default columns to use in dt,
default "percentage" (i.e. relative_to_max_score)}

\item{background}{character or NULL, default NULL. If character, must be subset
of levels in dt$variable, and differential usage will only be done with these
levels as denominator factor level (i.e. only compared too.)}
}
\value{
a data.table with additional attributes
}
\description{
Will create a fold change (ratio score) over your defined 'codon_score'.\cr
It then runs a wilcox p-value test on the differential dwell times of your scores.
It is set to signifcant = TRUE, if p-value < 0.05 & codon > min_total_N_codons &
(score > min_ratio_change | score < 1/min_ratio_change)
}
\details{
If a more rigerous model with estimated interaction terms between P and A sites is needed,
we advice RiboDT (see references).
}
\examples{
df <- ORFik.template.experiment()[9:10,] # Subset to 2 Ribo-seq libs
## For multiple libs
dt <- codon_usage_exp(df, outputLibs(df, type = "pshifted", output.mode = "list"),
                 min_counts_cds_filter = 10)
dt_diff_exp <- diff_exp_codon(dt, min_total_N_codons = 10)
dt_diff_exp[significant == TRUE,]
diff_exp_codon_plot(dt_diff_exp)
}
\references{
https://doi.org/10.1016/j.ymeth.2021.10.004
}
\seealso{
Other codon: 
\code{\link{codon_usage}()},
\code{\link{codon_usage_exp}()},
\code{\link{codon_usage_plot}()},
\code{\link{diff_exp_codon_plot}()}
}
\concept{codon}
