% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_usage.R
\name{diff_exp_codon_plot}
\alias{diff_exp_codon_plot}
\title{Differential codon dwell time plot}
\usage{
diff_exp_codon_plot(
  dt,
  min_total_N_codons = ifelse(!is.null(attr(dt, "min_total_N_codons")), attr(dt,
    "min_total_N_codons"), 100L),
  only_significant_difexp = FALSE,
  add_plotly_tooltip = TRUE
)
}
\arguments{
\item{dt}{a data.table created from ORFik codon dwell time analysis}

\item{min_total_N_codons}{integer, default
ifelse(!is.null(attr(dt, "min_total_N_codons")), attr(dt, "min_total_N_codons"), 100L).
How many codons are minimum to not color it red (i.e. "the not enough codons" group).}

\item{only_significant_difexp}{logical, default FALSE. If TRUE, don't show non
significant codons.}

\item{add_plotly_tooltip}{logical, default TRUE.
If you code \code{plotly::ggplotly(plot, tooltip = "text")}, you will get a nice
tooltip by default.}
}
\value{
a ggplot object
}
\description{
Uses the wilcox p-values and thresholds to create a ggplot dotplot,
with significance bars and color coded dots. Also mono font, to make sure
all letters are equal size.
}
\details{
If you want interactive plots, wrap output in
plotly::ggplotly(diff_exp_ggplot, tooltip = "text)
}
\examples{
# See ?diff_exp_codon
}
\seealso{
Other codon: 
\code{\link{codon_usage}()},
\code{\link{codon_usage_exp}()},
\code{\link{codon_usage_plot}()},
\code{\link{diff_exp_codon}()}
}
\concept{codon}
