% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{disengagementScore}
\alias{disengagementScore}
\title{Disengagement score (DS)}
\usage{
disengagementScore(grl, RFP, GtfOrTx)
}
\arguments{
\item{grl}{a GRangesList object with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{GtfOrTx}{if Gtf: a TxDb object of a gtf file,
if tx: a GrangesList of transcripts, called from:
exonsBy(Gtf, by = "tx", use.names = T)}
}
\value{
a named vector of numeric values of scores
}
\description{
is defined as (RPFs over ORF)/(RPFs downstream to tx end).
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1242/dev.098344
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
