% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods2.R
\name{distToCds}
\alias{distToCds}
\title{Get distances between ORF ends and starts of their transcripts cds'.}
\usage{
distToCds(ORFs, fiveUTRs, cds = NULL, extension = NULL)
}
\arguments{
\item{ORFs}{orfs as \code{\link[GenomicRanges]{GRangesList}},
names of orfs must be transcript names}

\item{fiveUTRs}{fiveUTRs as \code{\link[GenomicRanges]{GRangesList}},
must be original unchanged fiveUTRs}

\item{cds}{cds' as \code{\link[GenomicRanges]{GRangesList}},
only add if you used CageSeq to extend leaders}

\item{extension}{Numeric that needs set to 0 if you did not use CageSeq
if you used CageSeq to change tss' when finding the orfs, standard cage
extension is 1000.}
}
\value{
an integer vector, +1 means one base upstream of cds, -1 means
2nd base in cds, 0 means orf stops at cds start.
}
\description{
Will calculate distance between each ORF end and begining of the
corresponding cds. Matching is done by transcript names.
This is applicable practically to the upstream (fiveUTRs) ORFs.
The cds start site, will be presumed to be on + 1 of end of fiveUTRs.
}
\examples{
grl <- GRangesList(tx1_1 = GRanges("1", IRanges(1, 10), "+"))
fiveUTRs <- GRangesList(tx1 = GRanges("1", IRanges(1, 20), "+"))
distToCds(grl, fiveUTRs, extension = 0)

}
\references{
doi: 10.1074/jbc.R116.733899
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
