% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{download.SRA.metadata}
\alias{download.SRA.metadata}
\title{Downloads metadata from SRA}
\usage{
download.SRA.metadata(SRP, outdir, remove.invalid = TRUE)
}
\arguments{
\item{SRP}{a string, a study ID as either the SRP, ERP or PRJ of the study,
examples would be "SRP226389" or "ERP116106".}

\item{outdir}{directory to save file,
The file will be called "SraRunInfo_SRP.csv", where SRP is
the SRP argument.
The directory will be created if not existing.}

\item{remove.invalid}{logical, default TRUE. Remove Runs with 0 reads (spots)}
}
\value{
a data.table of the opened file
}
\description{
Downloads metadata from SRA
}
\examples{
## Originally on SRA
# outdir <- tempdir() # Specify output directory
# download.SRA.metadata("SRP226389", outdir)
## ORiginally on ENA
# download.SRA.metadata("ERP116106", outdir)
}
