% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{download.SRA}
\alias{download.SRA}
\title{Download SRR libraries from SRA}
\usage{
download.SRA(
  info,
  outdir,
  rename = TRUE,
  fastq.dump.path = install.sratoolkit(),
  settings = paste("--gzip", "--skip-technical", "--split-files"),
  subset = NULL,
  BPPARAM = bpparam()
)
}
\arguments{
\item{info}{a data.frame with information and SRR numbers or character vector of only SRR numbers.
If only SRR numbers can not rename, since no additional information is given.}

\item{outdir}{a string, default: cbu server}

\item{rename}{logical, default TRUE. Rename files according to info, only if info has a column called
`Experiment Title`}

\item{fastq.dump.path}{path to fastq-dump binary, default: path returned
from install.sratoolkit()}

\item{settings}{a string of arguments for fastq-dump,
default: paste("--gzip", "--skip-technical", "--split-files")}

\item{subset}{a numeric or NULL, default NULL (no subset). If defined as
a numeric will download only the first n reads specified by subset.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}}
}
\value{
a character vector of download files filepaths
}
\description{
Multicore version download, see documentation for SRA toolkit for more information.
}
\examples{
\dontrun{
## Simple single SRR run of YEAST
SRR <- c("SRR453566") # Can be more than one
outdir <- "/export/valenfs/data/raw_data/TCP-seq/testtest1/"
# Download, get 5 first reads
download.SRA(SRR, outdir, subset = 5)

## With additional info for renaming, info download manually from SRA
library(data.table)
# The SRR numbers as a list in file downloaded manually
path <- "/export/valenfs/data/raw_data/TCP-seq/sel-TCPSeq_Wagner/SRR.ids"
# The sra results as a file downloaded manually
info <- fread("/export/valenfs/data/raw_data/TCP-seq/sel-TCPSeq_Wagner/sra_result.csv")
SRR <- unlist(read.table(path,
                         stringsAsFactors = F), use.names = FALSE)
info <- cbind(info, SRR)
info <- info[1,]
# Download, 5 first reads
download.SRA(info, outdir, subset = 5)
}
}
\references{
https://ncbi.github.io/sra-tools/fastq-dump.html
}
