% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{download.SRA}
\alias{download.SRA}
\title{Download SRR libraries from SRA}
\usage{
download.SRA(
  info,
  outdir,
  rename = TRUE,
  fastq.dump.path = install.sratoolkit(),
  settings = paste("--gzip", "--skip-technical", "--split-files"),
  subset = NULL,
  BPPARAM = bpparam()
)
}
\arguments{
\item{info}{character vector of only SRR numbers or
a data.frame with SRA metadata information including the SRR numbers in a column called
"Run" or "SRR".
If only SRR numbers can not rename, since no additional information is given.}

\item{outdir}{a string, default: cbu server}

\item{rename}{logical, default TRUE. Rename files according to info, only if info has a column called
`Experiment Title`}

\item{fastq.dump.path}{path to fastq-dump binary, default: path returned
from install.sratoolkit()}

\item{settings}{a string of arguments for fastq-dump,
default: paste("--gzip", "--skip-technical", "--split-files")}

\item{subset}{a numeric or NULL, default NULL (no subset). If defined as
a numeric will download only the first n reads specified by subset.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}}
}
\value{
a character vector of download files filepaths
}
\description{
Multicore version download, see documentation for SRA toolkit for more information.
}
\examples{
\dontrun{
## Simple single SRR run of YEAST
SRR <- c("SRR453566") # Can be more than one
outdir <- tempdir() # Specify output directory
# Download, get 5 first reads
download.SRA(SRR, outdir, subset = 5)

## Using metadata column to get SRR numbers and to be able to rename samples
info <- download.SRA.metadata("SRP226389") # By study id
outdir <- tempdir() # Specify output directory
# Download, 5 first reads of each library and rename
download.SRA(info, outdir, subset = 5)
}
}
\references{
https://ncbi.github.io/sra-tools/fastq-dump.html
}
