% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{downstreamOfPerGroup}
\alias{downstreamOfPerGroup}
\title{get rest of objects downstream}
\usage{
downstreamOfPerGroup(tx, downstreamOf)
}
\arguments{
\item{tx}{a GRangesList, usually of Transcripts to be changed}

\item{downstreamOf}{a vector of integers, for each group in tx, where
is the new start point of first valid exon.}
}
\description{
per group get the part downstream of position
 defined in downstreamOf
 downstreamOf(tx, ORFik:::ORFStopSites(cds, asGR = F))
 will return the 3' utrs per transcript as GRangesList,
 usually used for interesting
 parts of the transcripts, like upstream open reading frames(uorf).
 downstreamOf +/- 1 is start/end site
 of transformed tx's, depending on strand
}
