% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_Analysis.R
\name{DTEG.analysis}
\alias{DTEG.analysis}
\title{Run differential TE analysis}
\usage{
DTEG.analysis(
  df.rfp,
  df.rna,
  output.dir = paste0(dirname(df.rfp$filepath[1]), "/QC_STATS/"),
  design = "stage",
  p.value = 0.05,
  plot.title = "",
  width = 6,
  height = 6,
  dot.size = 0.4
)
}
\arguments{
\item{df.rfp}{a \code{\link{experiment}} of Ribo-seq or 80S from TCP-seq.}

\item{df.rna}{a \code{\link{experiment}} of RNA-seq}

\item{output.dir}{output.dir directory to save plots,
plot will be named "TE_between.png". If NULL, will not save.}

\item{design}{a character vector, default "stage". The columns in the
experiment that creates the comparison contrasts.}

\item{p.value}{a numeric, default 0.05 in interval (0,1).
What p-value used for the analysis?}

\item{plot.title}{title for plots, usually name of experiment etc}

\item{width}{numeric, default 6 (in inches)}

\item{height}{numeric, default 6 (in inches)}

\item{dot.size}{numeric, default 0.4, size of point dots in plot.}
}
\value{
a data.table with 8 columns.
}
\description{
Using an equal reimplementation of the deltaTE algorithm (see reference).
You need at least 2 groups and 2 replicates per group. The Ribo-seq will
be over CDS and RNA-seq over mRNAs, per transcript, subset to genes if you
do not need isoform variants.
}
\details{
What the deltaTE plot calls intensified is here called mRNA abundance and
forwarded is called Buffering.
}
\examples{
#df.rfp <- read.experiment("Riboseq")
#df.rna <- read.experiment("RNAseq")
#DTEG.analysis(df.rfp, df.rna)
}
\references{
doi: 10.1002/cpmb.108
}
