% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_Analysis.R
\name{DTEG.analysis}
\alias{DTEG.analysis}
\title{Run differential TE analysis}
\usage{
DTEG.analysis(
  df.rfp,
  df.rna,
  output.dir = paste0(dirname(df.rfp$filepath[1]), "/QC_STATS/"),
  design = "stage",
  p.value = 0.05,
  plot.title = "",
  width = 6,
  height = 6,
  dot.size = 0.4
)
}
\arguments{
\item{df.rfp}{a \code{\link{experiment}} of Ribo-seq or 80S from TCP-seq.}

\item{df.rna}{a \code{\link{experiment}} of RNA-seq}

\item{output.dir}{output.dir directory to save plots,
plot will be named "TE_between.png". If NULL, will not save.}

\item{design}{a character vector, default "stage". The columns in the
experiment that creates the comparison contrasts.}

\item{p.value}{a numeric, default 0.05 in interval (0,1)
or "" to not show.
What p-value used for the analysis? Will be shown as a caption.}

\item{plot.title}{title for plots, usually name of experiment etc}

\item{width}{numeric, default 6 (in inches)}

\item{height}{numeric, default 6 (in inches)}

\item{dot.size}{numeric, default 0.4, size of point dots in plot.}
}
\value{
a data.table with 9 columns.
}
\description{
Using an equal reimplementation of the deltaTE algorithm (see reference).
You need at least 2 groups and 2 replicates per group. The Ribo-seq counts will
be over CDS and RNA-seq over mRNAs, per transcript. \cr
If you do not need isoform variants, subset to longest isoform in
the returned object.
}
\details{
Creates a total of 3 DESeq models (given x is design argument input
and libraryType is RNA-seq and Ribo-seq):\cr
1. Ribo-seq model: design = ~ x (differences between the x groups in Ribo-seq)\cr
2. RNA-seq model: design = ~ x (differences between the x groups in RNA-seq)\cr
3. TE model: design = ~ library type + x + libraryType + libraryType:x
(differences between the x and libraryType groups and the interaction between them)\cr
\cr The LFC values are shrunken by lfcShrink(type = "normal").\cr \cr
What the deltaTE plot calls intensified is here called mRNA abundance and
forwarded is called Buffering.
}
\examples{
#df.rfp <- read.experiment("Riboseq")
#df.rna <- read.experiment("RNAseq")
#DTEG.analysis(df.rfp, df.rna)
}
\references{
doi: 10.1002/cpmb.108
}
\seealso{
Other TE: 
\code{\link{DTEG.plot}()},
\code{\link{te.table}()},
\code{\link{te_rna.plot}()}
}
\concept{TE}
