% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE_plots.R
\name{DTEG.plot}
\alias{DTEG.plot}
\title{Plot DTEG result}
\usage{
DTEG.plot(
  dt,
  output.dir = NULL,
  p.value = 0.05,
  plot.title = "",
  width = 6,
  height = 6,
  dot.size = 0.4,
  xlim = c(-5, 5),
  ylim = c(-10, 10),
  relative.name = "DTEG_plot.png"
)
}
\arguments{
\item{dt}{a data.table with the results from \code{\link{DTEG.analysis}}}

\item{output.dir}{a character path, default NULL(no save), or a directory
to save to a file will be called "DTEG_plot.png"}

\item{p.value}{a numeric, default 0.05 in interval (0,1)
or "" to not show.
What p-value used for the analysis? Will be shown as a caption.}

\item{plot.title}{title for plots, usually name of experiment etc}

\item{width}{numeric, default 6 (in inches)}

\item{height}{numeric, default 6 (in inches)}

\item{dot.size}{numeric, default 0.4, size of point dots in plot.}

\item{xlim}{numeric vector, default c(-5, 5)}

\item{ylim}{numeric vector, default c(-10, 10)}

\item{relative.name}{character, Default: "DTEG_plot.png".
Relative name of file to be saved in folder specified in output.dir.
Change to .pdf if you want pdf file instead of png.}
}
\value{
a ggplot object
}
\description{
Plot DTEG result
}
\examples{
#df.rfp <- read.experiment("Riboseq")
#df.rna <- read.experiment("RNAseq")
#dt <- DTEG.analysis(df.rfp, df.rna)
#DTEG.plot(dt, xlim = c(-2, 2), ylim = c(-2, 2))
}
\seealso{
Other TE: 
\code{\link{DTEG.analysis}()},
\code{\link{te.table}()},
\code{\link{te_rna.plot}()}
}
\concept{TE}
