% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{entropy}
\alias{entropy}
\title{Calucalte Entropy value of input reads overlapping per group in grl
For example reads per orf.
The entropy interval per group is a real number in the interval (0:1)
Where 0 is no variance in reads over group.}
\usage{
entropy(grl, reads)
}
\arguments{
\item{grl}{a GRangesList that the reads can map to}

\item{reads}{a GAlignment object, ig. from ribo seq, or rna seq}
}
\value{
A numeric vector containing entropy per ORF
}
\description{
Calucalte Entropy value of input reads overlapping per group in grl
For example reads per orf.
The entropy interval per group is a real number in the interval (0:1)
Where 0 is no variance in reads over group.
}
