% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{entropy}
\alias{entropy}
\title{Calucalte Entropy value of input reads overlapping}
\usage{
entropy(grl, reads)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} that the reads can map to}

\item{reads}{a GAlignment object, ig. from ribo seq, or rna seq,
can also be GRanges or GRangesList}
}
\value{
A numeric vector containing entropy per ORF
}
\description{
For example reads per orf.
The entropy interval per group is a real number in the interval (0:1)
Where 0 is no variance in reads over group.
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
