% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exists.ftp.dir.fast}
\alias{exists.ftp.dir.fast}
\title{A fast ftp directory check}
\usage{
exists.ftp.dir.fast(url.dir, report.error = FALSE)
}
\arguments{
\item{url.dir}{character, url to a ftp directory.}

\item{report.error}{logical, FALSE. If TRUE,
stop and report error.}
}
\value{
logical, TRUE if url directory exists
}
\description{
Check if ftp directory exists
}
