% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exists.ftp.file.fast}
\alias{exists.ftp.file.fast}
\title{A fast ftp file check}
\usage{
exists.ftp.file.fast(url, report.error = FALSE)
}
\arguments{
\item{url}{character, url to a ftp file}

\item{report.error}{logical, FALSE. If TRUE,
stop and report error.}
}
\value{
logical, TRUE if file exists
}
\description{
Check if ftp file exists
}
