% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export.bedo}
\alias{export.bedo}
\title{Store GRanges object as .bedo}
\usage{
export.bedo(object, out)
}
\arguments{
\item{object}{a GRanges object}

\item{out}{a character, location on disc (full path)}
}
\value{
NULL, object saved to disc
}
\description{
.bedo is .bed ORFik, an optimized bed format for coverage reads with read lengths
.bedo is a text based format with columns:
chromosome, start, stop, width, strand, (cigar # M's, match/mismatch total)
, duplicates of that read
}
