% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exports.R
\name{export.bedo}
\alias{export.bedo}
\title{Store GRanges object as .bedo}
\usage{
export.bedo(object, out)
}
\arguments{
\item{object}{a GRanges object}

\item{out}{a character, location on disc (full path)}
}
\value{
NULL, object saved to disc
}
\description{
.bedo is .bed ORFik, an optimized bed format for coverage reads with
read lengths .bedo is a text based format with columns (6 maximum):\cr
1. chromosome\cr 2. start\cr 3. end\cr 4. strand\cr
5. ref width (cigar # M's, match/mismatch total)\cr
6. duplicates of that read\cr
}
\details{
Positions are 1-based, not 0-based as .bed.
End will be removed if all ends equals all starts.
Import with import.bedo
}
