% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{extendsTSSexons}
\alias{extendsTSSexons}
\title{Extend first exon of each transcript with length specified}
\usage{
extendsTSSexons(fiveUTRs, extension = 1000)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}

\item{extension}{The number of basses upstream to add on transcripts}
}
\value{
granges object of first exons
}
\description{
Extend first exon of each transcript with length specified
}
