% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{extendsTSSExons}
\alias{extendsTSSExons}
\title{Extend first exon of each transcript with length specified}
\usage{
extendsTSSExons(fiveUTRs, extension = 1000)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}

\item{extension}{The number of basses upstream to add on transcripts}
}
\description{
Extend first exon of each transcript with length specified
}
