% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{filepath}
\alias{filepath}
\title{Get filepaths to ORFik experiment}
\usage{
filepath(df, type, basename = FALSE)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "ofst" or "pshifted".
These are made with ORFik:::convertLibs() or shiftFootprintsByExperiment().
Can also be custom user made folders inside the experiments bam folder.
It acts in a recursive manner with priority: If you state "pshifted",
but it does not exist, it checks "ofst". If no .ofst files, it uses
"default", which always must exists.}

\item{basename}{logical, default (FALSE).
Get relative paths instead of full. Only use for inspection!}
}
\value{
a character vector of paths, or a list of character with 2 paths per,
if paired libraries exists
}
\description{
If other type than "default" is given and that type is not found,
it will return you ofst files, if they do not exist, then
default filepaths without warning. \cr
}
\details{
For pshifted libraries, it will load ".bedo"
prioritized over ".bed", if there exists both file types for the same file.
}
\examples{
df <- ORFik.template.experiment()
filepath(df, "default")
# If you have bedo files, see simpleLibs():
# filepath(df, "bedo")
# If you have pshifted files, see shiftFootprintsByExperiment():
# filepath(df, "pshifted")
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment.zf}()},
\code{\link{ORFik.template.experiment}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{libraryTypes}()},
\code{\link{organism,experiment-method}},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
