% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{filterCage}
\alias{filterCage}
\title{Filter peak of cage-data by value}
\usage{
filterCage(rawCage, filterValue = 1)
}
\arguments{
\item{rawCage}{The raw cage-data}

\item{filterValue}{The number of counts(score) to filter on
for a tss to pass as hit}
}
\value{
the filtered Granges object
}
\description{
Filter peak of cage-data by value
}
