% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterTranscripts}
\alias{filterTranscripts}
\title{Get the transcripts with accepted lengths of leaders, cds and trailer.}
\usage{
filterTranscripts(txdb, minFiveUTR = 30L, minCDS = 150L,
  minThreeUTR = 30L, stopOnEmpty = TRUE)
}
\arguments{
\item{txdb}{a TxDb file or a path to one of:
(.gtf ,.gff, .gff2, .gff2, .db or .sqlite)}

\item{minFiveUTR}{(integer) minimum bp for 5' UTR during filtering for the
transcripts. Set to NULL if no 5' UTRs exists for annotation.}

\item{minCDS}{(integer) minimum bp for CDS during filtering for the
transcripts}

\item{minThreeUTR}{(integer) minimum bp for 3' UTR during filtering for the
transcripts. Set to NULL if no 3' UTRs exists for annotation.}

\item{stopOnEmpty}{logical TRUE, stop if no valid names are found ?}
}
\value{
a character vector of valid tramscript names
}
\description{
Filter transcripts to those who have leaders, CDS, trailers of some lengths,
pick the longest per gene.
}
\details{
If a transcript does not have a trailer, then the length is 0,
so they will be filtered out. So only transcripts with leaders, cds and
trailers will be returned. You can set the integer to 0, that will return
all within that group.

If your annotation does not have leaders or trailers, set them to NULL.
}
\examples{
gtf_file <- system.file("extdata", "annotations.gtf", package = "ORFik")
txdb <- GenomicFeatures::makeTxDbFromGFF(gtf_file)
txNames <- filterTranscripts(txdb)

}
