% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fimport}
\alias{fimport}
\title{Load any type of sequencing reads}
\usage{
fimport(path, chrStyle = NULL)
}
\arguments{
\item{path}{a character path to file or a GRanges/Galignment object etc.
Any Ranged object.}

\item{chrStyle}{a GRanges object, or a character style (Default: NULL) to
get seqlevelsStyle from. Is chromosome 1 called chr1 or 1, is mitocondrial
chromosome called MT or chrM etc. Will use 1st seqlevel-
style if more are present. Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a GAlignment/GRanges object depending on input.
}
\description{
Wraps around rtracklayer::import and tries to speed up loading with the
use of data.table. Supports gzip, gz, bgz and bed formats.
Also safer chromosome naming with the argument chrStyle
}
