% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{find_in_frame_ORFs}
\alias{find_in_frame_ORFs}
\title{Creates list of IRanges with Open Reading Frames.}
\usage{
find_in_frame_ORFs(fastaSeq, startCodon = "ATG", stopCodon = "TAA|TAG|TGA",
  longestORF = F, minimumLength = 0)
}
\arguments{
\item{fastaSeq}{DNA sequence to search for Open Reading Frames.}

\item{startCodon}{string. Default is "ATG".}

\item{stopCodon}{string. Default is "TAA|TAG|TGA".}

\item{longestORF}{bolean. Default FALSE. Defines whether pick longest ORF only.
When FALSE will report all open reaidng frames, even overlapping small ones.
Currently not working - in preparation!}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs to be at least START + 8*3 [bp] + STOP.}
}
\value{
A List of IRanges objects of ORFs.
}
\description{
Creates list of IRanges with Open Reading Frames.
}
\examples{
#find_in_frame_ORFs()

}

