% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{find_url_ebi}
\alias{find_url_ebi}
\title{Locates and check if fastq files exists in ebi}
\usage{
find_url_ebi(SRR, stop.on.error = FALSE)
}
\arguments{
\item{SRR}{character, SRR, ERR or DRR numbers.}

\item{stop.on.error}{logical FALSE, if true will stop
if all files are not found.}
}
\value{
full url to fastq files, same length as input
(2 urls for paired end data). Returns empty character() if all
files not found.
}
\description{
Look for files in ebi following url: ftp://ftp.sra.ebi.ac.uk/vol1/fastq
Paired end and single end fastq files.\cr
EBI uses 3 ways to organize data inside vol1/fastq:\cr
- 1: Most common: SRR(3 first)/0(2 last)/whole\cr
- 2: less common: SRR(3 first)/00(1 last)/whole\cr
- 3: least common SRR(3 first)/whole
}
\examples{
# Test the 3 ways to get fastq files from EBI
# Both single end and paired end data

# Most common: SRR(3 first)/0(2 last)/whole
# Single
ORFik:::find_url_ebi("SRR10503056")
# Paired
ORFik:::find_url_ebi("SRR10500056")

# less common: SRR(3 first)/00(1 last)/whole
# Single
#ORFik:::find_url_ebi("SRR1562873")
# Paired
#ORFik:::find_url_ebi("SRR1560083")
# least common SRR(3 first)/whole
# Single
#ORFik:::find_url_ebi("SRR105687")
# Paired
#ORFik:::find_url_ebi("SRR105788")
}
