% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{findCageUTRFivelen}
\alias{findCageUTRFivelen}
\title{get length of leaders ordered after oldTxNames
Normally only a helper function for ORFik}
\usage{
findCageUTRFivelen(fiveUTRs, oldTxNames)
}
\arguments{
\item{fiveUTRs}{a GRangesList object of leaders}

\item{oldTxNames}{a character vector of names to group fiveUTRs by.}
}
\description{
get length of leaders ordered after oldTxNames
Normally only a helper function for ORFik
}
