% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFa}
\alias{findFa}
\title{Convenience wrapper for Rsamtools FaFile}
\usage{
findFa(faFile)
}
\arguments{
\item{faFile}{a character path or FaFile}
}
\value{
a FaFile or BSgenome
}
\description{
Convenience wrapper for Rsamtools FaFile
}
\seealso{
Other utils: \code{\link{bedToGR}},
  \code{\link{convertToOneBasedRanges}},
  \code{\link{fread.bed}}, \code{\link{is.gr_or_grl}},
  \code{\link{is.grl}}, \code{\link{validGRL}}
}
\concept{utils}
