% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{findFromPath}
\alias{findFromPath}
\title{Find all candidate library types filenames}
\usage{
findFromPath(filepaths, candidates = c("RNA", "rna-seq", "Rna-seq",
  "RFP", "RPF", "ribo-seq", "Ribo-seq", "mrna", "CAGE", "cage", "LSU",
  "SSU", "ATAC", "tRNA", "SHAPE"))
}
\arguments{
\item{filepaths}{path to all files}

\item{candidates}{Possible names to search for.}
}
\value{
a candidate library types (character vector)
}
\description{
Find all candidate library types filenames
}
