% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{findFromPath}
\alias{findFromPath}
\title{Find all candidate library types filenames}
\usage{
findFromPath(filepaths, candidates)
}
\arguments{
\item{filepaths}{path to all files}

\item{candidates}{a data.table with 2 columns,
Possible names to search for, see experiment_naming family for candidates.}
}
\value{
a candidate library types (character vector)
}
\description{
From the given \code{\link{experiment}}
}
