% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{findLibrariesInFolder}
\alias{findLibrariesInFolder}
\title{Get all library files in folder/folders of given types}
\usage{
findLibrariesInFolder(dir, types)
}
\arguments{
\item{dir}{The directory/directories to find bam, bed, wig files.}

\item{types}{All accepted types of bam, bed, wig files..}
}
\value{
(data.table) All files found from types parameter.
With 2 extra column (logical), is it wig pairs, and paired bam files.
}
\description{
Will try to guess paired / unpaired wig, bed, bam files.
}
