% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{findLibrariesInFolder}
\alias{findLibrariesInFolder}
\title{Get all library files in folder/folders of given types}
\usage{
findLibrariesInFolder(dir, types, pairedEndBam = FALSE)
}
\arguments{
\item{dir}{Which directory / directories to create experiment from}

\item{types}{Default (bam, bed, wig), which types of libraries to allow}

\item{pairedEndBam}{logical FALSE, else TRUE, or a logical list of
TRUE/FALSE per library you see will be included (run first without and check
what order the files will come in) 1 paired end file, then two single will
be c(T, F, F)}
}
\value{
(data.table) All files found from types parameter.
With 2 extra column (logical), is it wig pairs, and paired bam files.
}
\description{
Will try to guess paired / unpaired wig, bed, bam files.
}
\details{
Set pairedEndBam if you have paired end reads as a single bam file.
}
