% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findMapORFs}
\alias{findMapORFs}
\title{Find ORFs and immediately map them to their genomic positions.}
\usage{
findMapORFs(grl, seqs, startCodon = startDefinition(1),
  stopCodon = stopDefinition(1), longestORF = FALSE, minimumLength = 0)
}
\arguments{
\item{grl}{(\code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}}) of sequences
to search for ORFs, probably in genomic coordinates}

\item{seqs}{(DNAStringSet or character) DNA sequences to search for Open
Reading Frames.}

\item{startCodon}{(character) Possible START codons to search for. Check
\code{\link[=startDefinition]{startDefinition()}} for helper function.}

\item{stopCodon}{(character) Possible STOP codons to search for. Check
\code{\link[=stopDefinition]{stopDefinition()}} for helper function.}

\item{longestORF}{(logical) Default FALSE. When TRUE will only report ORFs
that are longest, all smaller overlapping ORFs will be ignored.
When FALSE will report all possible ORFs in all three reading frames.}

\item{minimumLength}{(integer) Default is 0. Minimum length of ORF, without
counting 3bp for START and STOP codons. For example minimumLength = 8 will
result in size of ORFs to be at least START + 8*3 (bp) + STOP.
Use this param to restrict search.}
}
\value{
A GRangesList of ORFs.
}
\description{
Finds ORFs on the sequences of interest, but returns relative positions to
the positions of \code{grl} argument. For example, \code{grl} can be exons
of known transcripts (with genomic coordinates), and \code{seq} sequences of
those transcripts, in that case, \code{\link[=findMapORFs]{findMapORFs()}} will return
genomic coordinates of ORFs found on transcript sequences.
}
\details{
This function assumes that \code{seq} is in widths relative to \code{grl},
and that their orders match.
}
\examples{
# This sequence has ORFs at 1-9 and 4-9
seqs <- c("ATGATGTAA") # the dna sequence
findORFs(seqs)
# lets assume that this sequence comes from two exons as follows
gr <- GRanges(seqnames = rep("1", 2), # chromosome 1
              ranges = IRanges(start = c(21, 10), end = c(23, 15)),
              strand = rep("-", 2), names = rep("tx1", 2))
grl <- GRangesList(tx1 = gr)
findMapORFs(grl, seqs) # ORFs are properly mapped to its genomic coordinates

grl <- c(grl, grl)
names(grl) <- c("tx1", "tx2")
findMapORFs(grl, c(seqs, seqs))

}
\seealso{
\code{\link[=findORFs]{findORFs()}}, \code{\link[=findORFsFasta]{findORFsFasta()}},
\code{\link[=startDefinition]{startDefinition()}}, \code{\link[=stopDefinition]{stopDefinition()}}

Other findORFs: \code{\link{findORFsFasta}},
  \code{\link{findORFs}}, \code{\link{startDefinition}},
  \code{\link{stopDefinition}}
}
