% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findMapORFs}
\alias{findMapORFs}
\title{Find Open Reading Frames  mapped to genomic coordinates.}
\usage{
findMapORFs(grl, fastaSeqs, startCodon = "ATG", stopCodon = "TAA|TAG|TGA",
  longestORF = FALSE, minimumLength = 0)
}
\arguments{
\item{grl}{\code{\link[GenomicRanges]{GRangesList}} of sequences
to search for orfs, in Genomic coordinates}

\item{fastaSeqs}{DNA sequences to search for Open Reading Frames,
must be DNAStringSet.}

\item{startCodon}{string. Default is "ATG".}

\item{stopCodon}{string. Default is "TAA|TAG|TGA".}

\item{longestORF}{bolean. Default FALSE. When TRUE pick
longest ORF per group/chromosome.
When FALSE will report all open reaidng frames, even overlapping small ones.}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs to be at least
 START + 8*3 [bp] + STOP.}
}
\value{
A GRangesList of ORFs.
}
\description{
Input is a Grangeslist of regions to search, together with a DNAStringSet
/ character vector with fastaSequence in same order as the grl.
}
\examples{
seqs <- c("ATGGGTATTTATA") # the dna sequence
startCodons <- "ATG|TGG|GGG"
stopCodons <- "TAA|AAT|ATA"
## gr is sequence mapping
gr <- GRanges(seqnames = rep("1", 2),
                ranges = IRanges(start = c(21, 10), end = c(23, 19)),
                strand = rep("-", 2), names = rep("tx1_1", 2))
grl <- GRangesList(tx1 = gr)
## now run
findMapORFs(grl,seqs,startCodons, stopCodons,
                   longestORF = FALSE, minimumLength = 0)
## You will find 1 orf splitted on 2 exons (10-19 with 21-22)

}
