% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{findNewTSS}
\alias{findNewTSS}
\title{Finds max peaks per trancsript from reads in the cagefile}
\usage{
findNewTSS(fiveUTRs, cageData, extension)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}

\item{extension}{The number of basses upstream to add on transcripts}

\item{filePath}{The location of the cage-file}
}
\description{
Finds max peaks per trancsript from reads in the cagefile
}
