% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findORFs}
\alias{findORFs}
\title{Find Open Reading Frames.}
\usage{
findORFs(seqs, startCodon = startDefinition(1),
  stopCodon = stopDefinition(1), longestORF = FALSE, minimumLength = 0)
}
\arguments{
\item{seqs}{(DNAStringSet or character) DNA sequences to search for Open
Reading Frames.}

\item{startCodon}{(character) Possible START codons to search for. Check
[startDefinition()] for helper function.}

\item{stopCodon}{(character) Possible STOP codons to search for. Check
[stopDefinition()] for helper function.}

\item{longestORF}{(logical) Default FALSE. When TRUE will only report ORFs
that are longest, all smaller overlapping ORFs will be ignored.
When FALSE will report all possible ORFs in all three reading frames.}

\item{minimumLength}{(integer) Default is 0. Minimum length of ORF, without
counting 3bp for START and STOP codons. For example minimumLength = 8 will
result in size of ORFs to be at least START + 8*3 (bp) + STOP.
Use this param to restrict search.}
}
\value{
(IRangesList) of ORFs locations incuding START and STOP codons
grouped by input seqeunces.
}
\description{
Find all Open Reading Frames (ORFs) on the input sequences
in 5'- 3' direction, but within all three possible reading frames. For each
sequence of the input vector \code{\link{IRanges}} with START and
STOP positions (inclusive) will be returned as
\code{\link{IRangesList}}. Returned coordinates are relative to the
input sequences.
}
\examples{
findORFs("ATGTAA")
findORFs("ATGTTAA") # not in frame anymore

findORFs("ATGATGTAA") # two ORFs
findORFs("ATGATGTAA", longestORF = TRUE) # only longest of two above

findORFs(c("ATGTAA", "ATGATGTAA"))

}
\seealso{
[findMapORFs()], [findORFsFasta()],
[startDefinition()], [stopDefinition()]

Other findORFs: \code{\link{findMapORFs}},
  \code{\link{findORFsFasta}},
  \code{\link{startDefinition}},
  \code{\link{stopDefinition}}
}
