% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findORFsFasta}
\alias{findORFsFasta}
\title{Creates GRanges with Open Reading Frames from fasta files.}
\usage{
findORFsFasta(file, startCodon = "ATG", stopCodon = "TAA|TAG|TGA",
  longestORF = TRUE, minimumLength = 0, is.circular = FALSE)
}
\arguments{
\item{file}{- Path to fasta file.}

\item{startCodon}{Default is "ATG".}

\item{stopCodon}{Default is "TAA|TAG|TGA".}

\item{longestORF}{bolean. Default FALSE. When TRUE pick
longest ORF per group/chromosome.
When FALSE will report all open reaidng frames, even overlapping small ones.}

\item{minimumLength}{numeric. Default is 0.
For example minimumLength = 8 will result in size of ORFs
 to be at least START + 8*3 [bp] + STOP.}

\item{is.circular}{a logical(F), should ORFs that start before end
and after start be searched for.}
}
\value{
A GRanges object of ORFs mapped from fasta file.
 Each ORF includes START and STOP codon, seqname, strand and frame.
}
\description{
Each fasta header is treated separately, and name of the sequence will
be used as seqname in returned GRanges object.
Frame of the Open Reading Frame is also
returned in metadata column 'frame'.
}
\examples{
filePath <- system.file("extdata", "orfFindingExample.fasta",
package = "ORFik") ## location of the fasta file

findORFsFasta(filePath)

}
