% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{findORFsFasta}
\alias{findORFsFasta}
\title{Finds Open Reading Frames in fasta files.}
\usage{
findORFsFasta(filePath, startCodon = startDefinition(1),
  stopCodon = stopDefinition(1), longestORF = TRUE, minimumLength = 0,
  is.circular = FALSE)
}
\arguments{
\item{filePath}{(character) Path to the fasta file.}

\item{startCodon}{(character) Possible START codons to search for. Check
\code{\link{startDefinition}} for helper function.}

\item{stopCodon}{(character) Possible STOP codons to search for. Check
\code{\link{stopDefinition}} for helper function.}

\item{longestORF}{(logical) Default FALSE. When TRUE will only report ORFs
that are longest, all smaller overlapping ORFs will be ignored.
When FALSE will report all possible ORFs in all three reading frames.}

\item{minimumLength}{(integer) Default is 0. Minimum length of ORF, without
counting 3bp for START and STOP codons. For example minimumLength = 8 will
result in size of ORFs to be at least START + 8*3 [bp] + STOP.
Use this param to restrict search.}

\item{is.circular}{(logical) Whether the genome in filePath is circular.
Prokaryotic genomes are usually circular.}
}
\value{
(GRanges) object of ORFs mapped from fasta file. Positions are
relative to the fasta file.
}
\description{
Searches through each fasta header and reports all ORFs found for sense (+)
and antisense strand (-) in all frames. Name of the header will be used as
seqnames of reported ORFs.
Each fasta header is treated separately, and name of the sequence will
be used as seqname in returned GRanges object.
}
\examples{
# location of the example fasta file
example_genome <- system.file("extdata", "genome.fasta", package = "ORFik")
findORFsFasta(example_genome)

}
\seealso{
\code{\link{findORFs}}, \code{\link{findMapORFs}},
\code{\link{startDefinition}}, \code{\link{stopDefinition}}

Other findORFs: \code{\link{findMapORFs}},
  \code{\link{findORFs}}, \code{\link{startDefinition}},
  \code{\link{stopDefinition}}
}
