% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{firstEndPerGroup}
\alias{firstEndPerGroup}
\title{Get first end per granges group}
\usage{
firstEndPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
a Rle(keep.names = T), or integer vector(F)
}
\description{
grl must be sorted, call ORFik:::sortPerGroup if needed
}
