% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{firstExonPerGroup}
\alias{firstExonPerGroup}
\title{get first exon per GRangesList group
grl must be sorted, call ORFik:::sortPerGroup if needed}
\usage{
firstExonPerGroup(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}
}
\description{
get first exon per GRangesList group
grl must be sorted, call ORFik:::sortPerGroup if needed
}
