% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{firstStartPerGroup}
\alias{firstStartPerGroup}
\title{Get first start per granges group}
\usage{
firstStartPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
a Rle(keep.names = TRUE), or integer vector(FALSE)
}
\description{
grl must be sorted, call ORFik:::sortPerGroup if needed
}
\examples{
gr_plus <- GRanges(seqnames = c("chr1", "chr1"),
                   ranges = IRanges(c(7, 14), width = 3),
                   strand = c("+", "+"))
gr_minus <- GRanges(seqnames = c("chr2", "chr2"),
                    ranges = IRanges(c(4, 1), c(9, 3)),
                    strand = c("-", "-"))
grl <- GRangesList(tx1 = gr_plus, tx2 = gr_minus)
firstStartPerGroup(grl)
}
