% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{floss}
\alias{floss}
\title{Fragment Length Organization Similarity Score}
\usage{
floss(grl, RFP, cds, start = 26, end = 34)
}
\arguments{
\item{grl}{a GRangesList object with ORFs}

\item{RFP}{ribozomal footprints, given as Galignment or GRanges object}

\item{cds}{a GRangesList of coding sequences}

\item{start}{usually 26, the start of the floss interval}

\item{end}{usually 34, the end of the floss interval}
}
\value{
a vector of FLOSS of length same as grl
}
\description{
See article: 10.1016/j.celrep.2014.07.045
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{fpkmRFP}}, \code{\link{fpkmRNA}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
