% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{fpkm}
\alias{fpkm}
\title{Create normalizations of counts}
\usage{
fpkm(grl, reads, pseudoCount = 0)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
can be either transcripts, 5' utrs, cds', 3' utrs or
ORFs are a special case (uORFs, potential new cds' etc).}

\item{reads}{a GAlignment, GRanges or GRangesList object,
usually of ribo-seq, rna seq, cage seq, etc.}

\item{pseudoCount}{an integer, 0,
set it to 1 if you want to avoid NA and inf values.}
}
\value{
a numeric vector with the fpkm values
}
\description{
Short for: Fragments per kilobase of transcript per million fragments
 For ribo seq on orfs, send orfs as grl,
 for rna seq always send transcripts as grl
see article: 10.1038/nbt.1621
}
\examples{
ORF <- GRanges(seqnames = "1",
ranges = IRanges(start = c(1, 10, 20),
                end = c(5, 15, 25)),
strand = "+")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25),"+")
fpkm(grl, RFP)
}
\seealso{
Other features: \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
