% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{fpkm}
\alias{fpkm}
\title{Creates normalizations of the counts, normally,}
\usage{
fpkm(counts, lengthSize, librarySize)
}
\arguments{
\item{counts}{a list of integer counts per object}

\item{lengthSize}{a list of integer lengths per object}

\item{librarySize}{a numeric of size 1, the size of the library}
}
\description{
used in Translations efficiency calculations
}
