% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{fpkm}
\alias{fpkm}
\title{Create normalizations of overlapping read counts.}
\usage{
fpkm(grl, reads, pseudoCount = 0)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}} object
can be either transcripts, 5' utrs, cds', 3' utrs or
ORFs as a special case (uORFs, potential new cds' etc).}

\item{reads}{a GAlignment, GRanges or GRangesList object,
usually of RiboSeq, RnaSeq, CageSeq, etc.}

\item{pseudoCount}{an integer, by default is 0, set it to 1 if you want to
avoid NA and inf values.}
}
\value{
a numeric vector with the fpkm values
}
\description{
FPKM is short for "Fragments Per Kilobase of transcript per Million
fragments". When calculating RiboSeq data FPKM over ORFs use ORFs as
\code{grl}. When calculating RNASeq data FPKM use full transcripts as
\code{grl}.
}
\examples{
ORF <- GRanges(seqnames = "1",
               ranges = IRanges(start = c(1, 10, 20),
               end = c(5, 15, 25)),
               strand = "+")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25),"+")
fpkm(grl, RFP)

}
\references{
doi: 10.1038/nbt.1621
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
