% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fread.bed}
\alias{fread.bed}
\title{Load bed file as GRanges.}
\usage{
fread.bed(filePath)
}
\arguments{
\item{filePath}{The location of the bed file}
}
\value{
a GRanges object
}
\description{
Wraps around rtracklayer::import.bed and tries to speed up loading with the
use of data.table. Supports gzip, gz, bgz and bed formats.
}
\examples{
# path to example CageSeq data from hg19 heart sample
cageData <- system.file("extdata", "cage-seq-heart.bed.bgz",
                        package = "ORFik")
fread.bed(cageData)

}
