% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{fread.bed}
\alias{fread.bed}
\title{Get bed file from a file-path}
\usage{
fread.bed(filePath)
}
\arguments{
\item{filePath}{The location of the bed file}
}
\value{
a GRanges object
}
\description{
Tries to speed up rtracklayer import.bed
If speedup is not supported, it will use normal
rtracklayer::import.bed
Supports gzip, gz, bgz and bed formats
}
