% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_metadata.R
\name{get_bioproject_candidates}
\alias{get_bioproject_candidates}
\title{Query eutils for bioproject IDs}
\usage{
get_bioproject_candidates(term = "Ribosome Profiling human", RetMax = 10000)
}
\arguments{
\item{term}{character, default "Ribosome Profiling human"}

\item{RetMax}{integer, default 10000. How many IDs to return maximum}
}
\value{
character vector of IDs
}
\description{
The default query of Ribosome Profiling human, will result in internal
entrez search of:
Ribosome[All Fields] AND Profiling[All Fields] AND ("Homo sapiens"[Organism]
OR human[All Fields])
}
\examples{
term <- "Ribosome Profiling Saccharomyces cerevisiae"
# get_bioproject_candidates(term)
}
\references{
https://www.ncbi.nlm.nih.gov/books/NBK25501/
}
\seealso{
Other sra: 
\code{\link{download.SRA.metadata}()},
\code{\link{download.SRA}()},
\code{\link{download.ebi}()},
\code{\link{install.sratoolkit}()},
\code{\link{rename.SRA.files}()}
}
\concept{sra}
