% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_helper.R
\name{get_genome_fasta}
\alias{get_genome_fasta}
\title{Download genome (fasta), annotation (GTF) and contaminants}
\usage{
get_genome_fasta(
  genome,
  output.dir,
  organism,
  assembly,
  assembly_type,
  db,
  gunzip
)
}
\arguments{
\item{genome}{logical, default: TRUE. Download the genome for the organism
given in `organism`. If `FALSE`, ORFik checks whether the file is already
available locally. To use a custom genome from disk, set
\code{genome = FALSE} and then assign:\cr
\code{annotation <- getGenomeAndAnnotation(genome = FALSE)}\cr
\code{annotation["genome"] = "path/to/genome.fasta"}.\cr
For Ensembl, the primary assembly is downloaded.}

\item{output.dir}{directory to save downloaded data}

\item{organism}{scientific name of organism, Homo sapiens,
Danio rerio, Mus musculus, etc. See \code{biomartr:::get.ensembl.info()}
for full list of supported organisms.}

\item{assembly}{character, default `assembly = organism`, which selects the
first assembly returned for that organism. You can also supply a specific
assembly name, for example `"GCA_000005845"` for the commonly used
E. coli K-12 reference. This argument is usually only needed for bacterial
genomes.}

\item{assembly_type}{character, default c("primary_assembly", "toplevel"). Used for ensembl only,
specifies the genome assembly type. Searches for both primary and toplevel, and if both are found, uses the
first by order (so primary is prioritized by default).
The Primary assembly should usually be used if it exists.
The "primary assembly" contains all the top-level sequence regions,
excluding alternative haplotypes and patches.
If the primary assembly file is not present for a species
(only defined for standard model organisms),
that indicates that there were no haplotype/patch regions,
and in such cases, the 'toplevel file is used.
For more details see:
\href{https://grch37.ensembl.org/info/genome/genebuild/assembly.html}{ensembl tutorial}}

\item{db}{database to use for the genome and annotation. The recommended
choice is `"ensembl"`. When using Ensembl, remember to set
\code{assembly_type = "primary_assembly"} unless you want haplotypes
included, which can make the files much larger. Alternatives are `"refseq"`
for reference assemblies and `"genbank"` for all available assemblies.}

\item{gunzip}{logical, default TRUE, uncompress downloaded files
that are zipped when downloaded, should be TRUE!}
}
\value{
A named character vector with the paths to the downloaded genome,
annotation, and any contaminant references. If `merge_contaminants = TRUE`,
only the merged contaminant fasta is returned, not the individual files.
}
\description{
This function downloads the genome, annotation, and contaminant references
used for alignment if they are not already present. By default it uses the
Ensembl reference. When the run completes, it stores a file called
\code{file.path(output.dir, "outputs.rds")} with the output paths for the
downloaded files.
}
\details{
For most non-model non-vertebrate organisms, you need
my fork of biomartr for it to work:
remotes::install_github("Roleren/biomartr")
If a download fails or you used the wrong organism or assembly, delete the
incorrect files and run again.\cr
Set \code{remake = TRUE} to rebuild everything from scratch.\cr

Files created during the download process include:\cr
- A fasta index for the genome\cr
- A TxDb to speed up GTF/GFF reading\cr
- Separate or merged contaminant files\cr
Optional files that can also be created:\cr
- Gene symbols (hgnc, etc)\cr
- Uniprot ids (For name of protein structures)\cr
If you want to use a custom genome or GTF from disk, assign the existing
paths like this: \cr
annotation <- getGenomeAndAnnotation(GTF = "path/to/gtf.gtf",
genome = "path/to/genome.fasta")\cr
}
\examples{

## Get Saccharomyces cerevisiae genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel")
## Download and add pseudo 5' UTRs
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel",
#  pseudo_5UTRS_if_needed = 100)
## Get Danio rerio genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Danio rerio", tempdir())

output.dir <- "/Bio_data/references/zebrafish"
## Get Danio rerio and Phix contamints to deplete during alignment
#getGenomeAndAnnotation("Danio rerio", output.dir, phix = TRUE)

## Optimize for ORFik (speed up for large annotations like human or zebrafish)
#getGenomeAndAnnotation("Danio rerio", tempdir(), optimize = TRUE)

# Drosophila melanogaster (toplevel exists only)
#getGenomeAndAnnotation("drosophila melanogaster", output.dir = file.path(config["ref"],
# "Drosophila_melanogaster_BDGP6"), assembly_type = "toplevel")
## How to save malformed refseq gffs:
## First run function and let it crash:
#annotation <- getGenomeAndAnnotation(organism = "Arabidopsis thaliana",
#  output.dir = "~/Desktop/test_plant/",
#  assembly_type = "primary_assembly", db = "refseq")
## Then apply a fix (example for linux, too long rows):
# fixed_gff <- fix_malformed_gff("~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.gff")
## Then updated arguments:
# annotation <- c(fixed_gff, "~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.fna")
# names(annotation) <- c("gtf", "genome")
# Then make the txdb (for faster R use)
# makeTxdbFromGenome(annotation["gtf"], annotation["genome"], organism = "Arabidopsis thaliana")
}
\references{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4919035/
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.allsteps.multiQC}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.multiQC}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{install.fastp}()}
}
\keyword{internal}
