% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_helper.R
\name{get_genome_fasta}
\alias{get_genome_fasta}
\title{Download genome (fasta), annotation (GTF) and contaminants}
\usage{
get_genome_fasta(genome, output.dir, organism, assembly_type, db, gunzip)
}
\arguments{
\item{genome}{logical, default: TRUE, download genome of organism
specified in "organism" argument. If FALSE, check if the downloaded
file already exist. If you want to use a custom gtf from you hard drive,
set GTF = FALSE,
and assign: \cr annotation <- getGenomeAndAnnotation(genome = FALSE)\cr
annotation["genome"] = "path/to/genome.fasta".\cr
Will download the primary assembly for ensembl}

\item{output.dir}{directory to save downloaded data}

\item{organism}{scientific name of organism, Homo sapiens,
Danio rerio, Mus musculus, etc. See \code{biomartr:::get.ensembl.info()}
for full list of supported organisms.}

\item{assembly_type}{a character string specifying from which assembly type
the genome shall be retrieved from (ensembl only, else this argument is ignored):
Default is
\code{assembly_type = "primary_assembly")}.
This will give you all no copies of any chromosomes.
As an example, the  primary_assembly fasta genome in human is only a few GB uncompressed.\cr
\code{assembly_type = "toplevel")}.
This will give you all multi-chromosomes (copies of the same chromosome with small variations).
As an example the toplevel fasta genome in human is over 70 GB uncompressed.}

\item{db}{database to use for genome and GTF,
default adviced: "ensembl" (will contain haplotypes, large file!).
Alternatives: "refseq" (primary assembly) and "genbank" (mix)}

\item{gunzip}{logical, default TRUE, uncompress downloaded files
that are zipped when downloaded, should be TRUE!}
}
\value{
a named character vector of path to genomes and gtf downloaded,
 and additional contaminants if used. If merge_contaminants is TRUE, will not
 give individual fasta files to contaminants, but only the merged one.
}
\description{
This function automatically downloads (if files not already exists)
genomes and contaminants specified for genome alignment.
Will create a R transcript database (TxDb object) from the annotation. \cr
It will also index the genome for you\cr
If you misspelled something or crashed, delete wrong files and
run again.\cr
Do remake = TRUE, to do it all over again.
}
\details{
If you want custom genome or gtf from you hard drive, assign it
after you run this function, like this:\cr
annotation <- getGenomeAndAnnotation(GTF = FALSE, genome = FALSE)\cr
annotation["genome"] = "path/to/genome.fasta"\cr
annotation["gtf"] = "path/to/gtf.gtf"
}
\examples{

## Get Saccharomyces cerevisiae genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel")
## Get Danio rerio genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Danio rerio", tempdir())

output.dir <- "/Bio_data/references/zebrafish"
## Get Danio rerio and Phix contamints to deplete during alignment
#getGenomeAndAnnotation("Danio rerio", output.dir, phix = TRUE)

## Optimize for ORFik (speed up for large annotations like human or zebrafish)
#getGenomeAndAnnotation("Danio rerio", tempdir(), optimize = TRUE)

}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.allsteps.multiQC}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.multiQC}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{install.fastp}()}
}
