% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download_helper.R
\name{get_genome_gtf}
\alias{get_genome_gtf}
\title{Download genome (fasta), annotation (GTF) and contaminants}
\usage{
get_genome_gtf(
  GTF,
  output.dir,
  organism,
  assembly,
  db,
  gunzip,
  genome,
  optimize = FALSE,
  uniprot_id = FALSE,
  gene_symbols = FALSE,
  pseudo_5UTRS_if_needed = NULL,
  remove_annotation_outliers = TRUE,
  refseq_genbank_format = c("gtf", "gff3")[1]
)
}
\arguments{
\item{GTF}{logical, default: TRUE. Download the annotation file for the
organism given in `organism`. If `FALSE`, ORFik checks whether the file is
already available locally. To use a custom annotation from disk, set
`GTF = FALSE` and then assign:\cr
annotation <- getGenomeAndAnnotation(gtf = FALSE)\cr
annotation["gtf"] = "path/to/gtf.gtf".\cr
If `db` is not `"ensembl"`, the downloaded annotation will be GFF instead
of GTF.}

\item{output.dir}{directory to save downloaded data}

\item{organism}{scientific name of organism, Homo sapiens,
Danio rerio, Mus musculus, etc. See \code{biomartr:::get.ensembl.info()}
for full list of supported organisms.}

\item{assembly}{character, default `assembly = organism`, which selects the
first assembly returned for that organism. You can also supply a specific
assembly name, for example `"GCA_000005845"` for the commonly used
E. coli K-12 reference. This argument is usually only needed for bacterial
genomes.}

\item{db}{database to use for the genome and annotation. The recommended
choice is `"ensembl"`. When using Ensembl, remember to set
\code{assembly_type = "primary_assembly"} unless you want haplotypes
included, which can make the files much larger. Alternatives are `"refseq"`
for reference assemblies and `"genbank"` for all available assemblies.}

\item{gunzip}{logical, default TRUE, uncompress downloaded files
that are zipped when downloaded, should be TRUE!}

\item{genome}{character path, default NULL.
Path to fasta genome, corresponding to the gtf. must be indexed
(.fai file must exist there).
If you want to make sure chromosome naming of the GTF matches the genome
and correct seqlengths. If value is NULL or FALSE, it will be ignored.}

\item{optimize}{logical, default FALSE. Create a folder
within the output folder (defined by txdb_file_out_path),
that includes optimized objects
to speed up loading of annotation regions from up to 15 seconds
on human genome down to 0.1 second. ORFik will then load these optimized
objects instead. Currently optimizes filterTranscript() function and
loadRegion() function for 5' UTRs, 3' UTRs, CDS,
 mRNA (all transcript with CDS) and tx (all transcripts).}

\item{uniprot_id}{logical default FALSE.  If TRUE, will download
and store all uniprot id for all transcripts (coding and noncoding)-
In a file called: "gene_symbol_tx_table.fst" in same folder as txdb.}

\item{gene_symbols}{logical default FALSE. If TRUE, will download
and store all gene symbols for all transcripts (coding and noncoding)-
In a file called: "gene_symbol_tx_table.fst" in same folder as txdb.
hgcn for human, mouse symbols for mouse and rat, more to be added.}

\item{pseudo_5UTRS_if_needed}{integer, default NULL. If defined > 0,
will add pseudo 5' UTRs of maximum this length if 'minimum_5UTR_percentage" (default 30%) of
mRNAs (coding transcripts) do not have a leader. (NULL and 0 are both the ignore command)}

\item{remove_annotation_outliers}{logical, default TRUE. Only used for
RefSeq annotations. If `TRUE`, malformed outlier lines are removed from the
input annotation file. The cleaned annotation overwrites the original file,
and the removed lines are stored in `tempdir()` for inspection. This is
needed for some RefSeq annotations, including Arabidopsis.}

\item{refseq_genbank_format}{= c("gtf", "gff3")[1] Annotation format to
request from RefSeq or GenBank. GTF is usually the safer choice for
downstream compatibility. GFF3 may contain information that is not present
in the GTF, so use it when you specifically need that extra annotation.}
}
\value{
A named character vector with the paths to the downloaded genome,
annotation, and any contaminant references. If `merge_contaminants = TRUE`,
only the merged contaminant fasta is returned, not the individual files.
}
\description{
This function downloads the genome, annotation, and contaminant references
used for alignment if they are not already present. By default it uses the
Ensembl reference. When the run completes, it stores a file called
\code{file.path(output.dir, "outputs.rds")} with the output paths for the
downloaded files.
}
\details{
For most non-model non-vertebrate organisms, you need
my fork of biomartr for it to work:
remotes::install_github("Roleren/biomartr")
If a download fails or you used the wrong organism or assembly, delete the
incorrect files and run again.\cr
Set \code{remake = TRUE} to rebuild everything from scratch.\cr

Files created during the download process include:\cr
- A fasta index for the genome\cr
- A TxDb to speed up GTF/GFF reading\cr
- Separate or merged contaminant files\cr
Optional files that can also be created:\cr
- Gene symbols (hgnc, etc)\cr
- Uniprot ids (For name of protein structures)\cr
If you want to use a custom genome or GTF from disk, assign the existing
paths like this: \cr
annotation <- getGenomeAndAnnotation(GTF = "path/to/gtf.gtf",
genome = "path/to/genome.fasta")\cr
}
\examples{

## Get Saccharomyces cerevisiae genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel")
## Download and add pseudo 5' UTRs
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel",
#  pseudo_5UTRS_if_needed = 100)
## Get Danio rerio genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Danio rerio", tempdir())

output.dir <- "/Bio_data/references/zebrafish"
## Get Danio rerio and Phix contamints to deplete during alignment
#getGenomeAndAnnotation("Danio rerio", output.dir, phix = TRUE)

## Optimize for ORFik (speed up for large annotations like human or zebrafish)
#getGenomeAndAnnotation("Danio rerio", tempdir(), optimize = TRUE)

# Drosophila melanogaster (toplevel exists only)
#getGenomeAndAnnotation("drosophila melanogaster", output.dir = file.path(config["ref"],
# "Drosophila_melanogaster_BDGP6"), assembly_type = "toplevel")
## How to save malformed refseq gffs:
## First run function and let it crash:
#annotation <- getGenomeAndAnnotation(organism = "Arabidopsis thaliana",
#  output.dir = "~/Desktop/test_plant/",
#  assembly_type = "primary_assembly", db = "refseq")
## Then apply a fix (example for linux, too long rows):
# fixed_gff <- fix_malformed_gff("~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.gff")
## Then updated arguments:
# annotation <- c(fixed_gff, "~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.fna")
# names(annotation) <- c("gtf", "genome")
# Then make the txdb (for faster R use)
# makeTxdbFromGenome(annotation["gtf"], annotation["genome"], organism = "Arabidopsis thaliana")
}
\references{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4919035/
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.allsteps.multiQC}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.multiQC}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{install.fastp}()}
}
\keyword{internal}
