% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_system_usage}
\alias{get_system_usage}
\title{System usage for Linux (Auto-detects correct drive if not provided)}
\usage{
get_system_usage(drive = detect_drive(), one_liner = FALSE)
}
\arguments{
\item{drive}{path, the Filesystem drive !(Not the mounted name),
use drive = ORFik:::detect_drive("My_directory_inside this mount_name") to
get custom drive}

\item{one_liner}{Logical, default FALSE. Instead return a length 1 character string
with all the info.}
}
\value{
A list with system info, if one_liner is TRUE, then a length 1 character string.
}
\description{
System usage for Linux (Auto-detects correct drive if not provided)
}
\examples{
get_system_usage()
}
