% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGAlignmentsPairs}
\alias{getGAlignmentsPairs}
\title{Internal GAlignmentPairs loader from fst data.frame}
\usage{
getGAlignmentsPairs(df)
}
\arguments{
\item{df}{a data.frame with columns minimum 6 columns:
seqnames, start1/start2 (integers), cigar1/cigar2 and strand\cr
Additional columns will be assigned as meta columns}
}
\value{
GAlignmentPairs object
}
\description{
Internal GAlignmentPairs loader from fst data.frame
}
