% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_download.R
\name{getGenomeAndAnnotation}
\alias{getGenomeAndAnnotation}
\title{Download genome (fasta), annotation (GTF) and contaminants}
\usage{
getGenomeAndAnnotation(
  organism,
  output.dir,
  db = "ensembl",
  GTF = TRUE,
  genome = TRUE,
  merge_contaminants = TRUE,
  phix = FALSE,
  ncRNA = FALSE,
  tRNA = FALSE,
  rRNA = FALSE,
  gunzip = TRUE,
  remake = FALSE,
  assembly_type = c("primary_assembly", "toplevel"),
  optimize = FALSE,
  gene_symbols = FALSE,
  uniprot_id = FALSE,
  pseudo_5UTRS_if_needed = NULL,
  remove_annotation_outliers = TRUE,
  notify_load_existing = TRUE,
  assembly = organism,
  refseq_genbank_format = c("gtf", "gff3")[1]
)
}
\arguments{
\item{organism}{scientific name of organism, Homo sapiens,
Danio rerio, Mus musculus, etc. See \code{biomartr:::get.ensembl.info()}
for full list of supported organisms.}

\item{output.dir}{directory to save downloaded data}

\item{db}{database to use for the genome and annotation. The recommended
choice is `"ensembl"`. When using Ensembl, remember to set
\code{assembly_type = "primary_assembly"} unless you want haplotypes
included, which can make the files much larger. Alternatives are `"refseq"`
for reference assemblies and `"genbank"` for all available assemblies.}

\item{GTF}{logical, default: TRUE. Download the annotation file for the
organism given in `organism`. If `FALSE`, ORFik checks whether the file is
already available locally. To use a custom annotation from disk, set
`GTF = FALSE` and then assign:\cr
annotation <- getGenomeAndAnnotation(gtf = FALSE)\cr
annotation["gtf"] = "path/to/gtf.gtf".\cr
If `db` is not `"ensembl"`, the downloaded annotation will be GFF instead
of GTF.}

\item{genome}{logical, default: TRUE. Download the genome for the organism
given in `organism`. If `FALSE`, ORFik checks whether the file is already
available locally. To use a custom genome from disk, set
\code{genome = FALSE} and then assign:\cr
\code{annotation <- getGenomeAndAnnotation(genome = FALSE)}\cr
\code{annotation["genome"] = "path/to/genome.fasta"}.\cr
For Ensembl, the primary assembly is downloaded.}

\item{merge_contaminants}{logical, default TRUE. Merge the requested
contaminant references into one fasta file. This usually saves space and is
faster to align with STAR than keeping each contaminant as a separate
reference. Ignored if no contaminants are requested.}

\item{phix}{logical, default FALSE, download phiX sequence to filter
 out Illumina control reads. ORFik defines Phix as a contaminant genome.
Phix is used in Illumina sequencers for sequencing quality control.
Genome is: refseq, Escherichia phage phiX174.
If sequencing facility created fastq files with the command \code{bcl2fastq},
then there should be very few phix reads left in the fastq files received.}

\item{ncRNA}{logical or character, default FALSE. If `TRUE` or a path is
supplied, ncRNA is used as a contaminant reference. If `TRUE`, ORFik tries
to find ncRNA sequences from the GTF file, usually represented as lncRNA
entries, and reports if none are found.\cr If that does not work, try a
character input.\cr A value of `"auto"` tells ORFik to look for an ncRNA
file for the organism on NONCODE, using the organism's common name
(for example, Homo sapiens -> human). `"auto"` does not work for every
species, so you may need to provide the common name used by NONCODE
directly. If the value is not `"auto"` or `""`, it must be a character
vector of common species names, not scientific names. See
http://www.noncode.org/download.php/ if you need to look up the correct
NONCODE name.}

\item{tRNA}{logical or character, default FALSE. If enabled, tRNA is used
as a contaminant reference. If `TRUE`, ORFik tries to extract tRNA
sequences from the GTF file, usually represented as `Mt_tRNA` entries, and
reports if none are found. If that is not sufficient, provide a character
vector with a valid path to a fasta file containing mature tRNAs on disk.
You can obtain these from http://gtrnadb.ucsc.edu/ or by running tRNAscan
on the genome.}

\item{rRNA}{logical or character, default FALSE. If enabled, rRNA is used
as a contaminant reference. If `TRUE`, ORFik tries to extract rRNA
sequences from the GTF file and reports if none are found. If set to
`"silva"`, ORFik downloads the Silva SSU and LSU reference files for all
species. If you need a smaller reference, download one manually from
https://www.arb-silva.de/.\cr
If the value is neither `""` nor `"silva"`, it must be a character vector
giving a valid path to an rRNA fasta file on disk.}

\item{gunzip}{logical, default TRUE, uncompress downloaded files
that are zipped when downloaded, should be TRUE!}

\item{remake}{logical, default: FALSE, if TRUE remake everything specified}

\item{assembly_type}{character, default c("primary_assembly", "toplevel"). Used for ensembl only,
specifies the genome assembly type. Searches for both primary and toplevel, and if both are found, uses the
first by order (so primary is prioritized by default).
The Primary assembly should usually be used if it exists.
The "primary assembly" contains all the top-level sequence regions,
excluding alternative haplotypes and patches.
If the primary assembly file is not present for a species
(only defined for standard model organisms),
that indicates that there were no haplotype/patch regions,
and in such cases, the 'toplevel file is used.
For more details see:
\href{https://grch37.ensembl.org/info/genome/genebuild/assembly.html}{ensembl tutorial}}

\item{optimize}{logical, default FALSE. Create a folder
within the output folder (defined by txdb_file_out_path),
that includes optimized objects
to speed up loading of annotation regions from up to 15 seconds
on human genome down to 0.1 second. ORFik will then load these optimized
objects instead. Currently optimizes filterTranscript() function and
loadRegion() function for 5' UTRs, 3' UTRs, CDS,
 mRNA (all transcript with CDS) and tx (all transcripts).}

\item{gene_symbols}{logical default FALSE. If TRUE, will download
and store all gene symbols for all transcripts (coding and noncoding)-
In a file called: "gene_symbol_tx_table.fst" in same folder as txdb.
hgcn for human, mouse symbols for mouse and rat, more to be added.}

\item{uniprot_id}{logical default FALSE.  If TRUE, will download
and store all uniprot id for all transcripts (coding and noncoding)-
In a file called: "gene_symbol_tx_table.fst" in same folder as txdb.}

\item{pseudo_5UTRS_if_needed}{integer, default NULL. If defined > 0,
will add pseudo 5' UTRs of maximum this length if 'minimum_5UTR_percentage" (default 30%) of
mRNAs (coding transcripts) do not have a leader. (NULL and 0 are both the ignore command)}

\item{remove_annotation_outliers}{logical, default TRUE. Only used for
RefSeq annotations. If `TRUE`, malformed outlier lines are removed from the
input annotation file. The cleaned annotation overwrites the original file,
and the removed lines are stored in `tempdir()` for inspection. This is
needed for some RefSeq annotations, including Arabidopsis.}

\item{notify_load_existing}{logical, default TRUE. If a previous download is
already present in `output.dir` and recorded in `outputs.rds`, print a short
message instead of silently reusing it. Set to `FALSE` to suppress this.}

\item{assembly}{character, default `assembly = organism`, which selects the
first assembly returned for that organism. You can also supply a specific
assembly name, for example `"GCA_000005845"` for the commonly used
E. coli K-12 reference. This argument is usually only needed for bacterial
genomes.}

\item{refseq_genbank_format}{= c("gtf", "gff3")[1] Annotation format to
request from RefSeq or GenBank. GTF is usually the safer choice for
downstream compatibility. GFF3 may contain information that is not present
in the GTF, so use it when you specifically need that extra annotation.}
}
\value{
A named character vector with the paths to the downloaded genome,
annotation, and any contaminant references. If `merge_contaminants = TRUE`,
only the merged contaminant fasta is returned, not the individual files.
}
\description{
This function downloads the genome, annotation, and contaminant references
used for alignment if they are not already present. By default it uses the
Ensembl reference. When the run completes, it stores a file called
\code{file.path(output.dir, "outputs.rds")} with the output paths for the
downloaded files.
}
\details{
For most non-model non-vertebrate organisms, you need
my fork of biomartr for it to work:
remotes::install_github("Roleren/biomartr")
If a download fails or you used the wrong organism or assembly, delete the
incorrect files and run again.\cr
Set \code{remake = TRUE} to rebuild everything from scratch.\cr

Files created during the download process include:\cr
- A fasta index for the genome\cr
- A TxDb to speed up GTF/GFF reading\cr
- Separate or merged contaminant files\cr
Optional files that can also be created:\cr
- Gene symbols (hgnc, etc)\cr
- Uniprot ids (For name of protein structures)\cr
If you want to use a custom genome or GTF from disk, assign the existing
paths like this: \cr
annotation <- getGenomeAndAnnotation(GTF = "path/to/gtf.gtf",
genome = "path/to/genome.fasta")\cr
}
\examples{

## Get Saccharomyces cerevisiae genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel")
## Download and add pseudo 5' UTRs
#getGenomeAndAnnotation("Saccharomyces cerevisiae", tempdir(), assembly_type = "toplevel",
#  pseudo_5UTRS_if_needed = 100)
## Get Danio rerio genome and gtf (create txdb for R)
#getGenomeAndAnnotation("Danio rerio", tempdir())

output.dir <- "/Bio_data/references/zebrafish"
## Get Danio rerio and Phix contamints to deplete during alignment
#getGenomeAndAnnotation("Danio rerio", output.dir, phix = TRUE)

## Optimize for ORFik (speed up for large annotations like human or zebrafish)
#getGenomeAndAnnotation("Danio rerio", tempdir(), optimize = TRUE)

# Drosophila melanogaster (toplevel exists only)
#getGenomeAndAnnotation("drosophila melanogaster", output.dir = file.path(config["ref"],
# "Drosophila_melanogaster_BDGP6"), assembly_type = "toplevel")
## How to save malformed refseq gffs:
## First run function and let it crash:
#annotation <- getGenomeAndAnnotation(organism = "Arabidopsis thaliana",
#  output.dir = "~/Desktop/test_plant/",
#  assembly_type = "primary_assembly", db = "refseq")
## Then apply a fix (example for linux, too long rows):
# fixed_gff <- fix_malformed_gff("~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.gff")
## Then updated arguments:
# annotation <- c(fixed_gff, "~/Desktop/test_plant/Arabidopsis_thaliana_genomic_refseq.fna")
# names(annotation) <- c("gtf", "genome")
# Then make the txdb (for faster R use)
# makeTxdbFromGenome(annotation["gtf"], annotation["genome"], organism = "Arabidopsis thaliana")
}
\references{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4919035/
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.allsteps.multiQC}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.multiQC}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{install.fastp}()}
}
\concept{STAR}
