% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{getNGenesCoverage}
\alias{getNGenesCoverage}
\title{Get number of genes per coverage table}
\usage{
getNGenesCoverage(coverage)
}
\arguments{
\item{coverage}{a data.table with coverage}
}
\value{
number of genes in coverage
}
\description{
Used to count genes in ORFik meta plots
}
\keyword{internal}
