% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_helpers.R
\name{getWeights}
\alias{getWeights}
\title{Get weights from a subject GenomicRanges object}
\usage{
getWeights(subject, weight = 1L)
}
\arguments{
\item{subject}{a GRanges, IRanges or GAlignment object}

\item{weight}{a vector (default: 1L, if 1L it is identical to
countOverlaps()),
if single number (!= 1), it applies for all,
if more than one must be equal size of 'reads'.
else it must be the string name of a defined meta column in subject
"reads", that gives number of times a read was found.
GRanges("chr1", 1, "+", score = 5), would mean "score" column tells
that this alignment region was found 5 times.}
}
\value{
a numeric vector of weights of equal size to subject
}
\description{
Get weights from a subject GenomicRanges object
}
