% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_helpers.R
\name{groupGRangesBy}
\alias{groupGRangesBy}
\title{make GRangesList from GRanges,}
\usage{
groupGRangesBy(gr, other = NULL)
}
\arguments{
\item{gr}{a GRanges object}

\item{other}{a vector of names to group, no 2 groups can have same name}
}
\value{
a GRangesList named after names(Granges) if other is NULL, else
names are from unique(other)
}
\description{
Grouped by names or another column(other)
 ig. if GRanges should be grouped by gene,
 give gene column as other
}
\examples{
ORFranges <- GRanges(seqnames = Rle(rep("1", 3)),
           ranges = IRanges(start = c(1, 10, 20),
           end = c(5, 15, 25)),
           strand = "+")

ORFranges2 <- GRanges("1",
                     ranges = IRanges(start = c(20, 30, 40),
                                      end = c(25, 35, 45)),
                     strand = "+")
names(ORFranges) = rep("tx1_1",3)
names(ORFranges2) = rep("tx1_2",3)
grl <- GRangesList(tx1_1 = ORFranges, tx1_2 = ORFranges2)
gr <- unlist(grl, use.names = FALSE)
## now recreate the grl
## group by orf
grltest <- groupGRangesBy(gr) # using the names to group
identical(grl, grltest) ## they are identical

## group by transcript
names(gr) <- OrfToTxNames(gr)
grltest <- groupGRangesBy(gr)
identical(grl, grltest) ## they are not identical

}
