% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{gSort}
\alias{gSort}
\title{Sort a GRangesList, helper.}
\usage{
gSort(grl, decreasing = FALSE, byStarts = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}}}

\item{decreasing}{should the first in each group have max(start(group))
->T or min-> default(F) ?}

\item{byStarts}{a logical T, should it order by starts or ends F.}
}
\value{
an equally named GRangesList, where each group is sorted within
group.
}
\description{
A helper for \code{\link[=sortPerGroup]{sortPerGroup()}}.
A faster, more versatile reimplementation of GenomicRanges::sort()
Normally not used directly.
Groups first each group, then either decreasing or increasing
(on starts if byStarts == T, on ends if byStarts == F)
}
