% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{hasHits}
\alias{hasHits}
\title{Hits from reads}
\usage{
hasHits(grl, reads, keep.names = FALSE, overlaps = NULL)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} or GRanges object}

\item{reads}{a GRanges, GAlignment or GAlignmentPairs object}

\item{keep.names}{logical (F), keep names or not}

\item{overlaps}{default NULL, if not null must be
countOverlaps(grl, reads), input if you have it already.}
}
\value{
a list of logicals, T == hit, F == no hit
}
\description{
Finding GRanges groups that have overlap hits with reads
Similar to %over%.
}
