% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_heatmaps.R
\name{heatMapRegion}
\alias{heatMapRegion}
\title{Create heatmap around specified position}
\usage{
heatMapRegion(
  df,
  region = "TIS",
  outdir = "default",
  scores = c("transcriptNormalized", "sum"),
  type = "ofst",
  cage = NULL,
  format = ".png",
  acceptedLengths = 21:75,
  upstream = c(50, 30),
  downstream = c(29, 69),
  shifting = c("5prime", "3prime")
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{region}{a character, default "TIS", can be any combination of the
set: c("TSS", "TIS", "TTS"), which are: Transcription start site
(5' end of mrna), Translation initation site (5' end of CDS),
Translation termination site (3' end of CDS)}

\item{outdir}{a character path, default: "default", saves to:
\code{paste0(dirname(df$filepath[1]), "/QC_STATS/heatmaps/")}, a created
folder within the ORFik experiment data folder for plots. Change if you
want custom location.}

\item{scores}{character vector, default c("transcriptNormalized", "sum"),
either of zscore, transcriptNormalized, sum, mean, median, ..
see ?coverageScorings for info and more alternatives.}

\item{type}{character, default: "ofst". Type of library:
either "default", usually bam format (the one you gave to experiment),
"pshifted" pshifted reads, "ofst", "bed", "bedo" optimized bed, or "wig"}

\item{cage}{a character path to library file or a \code{\link{GRanges}},
\code{\link{GAlignments}} preloaded file of CAGE data. Only used if
"TSS" is defined as region, to redefine 5' leaders.}

\item{format}{a character, default ".png", alternative ".pdf"}

\item{acceptedLengths}{an integer vector (NULL), the read lengths accepted.
Default NULL, means all lengths accepted.}

\item{upstream}{2 integers, default c(50, 30), how long upstream from 0
should window extend (first index is 5' end extension, second is 3' end extension)}

\item{downstream}{2 integers, default c(29, 69), how long upstream from 0
should window extend (first index is 5' end extension, second is 3' end extension)}

\item{shifting}{a character, default c("5prime", "3prime"), can also be
either or NULL (no shifting of reads)}
}
\value{
invisible(NULL), plots are saved
}
\description{
Simplified input space for easier abstraction of coverage heatmaps\cr
Pick your region and plot \cr
Input CAGE file if you use TSS and want improved 5' annotation.
}
\examples{

df <- ORFik.template.experiment()[3,] # Only third library
#heatMapRegion(df, "TIS", outdir = "PATH/TO/SAVE/)
# Do also TSS, add cage for specific TSS
# heatMapRegion(df, c("TSS", "TIS"), cage = "path/to/cage.bed")

# Do on pshifted reads instead of original files
# outputLibs(df, type = "pshifted")
# heatMapRegion(df, "TIS")
}
\seealso{
Other heatmaps: 
\code{\link{coverageHeatMap}()},
\code{\link{heatMapL}()},
\code{\link{heatMap_single}()}
}
\concept{heatmaps}
