% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{inFrameWithCDS}
\alias{inFrameWithCDS}
\title{find frame for each orf relative to cds}
\usage{
inFrameWithCDS(dists)
}
\arguments{
\item{dists}{a vector of distances between ORF and cds}
}
\description{
Input of this function, is the output of the function distOrfToCds
See article:  10.1074/jbc.R116.733899
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
