% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{insideOutsideORF}
\alias{insideOutsideORF}
\title{Inside/outside score (IO)}
\usage{
insideOutsideORF(grl, RFP, GtfOrTx)
}
\arguments{
\item{grl}{a GRangesList object with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{GtfOrTx}{if Gtf: a TxDb object of a gtf file,
if tx: a GrangesList of transcripts, called from:
exonsBy(Gtf, by = "tx", use.names = T)}
}
\value{
a named vector of numeric values of scores
}
\description{
is defined as (RPFs over ORF)/(RPFs downstream to tx end).
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1242/dev.098345
}
