% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{install.fastp}
\alias{install.fastp}
\title{Download and prepare fastp trimmer}
\usage{
install.fastp(folder = "~/bin")
}
\arguments{
\item{folder}{path to folder for download, fille will be named
"fastp", this should be most recent version}
}
\value{
path to runnable fastp
}
\description{
Will not run "make", only use precompiled fastp file.\cr
Only works for Linux
}
\examples{
#install.fastp()
}
\references{
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6129281/
}
