% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{install.sratoolkit}
\alias{install.sratoolkit}
\title{Download sra toolkit}
\usage{
install.sratoolkit(folder = "~/bin", version = "2.10.8")
}
\arguments{
\item{folder}{default folder, "~/bin"}

\item{version}{a string, default "2.10.8"}
}
\value{
path to fastq-dump in sratoolkit
}
\description{
Download sra toolkit
}
\references{
https://ncbi.github.io/sra-tools/fastq-dump.html
}
