% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRA_helper.R
\name{install.sratoolkit}
\alias{install.sratoolkit}
\title{Download sra toolkit}
\usage{
install.sratoolkit(folder = "~/bin", version = "2.10.9")
}
\arguments{
\item{folder}{default folder, "~/bin"}

\item{version}{a string, default "2.10.9"}
}
\value{
path to fastq-dump in sratoolkit
}
\description{
Currently supported for Linux (64 bit centos and ubunutu is tested to work)
and Mac-OS(64 bit)
}
\examples{
# install.sratoolkit()
## Custom folder and version
folder <- "/I/WANT/IT/HERE/"
# install.sratoolkit(folder, version = "2.10.7")

}
\references{
https://ncbi.github.io/sra-tools/fastq-dump.html
}
\seealso{
Other sra: 
\code{\link{download.SRA.metadata}()},
\code{\link{download.SRA}()},
\code{\link{download.ebi}()},
\code{\link{get_bioproject_candidates}()},
\code{\link{rename.SRA.files}()}
}
\concept{sra}
