% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.grl}
\alias{is.grl}
\title{Helper function to check for GRangesList}
\usage{
is.grl(class)
}
\arguments{
\item{class}{the class you want to check if is GRL,
either a character from class or the object itself.}
}
\value{
a boolean
}
\description{
Helper function to check for GRangesList
}
