% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{is.ORF}
\alias{is.ORF}
\title{Check if all requirements for an ORFik ORF is accepted.}
\usage{
is.ORF(grl)
}
\arguments{
\item{grl}{a GRangesList or GRanges to check}
}
\value{
a logical (TRUE/FALSE)
}
\description{
Check if all requirements for an ORFik ORF is accepted.
}
\seealso{
Other validity: 
\code{\link{checkRFP}()},
\code{\link{checkRNA}()},
\code{\link{is.gr_or_grl}()},
\code{\link{is.grl}()},
\code{\link{is.range}()},
\code{\link{validGRL}()},
\code{\link{validSeqlevels}()}
}
\concept{validity}
