% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{is.range}
\alias{is.range}
\title{Helper function to check for ranged object}
\usage{
is.range(class)
}
\arguments{
\item{class}{the class you want to check if is GRL or GR,
either a character from class or the object itself.}
}
\value{
a boolean
}
\description{
Helper function to check for ranged object
}
\seealso{
Other validity: 
\code{\link{checkRFP}()},
\code{\link{checkRNA}()},
\code{\link{is.ORF}()},
\code{\link{is.gr_or_grl}()},
\code{\link{is.grl}()},
\code{\link{validGRL}()},
\code{\link{validSeqlevels}()}
}
\concept{validity}
\keyword{internal}
