% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods2.R
\name{isOverlapping}
\alias{isOverlapping}
\title{Find frame for each orf relative to cds}
\usage{
isOverlapping(dists)
}
\arguments{
\item{dists}{a vector of distances between ORF and cds}
}
\value{
a logical vector
}
\description{
Input of this function, is the output of the function
[distToCds()]
}
\examples{
# simple example
isOverlapping(c(-3,-6,8,11,15))

# GRangesList example
grl <- GRangesList(tx1_1 = GRanges("1", IRanges(1,10), "+"))
fiveUTRs <- GRangesList(tx1 = GRanges("1", IRanges(1,20), "+"))
dist <- distToCds(grl, fiveUTRs, extension = 0)
isOverlapping <- isOverlapping(dist)
}
\references{
doi: 10.1074/jbc.R116.733899
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{distToTSS}},
  \code{\link{entropy}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{initiationScore}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
\concept{features}
