% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{isPeriodic}
\alias{isPeriodic}
\title{Find if there is a periodicity of 3 in the vector}
\usage{
isPeriodic(x)
}
\arguments{
\item{x}{(numeric) Vector of values to detect periodicity of 3 like in
RiboSeq data.}
}
\value{
a logical, if it is periodic.
}
\description{
It uses Fourier transform for finding periodic vectors on the transcript
normalized counts over all CDS TIS regions from -30 to 29, where TIS is
0.\cr
Checks if there is a periodicity and if the periodicity is 3,
more precisely between 2.9 and 3.1.
}
\details{
Transcript normalized means per CDS TIS region, count reads per position,
divide that number per position by the total of that transcript, then sum
up these numbers per position for all transcripts.
}
