% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_footprints_helpers.R
\name{isPeriodic}
\alias{isPeriodic}
\title{Find if there is periodicity in the vector}
\usage{
isPeriodic(x)
}
\arguments{
\item{x}{(numeric) Vector of values to detect periodicity of 3 like in
RiboSeq data.}
}
\value{
a logical, if it is periodic.
}
\description{
Checks if there is a periodicity and if the periodicity is 3,
more precisely between 2.9 and 3.1.
}
\details{
It uses Fourier transform for finding periodic vectors
}
