% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{kozakSequenceScore}
\alias{kozakSequenceScore}
\title{Make a score for each ORFs start region}
\usage{
kozakSequenceScore(grl, faFile, species = "human")
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} grouped by ORF}

\item{faFile}{a FaFile from the fasta file, see ?FaFile}

\item{species}{which species to use, currently only support human}
}
\value{
an integer vector, one score per orf
}
\description{
The closer the sequence is to the kozak sequence
The higher the score, based on simplification of PWM
score system: 4 upstream, 5 downstream of start
CACCATGGC, 1+3+1+2, skip ATG, +2+1 = 10
CGCCATGGC, 1+!2+1+2, skip ATG, +2+1 = 9
Inspired by experimental bit values for each position
}
\seealso{
Other features: \code{\link{ORFScores}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{allFeatures}},
  \code{\link{calculateCoverage}},
  \code{\link{disengagementScore}},
  \code{\link{distOrfToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkmRFP}},
  \code{\link{fpkmRNA}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{inFrameWithCDS}},
  \code{\link{insideOutsideORF}},
  \code{\link{isOverlappingCds}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
