% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{lastExonPerGroup}
\alias{lastExonPerGroup}
\title{get first exon per GRangesList group
grl must be sorted, call ORFik:::sortPerGroup if needed}
\usage{
lastExonPerGroup(grl)
}
\arguments{
\item{grl}{a GRangesList,}
}
\value{
a GRangesList of last exons per group
}
\description{
get first exon per GRangesList group
grl must be sorted, call ORFik:::sortPerGroup if needed
}
