% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{lastExonPerGroup}
\alias{lastExonPerGroup}
\title{get last exon per granges group}
\usage{
lastExonPerGroup(grl)
}
\arguments{
\item{grl}{a GRangesList}
}
\description{
get last exon per granges group
}
