% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{lastExonPerGroup}
\alias{lastExonPerGroup}
\title{get last exon per GRangesList group
grl must be sorted, call ORFik:::sortPerGroup if needed}
\usage{
lastExonPerGroup(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}
}
\value{
a GRangesList of the last exon per group
}
\description{
get last exon per GRangesList group
grl must be sorted, call ORFik:::sortPerGroup if needed
}
