% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{lastExonStartPerGroup}
\alias{lastExonStartPerGroup}
\title{get last start per granges group}
\usage{
lastExonStartPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a GRangesList}

\item{keep.names}{a boolean, keep names or not}
}
\description{
get last start per granges group
}
