% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{libFolder,experiment-method}
\alias{libFolder,experiment-method}
\title{Get path to ORFik experiment library folder}
\usage{
\S4method{libFolder}{experiment}(x, mode = "first", unique_mappers = uniqueMappers(x))
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}

\item{mode}{character, default "first". Alternatives: "unique", "all". Unique
means the unique directories, not to be confused with unique_mappers argument below.}

\item{unique_mappers}{logical, default uniqueMappers(x) If true appends unique_mappers to path}
}
\value{
a character path
}
\description{
Get path to ORFik experiment library folder
}
