% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{libFolder,experiment-method}
\alias{libFolder,experiment-method}
\title{Get ORFik experiment library folder}
\usage{
\S4method{libFolder}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get ORFik experiment library folder
}
