% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{libFolder}
\alias{libFolder}
\title{Get ORFik experiment library folder}
\usage{
libFolder(x, mode = "first")
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}

\item{mode}{character, default "first". Alternatives: "unique", "all".}
}
\value{
a character path
}
\description{
Get ORFik experiment library folder
}
