% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{libraryTypes}
\alias{libraryTypes}
\title{Which type of library type in \code{\link{experiment}}?}
\usage{
libraryTypes(df)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}
}
\value{
library types (character vector)
}
\description{
Which type of library type in \code{\link{experiment}}?
}
\seealso{
Other ORFik_experiment: 
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{organism.df}()},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
