% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadRegion}
\alias{loadRegion}
\title{Load transcript region}
\usage{
loadRegion(txdb, part = "tx")
}
\arguments{
\item{txdb}{a TxDb file or a path to one of:
(.gtf ,.gff, .gff2, .gff2, .db or .sqlite), if it is a GRangesList,
it will return it self.}

\item{part}{a character, one of: tx, leader, cds, trailer, introns}
}
\value{
a GrangesList of region
}
\description{
Load GRangesList if input is not already GRangesList.
}
