% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadRegion}
\alias{loadRegion}
\title{Load transcript region}
\usage{
loadRegion(txdb, part = "tx", names.keep = NULL, by = "tx")
}
\arguments{
\item{txdb}{a TxDb file or a path to one of:
(.gtf ,.gff, .gff2, .gff2, .db or .sqlite), if it is a GRangesList,
it will return it self.}

\item{part}{a character, one of: tx, leader, cds, trailer, intron, mrna
NOTE: difference between tx and mrna is that tx are all transcripts, while
mrna are all transcripts with a cds}

\item{names.keep}{a character vector of subset of names to keep. Example:
loadRegions(txdb, names = "ENST1000005"), will return only that transcript.
Remember if you set by to "gene", then this list must be with gene names.}

\item{by}{a character, default "tx" Either "tx" or "gene". What names to
output region by, the transcript name "tx" or gene names "gene"}
}
\value{
a GrangesList of region
}
\description{
Usefull to simplify loading of standard regions, like cds' and leaders.
Adds another safety in that seqlevels will be set
}
\details{
Load as GRangesList if input is not already GRangesList.
}
\examples{
gtf <- system.file("extdata", "annotations.gtf", package = "ORFik")
loadRegion(gtf, "cds")
loadRegion(gtf, "intron")
}
