% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadRegions}
\alias{loadRegions}
\title{Get all regions of transcripts specified to environment}
\usage{
loadRegions(
  txdb,
  parts = c("mrna", "leaders", "cds", "trailers"),
  extension = "",
  names.keep = NULL,
  envir = .GlobalEnv
)
}
\arguments{
\item{txdb}{a TxDb file, a path to one of:
(.gtf ,.gff, .gff2, .gff2, .db or .sqlite)
or an ORFik experiment}

\item{parts}{the transcript parts you want}

\item{extension}{What to add on the name after leader, like: B -> leadersB}

\item{names.keep}{a character vector of subset of names to keep. Example:
loadRegions(txdb, names = ENST1000005), will return only that transcript}

\item{envir}{Which environment to save to, default (.GlobalEnv)}
}
\value{
NULL (regions set by envir assignment)
}
\description{
By default loads all parts to .GlobalEnv (global environemnt)
Useful to not spend time on finding the functions to load regions.
}
\examples{
# Load all mrna regions to Global environment
gtf <- system.file("extdata", "annotations.gtf", package = "ORFik")
loadRegions(gtf, parts = c("mrna", "leaders", "cds", "trailers"))
}
