% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadRegions}
\alias{loadRegions}
\title{Get all regions of transcripts specified}
\usage{
loadRegions(txdb, parts = c("mrna", "leaders", "cds", "trailers"),
  extension = "", envir = .GlobalEnv)
}
\arguments{
\item{txdb}{a TxDb file, an ORFik experiment or
a path to one of:
 (.gtf ,.gff, .gff2, .gff2, .db or .sqlite)}

\item{parts}{the transcript parts you want}

\item{extension}{What to add on the name after leader, like: B -> leadersB}

\item{envir}{Which environment to save to, default (.GlobalEnv)}
}
\value{
NULL (regions set by envir assignment)
}
\description{
Get all regions of transcripts specified
}
