% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{loadTxdb}
\alias{loadTxdb}
\title{General loader for txdb}
\usage{
loadTxdb(txdb, chrStyle = NULL)
}
\arguments{
\item{txdb}{a TxDb file, an ORFik experiment or
a path to one of:
 (.gtf ,.gff, .gff2, .gff2, .db or .sqlite)}

\item{chrStyle}{a GRanges object, or a character style (Default: NULL) to
get seqlevelsStyle from. Is chromosome 1 called chr1 or 1, is mitocondrial
chromosome called MT or chrM etc. Will use 1st seqlevel-
style if more are present. Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a TxDb object
}
\description{
Useful to allow fast TxDb loader like .db
}
\examples{
library(GenomicFeatures)
# Get the gtf txdb file
txdbFile <- system.file("extdata", "hg19_knownGene_sample.sqlite",
                        package = "GenomicFeatures")
txdb <- loadDb(txdbFile)

}
