% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{makeExonRanks}
\alias{makeExonRanks}
\title{make a meta column with exon ranks}
\usage{
makeExonRanks(grl, byTranscript = F)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{byTranscript}{if ORfs are by transcript, check duplicates}
}
\description{
make a meta column with exon ranks
}
