% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_helpers.R
\name{makeExonRanks}
\alias{makeExonRanks}
\title{Make a meta column with exon ranks}
\usage{
makeExonRanks(grl, byTranscript = F)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{byTranscript}{if ORfs are by transcript, check duplicates}
}
\value{
an integer vector of indices for exon ranks
}
\description{
Must be ordered, so that same transcripts are ordered together.
}
