% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{makeORFNames}
\alias{makeORFNames}
\title{Make ORF names per orf}
\usage{
makeORFNames(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}
}
\value{
a GRangesList now with ORF names, grouped by transcripts.
 They are also sorted
}
\description{
grl must be grouped by transcript
If a list of orfs are grouped by transcripts, but does not have
ORF names, then create them and return the new GRangesList
}
