% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{makeSummarizedExperimentFromBam}
\alias{makeSummarizedExperimentFromBam}
\title{Make a summerizedExperiment object from bam files}
\usage{
makeSummarizedExperimentFromBam(
  df,
  saveName = NULL,
  longestPerGene = TRUE,
  geneOrTxNames = "gene",
  region = "mrna",
  type = "count"
)
}
\arguments{
\item{df}{an ORFik experiment, to make it, see: ?experiment}

\item{saveName}{a character (default NULL),
if set save experiment to path given. Always saved as .rds.,
it is optional to add .rds, it will be added for you if not present.}

\item{longestPerGene}{a logical (default TRUE), if FALSE all transcript
isoforms per gene.}

\item{geneOrTxNames}{a character (default gene), if tx use with tx names}

\item{region}{(default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{type}{default: "count" (raw counts matrix), alternative is "fpkm",
"log2fpkm" or "log10fpkm"}
}
\value{
a summerizedExperiment object
}
\description{
If txdb or gtf path is added, it is a rangedSummerizedExperiment
}
