% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{makeSummarizedExperimentFromBam}
\alias{makeSummarizedExperimentFromBam}
\title{Make a count matrix from a library or experiment}
\usage{
makeSummarizedExperimentFromBam(
  df,
  saveName = NULL,
  longestPerGene = FALSE,
  geneOrTxNames = "tx",
  region = "mrna",
  type = "count",
  lib.type = "ofst",
  weight = "score"
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{saveName}{a character (default NULL),
if set save experiment to path given. Always saved as .rds.,
it is optional to add .rds, it will be added for you if not present.
Also used to load existing file with that name.}

\item{longestPerGene}{a logical (default FALSE), if FALSE all transcript
isoforms per gene. Ignored if "region" is not a character of either:
"mRNA","tx", "cds", "leaders" or "trailers".}

\item{geneOrTxNames}{a character vector (default "tx"), should row names
keep trancript names ("tx") or change to gene names ("gene")}

\item{region}{a character vector (default: "mrna"), make raw count matrices
of whole mrnas or one of (leaders, cds, trailers).
Can also be a \code{\link{GRangesList}}, then it uses this region directly.
Can then be uORFs or a subset of CDS etc.}

\item{type}{default: "count" (raw counts matrix), alternative is "fpkm",
"log2fpkm" or "log10fpkm"}

\item{lib.type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "ofst", "bedo", "bedoc" or "pshifted".
These are made with ORFik:::convertLibs() or shiftFootprintsByExperiment().
Can also be custom user made folders inside the experiments bam folder.}

\item{weight}{numeric or character, a column to score overlaps by. Default "score",
will check for a metacolumn called "score" in libraries. If not found,
will not use weights.}
}
\value{
a \code{\link{SummarizedExperiment}} object or data.table if
"type" is not "count, with rownames as transcript / gene names.
}
\description{
Make a summerizedExperiment / matrix object from bam files or
other library formats sepcified by lib.type argument. Works
like HTSeq, to give you count tables per library.
}
\details{
If txdb or gtf path is added, it is a rangedSummerizedExperiment
NOTE: If the file called saveName exists, it will then load file,
not remake it!
}
\examples{
##Make experiment
df <- ORFik.template.experiment()
# makeSummarizedExperimentFromBam(df)
## Only cds (coding sequences):
# makeSummarizedExperimentFromBam(df, region = "cds")
## FPKM instead of raw counts on whole mrna regions
# makeSummarizedExperimentFromBam(df, type = "fpkm")
## Make count tables of pshifted libraries over uORFs
#uORFs <- GRangesList("chr1", 1:9, "+)
#saveName <- file.path(dirname(df$filepath[1]), "uORFs", "countTable_uORFs")
#makeSummarizedExperimentFromBam(df, , region = uORFs)
## To load the uORFs later
# c
}
