% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{makeTxdbFromGenome}
\alias{makeTxdbFromGenome}
\title{Make txdb from genome}
\usage{
makeTxdbFromGenome(gtf, genome = NULL, organism)
}
\arguments{
\item{gtf}{path to gtf file}

\item{genome}{character, default NULL. Path to fasta genome
corresponding to the gtf. If NULL, can not set seqlevels.}

\item{organism}{Scientific name of organism, first letter
must be capital! Example: Homo sapiens. Will force first letter
to capital and convert any "_" (underscore) to " " (space)}
}
\value{
NULL, Txdb saved to disc named paste0(gtf, ".db")
}
\description{
Make a Txdb with defined seqlevels and
seqlevelsstyle from the fasta genome.
This makes it more fail safe than standard Txdb creation.
Example is that you can not create a coverage window outside
the chromosome boundary, this is only possible if you have
set the seqlengths.
}
