% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{map_to_GRanges}
\alias{map_to_GRanges}
\title{Creates list of IRanges with Open Reading Frames.}
\usage{
map_to_GRanges(ORFdef, grangesObj, transcriptName = "")
}
\arguments{
\item{ORFdef}{List of IRanges objects representing found ORFs.}

\item{grangesObj}{GRanges object to map coordinates back from each IRanges ORF.}

\item{transcriptName}{String of name that will be added as metadata name column.}
}
\value{
A GRanges objects of ORFs mapped to grangesObj.
}
\description{
Creates list of IRanges with Open Reading Frames.
}
\examples{
#map_to_GRanges() #rewrite into C++

}
