% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{map_to_GRanges}
\alias{map_to_GRanges}
\title{Creates GRangesList from the results of get_all_ORFs_as_GRangesList and
 a list of group indeces}
\usage{
map_to_GRanges(grl, result)
}
\arguments{
\item{grl}{GRangesList. A GRangesList of the original sequences that gave the orfs}

\item{result}{List. A List of the results of finding uorfs
List syntax is: result[1] contain grouping indeces,
result[2] countains two columns of start and stops}
}
\value{
A GRangesList of ORFs.
}
\description{
Creates GRangesList from the results of get_all_ORFs_as_GRangesList and
 a list of group indeces
}
