% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{mapToGRanges}
\alias{mapToGRanges}
\title{Map orfs to genomic coordinates}
\usage{
mapToGRanges(grl, result)
}
\arguments{
\item{grl}{A \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}} of the original
sequences that gave the orfs}

\item{result}{List. A list of the results of finding uorfs
list syntax is: result(1) contain grouping indeces, named index
result(2) countains two columns of start and stops,  named orf}
}
\value{
A \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}} of ORFs.
}
\description{
Creates GRangesList from the results of ORFs_as_List and
the GRangesList used to find the ORFs
}
