% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{matchSeqlevels}
\alias{matchSeqlevels}
\title{Match seqnames}
\usage{
matchSeqlevels(filteredCage, fiveSeqlevels)
}
\arguments{
\item{filteredCage}{Cage-data to check seqnames in}

\item{fiveSeqlevels}{A vector of seqlevels for 5' UTRs (seqlevels(fiveUTRs))}
}
\value{
filteredCage with matched seqnames convention
}
\description{
Check that seqlevels of fiveUTRs and cage uses
the same standard, i.g chr1 vs 1.
}
