% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CageData_Integration.R
\name{matchSeqnames}
\alias{matchSeqnames}
\title{Check that seqnames of fiveUTRs and cage uses same standard, i.g chr1 vs 1.}
\usage{
matchSeqnames(filteredCage, fiveUTRs)
}
\arguments{
\item{filteredCage}{Cage-data to check seqnames in}

\item{fiveUTRs}{The 5' leader sequences as GRangesList}
}
\description{
Check that seqnames of fiveUTRs and cage uses same standard, i.g chr1 vs 1.
}
