% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchSeqStyle}
\alias{matchSeqStyle}
\title{A wrapper for seqlevelsStyle}
\usage{
matchSeqStyle(range, chrStyle = NULL)
}
\arguments{
\item{range}{a ranged object, (GRanges, GAlignment etc)}

\item{chrStyle}{a GRanges object, TxDb, FaFile,
, a \code{\link{seqlevelsStyle}} or \code{\link{Seqinfo}}.
(Default: NULL) to get seqlevelsStyle from. In addition if it
is a Seqinfo object, seqinfo will be updated.
Example of seqlevelsStyle update:
Is chromosome 1 called chr1 or 1,
 is mitocondrial chromosome called MT or chrM etc.
Will use 1st seqlevel-style if more are present.
Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a GAlignment/GRanges object depending on input.
}
\description{
To make sure chromosome naming is correct (chr1 vs 1 vs I etc)
}
\keyword{internal}
