% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{mergeLibs}
\alias{mergeLibs}
\title{Merge and save libraries of experiment}
\usage{
mergeLibs(
  df,
  out_dir = file.path(dirname(df$filepath[1]), "ofst_merged"),
  mode = "all",
  type = "ofst",
  keep_all_scores = TRUE
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out_dir}{Ouput directory, default \code{file.path(dirname(df$filepath[1]), "ofst_merged")},
saved as "all.ofst" in this folder if mode is "all". Use a folder called pshifted_merged, for
default Ribo-seq ofst files.}

\item{mode}{character, default "all". Merge all or "rep" for collapsing replicates only.
lib}

\item{type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "ofst" or "pshifted".
These are made with ORFik:::convertLibs() or shiftFootprintsByExperiment().
Can also be custom user made folders inside the experiments bam folder.
It acts in a recursive manner with priority: If you state "pshifted",
but it does not exist, it checks "ofst". If no .ofst files, it uses
"default", which always must exists.}

\item{keep_all_scores}{logical, default TRUE, keep all library scores in the merged file. These
score columns are named the libraries full name from \code{bamVarName(df)}.}
}
\value{
NULL, files saved to disc. A data.table with a score column that now contains the sum
of scores per merge setting.
}
\description{
Aggregate count of reads (from the "score" column) by making a merged library.
Only allowed for .ofst files!
}
\examples{
df2 <- ORFik.template.experiment()
df2 <- df2[df2$libtype == "RFP",]
# Merge all
#mergeLibs(df2, tempdir(), mode = "all", type = "default")
# Read as GRanges with mcols
#fimport(file.path(tempdir(), "all.ofst"))
# Read as direct fst data.table
#read_fst(file.path(tempdir(), "all.ofst"))
# Collapse replicates
#mergeLibs(df2, tempdir(), mode = "all", type = "default")
}
