% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{nrow,experiment-method}
\alias{nrow,experiment-method}
\title{Internal nrow function for ORFik experiment
Number of runs in experiment}
\usage{
\S4method{nrow}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
number of rows in experiment (integer)
}
\description{
Internal nrow function for ORFik experiment
Number of runs in experiment
}
