% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{methods}
\name{nrow,experiment-method}
\alias{nrow,experiment-method}
\title{Internal nrow function for ORFik experiment}
\usage{
\S4method{nrow}{experiment}(x)
}
\arguments{
\item{object}{an ORFik experiment}
}
\value{
number of rows in experiment (integer)
}
\description{
Internal nrow function for ORFik experiment
}
