% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{numCodons}
\alias{numCodons}
\title{Get number of codons}
\usage{
numCodons(grl, as.integer = TRUE, keep.names = FALSE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object}

\item{as.integer}{a logical (TRUE), remove stub codons}

\item{keep.names}{a logical (FALSE)}
}
\value{
an integer vector
}
\description{
Length of object / 3.
Choose either only whole codons, or with stubs.
For orfs stubs are not relevant, since there are no correctly defined
ORFs that are 17 bases long etc.
}
