% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{numExonsPerGroup}
\alias{numExonsPerGroup}
\title{Get list of the number of exons per group}
\usage{
numExonsPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a GRangesList}

\item{keep.names}{a boolean, keep names or not}
}
\value{
an integer vector of counts
}
\description{
Can also be used generaly to get number of GRanges object
 per GRangesList group
}
