% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{optimized_txdb_path}
\alias{optimized_txdb_path}
\title{Get path for optimization files for txdb}
\usage{
optimized_txdb_path(txdb, create.dir = FALSE, stop.error = TRUE)
}
\arguments{
\item{txdb}{a loaded TxDb object}

\item{create.dir}{logical FALSE, if TRUE create the
optimization directory, this should only be called first time used.}

\item{stop.error}{logical TRUE}
}
\value{
a character file path, returns NULL if not valid
and stop.error is FALSE.
}
\description{
Get path for optimization files for txdb
}
