% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimizeReads}
\alias{optimizeReads}
\title{Find optimized subset of valid reads}
\usage{
optimizeReads(grl, reads)
}
\arguments{
\item{grl}{a GRangesList or GRanges object}

\item{reads}{a GRanges or GAlignment object}
}
\value{
the reads as GRanges or GAlignment
}
\description{
Find optimized subset of valid reads
}
