% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF.R
\docType{class}
\name{ORF-class}
\alias{ORF-class}
\title{An S4 class to represent an ORF.}
\description{
An S4 class to represent an ORF.
}
\section{Fields}{

\describe{
\item{\code{orfRanges}}{A GRanges of ORF}

\item{\code{trailerRanges}}{A GRanges of trailer}

\item{\code{family}}{character. Family of ORF e.g. 'lincRNA'}

\item{\code{isoform}}{character. What is isoform of this ORF toward the transcript}

\item{\code{frame}}{numeric.}

\item{\code{ORFLength}}{numeric. Length of ORF.}

\item{\code{trailerLength}}{numeric. Length of trailer.}

\item{\code{transcript}}{character. Name of the transcript.}

\item{\code{start}}{character. Defined start of the ORF. e.g. 'ATG'}

\item{\code{identifier}}{character. Unique identifier of this ORF.}
}}

