% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFik.R
\docType{package}
\name{ORFik-package}
\alias{ORFik}
\alias{ORFik-package}
\title{ORFik for analysis of open reading frames.}
\description{
Main goals:
\enumerate{
\item Find open reading frames
\item Compatible with GRanges
\item Make it fast enough for big data
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JokingHero/ORFik}
  \item Report bugs at \url{https://github.com/JokingHero/ORFik/issues}
}

}
\author{
\strong{Maintainer}: Kornel Labun \email{kornellabun@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Haakon Tjeldnes \email{hauken_heyken@hotmail.com} [data contributor]
}

Other contributors:
\itemize{
  \item Katazyna Chyzynska \email{katchyz@gmail.com} [contributor, data contributor]
  \item Evind Valen \email{eivind.valen@gmail.com} [thesis advisor, funder]
}

}
