% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFik.R
\docType{package}
\name{ORFik-package}
\alias{ORFik}
\alias{ORFik-package}
\title{ORFik for analysis of open reading frames.}
\description{
Main goals:
\enumerate{
\item Finding Open Reading Frames (very fast) in the genome of interest or
 on the set of transcripts/sequences.
\item Utilities for metaplots of RiboSeq coverage over gene START and STOP
codons allowing to spot the shift.
\item Shifting functions for the RiboSeq data.
\item Finding new Transcription Start Sites with the use of CageSeq data.
\item Various measurements of gene identity e.g. FLOSS, coverage, ORFscore,
entropy that are recreated based on many scientific publications.
\item Utility functions to extend GenomicRanges for faster grouping,
splitting, tiling etc.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JokingHero/ORFik}
  \item Report bugs at \url{https://github.com/JokingHero/ORFik/issues}
}

}
\author{
\strong{Maintainer}: Kornel Labun \email{kornellabun@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Haakon Tjeldnes \email{hauken_heyken@hotmail.com} [data contributor]
}

Other contributors:
\itemize{
  \item Katarzyna Chyzynska \email{katchyz@gmail.com} [contributor, data contributor]
  \item Evind Valen \email{eivind.valen@gmail.com} [thesis advisor, funder]
}

}
