% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{ORFik.template.experiment}
\alias{ORFik.template.experiment}
\title{An ORFik experiment to see how it looks}
\usage{
ORFik.template.experiment()
}
\value{
an ORFik \code{\link{experiment}}
}
\description{
NOTE! This experiment should only be used for testing, since
it is just sampled data internal in ORFik.
}
\examples{
ORFik.template.experiment()
}
