% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{ORFikQC}
\alias{ORFikQC}
\title{A post Alignment quality control of reads}
\usage{
ORFikQC(df)
}
\arguments{
\item{df}{an ORFik experiment, to make it, see: ?experiment}
}
\description{
From this report you will get a summary csv table, with distribution of
aligned reads, overlap of reads to transcript regions, like
leader, cds, trailer, tRNAs, rRNAs, snoRNAs etc.
It will also make you some correlation plots and meta coverage plots,
so you get a good understanding of how good the quality of your NGS
data production + aligner step were.
Everything will be outputed in the directory of your NGS data,
inside the folder QC_STATS/, relative to data location in 'df'
}
\details{
To make a ORFik experiment, see ?ORFik::experiment
}
