% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{ORFikQC}
\alias{ORFikQC}
\title{A post Alignment quality control of reads}
\usage{
ORFikQC(df, out.dir = dirname(df$filepath[1]))
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default: dirname(df$filepath[1])}
}
\value{
NULL (objects stored to disc)
}
\description{
From this report you will get a summary csv table, with distribution of
aligned reads, overlap of reads to transcript regions, like
leader, cds, trailer, tRNAs, rRNAs, snoRNAs etc.
It will also make you some correlation plots and meta coverage plots,
so you get a good understanding of how good the quality of your NGS
data production + aligner step were.
You will also get count tables over mrna, leader, cds and trailer
separately, similar to HTseq count tables. \cr
Everything will be outputed in the directory of your NGS data,
inside the folder ./QC_STATS/, relative to data location in 'df'.
You can specify new out location with out.dir if you want.\cr
To make a ORFik experiment, see ?ORFik::experiment \cr
To see some normal mrna coverage profiles of different RNA-seq protocols:
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4310221/figure/F6/
}
\examples{
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
# Save with: save.experiment(df, file = "path/to/save/experiment.csv")

# ORFikQC(df)
}
\seealso{
Other ORFikQC: 
\code{\link{QCplots}()},
\code{\link{QCreport}()}
}
\concept{ORFikQC}
