% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods2.R
\name{orfScore}
\alias{orfScore}
\title{Get ORFscore for a GRangesList of ORFs}
\usage{
orfScore(grl, RFP)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object with ORFs}

\item{RFP}{ribozomal footprints, given as Galignment object,
Granges or GRangesList}
}
\value{
a matrix with 4 columns, the orfscore and score of
each of the 3 tiles
}
\description{
ORFscore tries to check whether the first frame of the 3 possible frames in
an ORF has more reads than second and third frame.
}
\details{
Pseudocode:
assume rff - is reads fraction in specific frame
\preformatted{ORFScore = log(rrf1 + rrf2 + rrf3)}
For all ORFs where rrf2 or rrf3 is bigger than rff1,
negate the resulting value.
\preformatted{ORFScore[rrf1Smaller] <- ORFScore[rrf1Smaller] * -1}

As result there is one value per ORF:
Positive values say that the first frame have the most reads,
negative values say that the first frame does not have the most reads.
}
\examples{
ORF <- GRanges(seqnames = "1",
               ranges = IRanges(start = c(1, 10, 20), end = c(5, 15, 25)),
               strand = "+")
names(ORF) <- c("tx1", "tx1", "tx1")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25), "+")
orfScore(grl, RFP)

}
\references{
doi: 10.1002/embj.201488411
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
