% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORF_features.R
\name{ORFScores}
\alias{ORFScores}
\title{Get ORF score for a GRangesList of ORFs}
\usage{
ORFScores(grl, RFP)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object with ORFs}

\item{RFP}{ribozomal footprints, given as Galignment object,
Granges or GRangesList}
}
\value{
a matrix with 4 columns, the orfscore and score of
each of the 3 tiles
}
\description{
That is, of the 3 possible frame in an ORF
Is the first one most important, by how much ?
NB! Only support + and - strand, not *
See article: 10.1002/embj.201488411
}
\seealso{
Other features: \code{\link{ORFDistToCds}},
  \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{RibosomeStallingScore}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subset_coverage}}, \code{\link{te}}
}
