% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{ORFStartSites}
\alias{ORFStartSites}
\title{get the start sites from a GRangesList of orfs grouped by orfs}
\usage{
ORFStartSites(grl, asGR = F, keep.names = F, is.sorted = F)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object}

\item{asGR}{a boolean, return as GRanges object}

\item{keep.names}{if asGR is False, do you still want
to keep a named vector}

\item{is.sorted}{a speedup, if you know the ranges are sorted}
}
\value{
if asGR is False, a vector, if True a GRanges object
}
\description{
In ATGTTTTGG, get the position of the A.
}
