% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{OrfToTxNames}
\alias{OrfToTxNames}
\title{Get transcript names from orf names}
\usage{
OrfToTxNames(grl, unique = F)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} grouped by ORF
or GRanges object}

\item{unique}{a boolean, if true unique the names,
used if several orfs map to same transcript and you only
want the unique groups}
}
\value{
a character vector of transcript names,
 without _* naming
}
\description{
names must either be a column called names, or the names of the
grl object
}
