% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{OrfToTxNames}
\alias{OrfToTxNames}
\title{Get transcript names from orf names
names must either be a column called names, or the names of the
grl object}
\usage{
OrfToTxNames(grl, unique = F)
}
\arguments{
\item{grl}{a GRangeslist grouped by ORF}

\item{unique}{a boolean, if true unique the names,
used if several orfs map to same transcript and you only
want the unique groups
@export}
}
\description{
Get transcript names from orf names
names must either be a column called names, or the names of the
grl object
}
