% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{outputLibs}
\alias{outputLibs}
\title{Output bam/bed/bedo/bedoc/ofst/wig files to R as variables}
\usage{
outputLibs(
  df,
  chrStyle = NULL,
  type = "default",
  envir = .GlobalEnv,
  BPPARAM = bpparam()
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{chrStyle}{a GRanges object, TxDb, FaFile,
or a \code{\link{seqlevelsStyle}}
(Default: NULL) to get seqlevelsStyle from. Is chromosome 1
called chr1 or 1, is mitocondrial chromosome called MT or chrM etc.
Will use 1st seqlevel-style if more are present.
Like: c("NCBI", "UCSC") -> pick "NCBI"}

\item{type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "bedo", "bedoc", "ofst or "pshifted".
These are made with ORFik:::simpleLibs(), shiftFootprintsByExperiment()..}

\item{envir}{environment to save to, default (.GlobalEnv)}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}}
}
\value{
NULL (libraries set by envir assignment)
}
\description{
Variable names defined by df (ORFik experiment DataFrame)
Uses multiple cores to load, defined by multicoreParam
}
\examples{
## Load a template ORFik experiment
df <- ORFik.template.experiment()
## Default library type load, usually bam files
# outputLibs(df, type = "default")
## .ofst file load, if ofst files does not exists
## it will load default
# outputLibs(df, type = "ofst")
## .wig file load, if wiggle files does not exists
## it will load default
# outputLibs(df, type = "wig")
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism.df}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
