% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{outputLibs}
\alias{outputLibs}
\title{Output bam/bed/wig files to R as variables}
\usage{
outputLibs(df, chrStyle = NULL, envir = .GlobalEnv, BPPARAM = bpparam())
}
\arguments{
\item{df}{an ORFik experiment, to make it, see: ?experiment}

\item{chrStyle}{the sequencelevels style (GRanges object or chr)}

\item{envir}{environment to save to, default (.GlobalEnv)}

\item{BPPARAM}{how many cores? default: bpparam()}
}
\value{
NULL (libraries set by envir assignment)
}
\description{
Variable names defined by df (ORFik experiment DataFrame)
Uses multiple cores to load, defined by multicoreParam
}
\examples{
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
# Output to .GlobalEnv with:
# outputLibs(df)
}
