% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{outputLibs}
\alias{outputLibs}
\title{Output bam/bed/bedo/wig files to R as variables}
\usage{
outputLibs(
  df,
  chrStyle = NULL,
  type = "default",
  envir = .GlobalEnv,
  BPPARAM = bpparam()
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{chrStyle}{a GRanges object, TxDb, FaFile,
or a \code{\link{seqlevelsStyle}}
(Default: NULL) to get seqlevelsStyle from. Is chromosome 1
called chr1 or 1, is mitocondrial chromosome called MT or chrM etc.
Will use 1st seqlevel-style if more are present.
Like: c("NCBI", "UCSC") -> pick "NCBI"}

\item{type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "bedo", "bedoc" or "pshifted".
These are made with ORFik:::simpleLibs(), shiftFootprintsByExperiment()..}

\item{envir}{environment to save to, default (.GlobalEnv)}

\item{BPPARAM}{how many cores/threads to use? default: bpparam()}
}
\value{
NULL (libraries set by envir assignment)
}
\description{
Variable names defined by df (ORFik experiment DataFrame)
Uses multiple cores to load, defined by multicoreParam
}
\examples{
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
# Output to .GlobalEnv with:
# outputLibs(df)
}
\seealso{
Other ORFik_experiment: 
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism.df}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
