% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plots.R
\name{pcaExperiment}
\alias{pcaExperiment}
\title{Simple PCA analysis}
\usage{
pcaExperiment(
  df,
  output.dir = NULL,
  table = countTable(df, "cds", type = "fpkm"),
  title = "CDS fpkm (All genes)",
  plot.ext = ".pdf"
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{output.dir}{default NULL, else character path to directory.
File saved as "PCAplot_(experiment name)(plot.ext)"}

\item{table}{data.table, default countTable(df, "cds", type = "fpkm"),
a data.table of counts per column (default normalized fpkm values).}

\item{title}{character, default "CDS fpkm (All genes)".}

\item{plot.ext}{character, default: ".pdf". Alternatives: ".png" or ".jpg".}
}
\value{
ggplot or invisible(NULL) if output.dir is defined
}
\description{
Detect outlier libraries with PCA analysis.
Will output PCA plot of PCA component 1 (x-axis) vs
PCA component 2 (y-axis) for each library (colored by library),
shape by replicate. Will be extended to allow batch correction
in the future.
}
