% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{pmapFromTranscriptF}
\alias{pmapFromTranscriptF}
\title{Faster more secure version of mapFromTranscripts}
\usage{
pmapFromTranscriptF(ranges, grl, indices)
}
\arguments{
\item{ranges}{IRanges of ranges within grl}

\item{grl}{the "transcripts" that contain ranges, GRangesList}

\item{indices}{integer vector of which index of grl ranges are from:
(c(1,1,2)) means first two ranges are from grl[1], third from grl[2])}
}
\value{
A GrangesList of ranges mapped from transcripts
}
\description{
Fixes a bug in function, and should have 10x speedup
Also removes hit column for you
}
\seealso{
Other ExtendGenomicRanges: \code{\link{asTX}},
  \code{\link{coveragePerTiling}},
  \code{\link{overlapsToCoverage}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{txSeqsFromFa}}, \code{\link{windowPerGroup}}
}
\concept{ExtendGenomicRanges}
