% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{pSitePlot}
\alias{pSitePlot}
\title{Plot area around TIS for p-shifted reads}
\usage{
pSitePlot(hitMap, length = 29, region = "start")
}
\arguments{
\item{hitMap}{a data.frame, given from metaWindow}

\item{length}{an integer (29), which length is this for?}

\item{region}{a character (start), either "start or "stop".}
}
\value{
a ggplot
}
\description{
Usefull to validate p-shifting is correct
}
