% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{QCfolder,experiment-method}
\alias{QCfolder,experiment-method}
\title{Get ORFik experiment QC folder path}
\usage{
\S4method{QCfolder}{experiment}(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get ORFik experiment QC folder path
}
