% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{QCfolder}
\alias{QCfolder}
\title{Get ORFik experiment QC folder path}
\usage{
QCfolder(x, value)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a character path
}
\description{
Get ORFik experiment QC folder path
}
