% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCplots}
\alias{QCplots}
\title{Correlation and coverage plots for ORFikQC}
\usage{
QCplots(df, region, stats_folder)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{region}{(default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{stats_folder}{directory to save}
}
\value{
NULL (objects stored to disc)
}
\description{
Output will be stored in same folder as the
libraries in df.
}
