% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCplots}
\alias{QCplots}
\title{Correlation and coverage plots for ORFikQC}
\usage{
QCplots(
  df,
  region = "mrna",
  stats_folder = paste0(dirname(df$filepath[1]), "/QC_STATS/"),
  plot.ext = ".pdf",
  BPPARAM
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{region}{a character (default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{stats_folder}{directory to save, default:
paste0(dirname(df$filepath[1]), "/QC_STATS/")}

\item{plot.ext}{character, default: ".pdf". Alternatives: ".png" or ".jpg".}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}.
You can also add a time remaining bar, for a more detailed pipeline.}
}
\value{
invisible(NULL) (objects stored to disc)
}
\description{
Correlation plots default to mRNA covering reads.
Meta plots defaults to leader, cds, trailer.\cr
Output will be stored in same folder as the
libraries in df.\cr
Correlation plots will be fpkm correlation and
log2(fpkm + 1) correlation between samples.
}
\details{
Is part of \code{\link{QCreport}}
}
\seealso{
Other QC report: 
\code{\link{QCreport}()},
\code{\link{QCstats}()}
}
\concept{QC report}
\keyword{internal}
