% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCplots}
\alias{QCplots}
\title{Correlation and coverage plots for ORFikQC}
\usage{
QCplots(
  df,
  region = "mrna",
  stats_folder = QCfolder(df),
  plot.ext = ".pdf",
  complex.correlation.plots = TRUE,
  library.names = bamVarName(df),
  force = TRUE,
  BPPARAM
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{region}{a character (default: mrna), make raw count matrices of
whole mrnas or one of (leaders, cds, trailers)}

\item{stats_folder}{directory to save, default:
\code{QCfolder(df)}}

\item{plot.ext}{character, default: ".pdf". Alternatives: ".png" or ".jpg".
Note that in pdf format the complex correlation plots become very slow to load!}

\item{complex.correlation.plots}{logical, default TRUE. Add in addition
to simple correlation plot two computationally heavy dots + correlation plots.
Useful for deeper analysis, but takes longer time to run, especially on low-quality
gpu computers. Set to FALSE to skip these.}

\item{library.names}{character, default: bamVarName(df). Names to load
libraries as to environment and names to display in plots.}

\item{force}{logical, default TRUE If TRUE, reload library files even if
matching named variables are found in environment used by experiment
 (see \code{\link{envExp}}) A simple way to make
sure correct libraries are always loaded. FALSE is faster if data
is loaded correctly already.}

\item{BPPARAM}{how many cores/threads to use? default: bpparam().
To see number of threads used, do \code{bpparam()$workers}.
You can also add a time remaining bar, for a more detailed pipeline.}
}
\value{
invisible(NULL) (objects stored to disc)
}
\description{
Correlation plots default to mRNA covering reads.
Meta plots defaults to leader, cds, trailer.\cr
Output will be stored in same folder as the
libraries in df.\cr
Correlation plots will be fpkm correlation and
log2(fpkm + 1) correlation between samples.
}
\details{
Is part of \code{\link{QCreport}}
}
\seealso{
Other QC report: 
\code{\link{QCreport}()},
\code{\link{QCstats}()}
}
\concept{QC report}
\keyword{internal}
