% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{QCreport}
\alias{QCreport}
\title{A post Alignment quality control of reads}
\usage{
QCreport(df, out.dir = dirname(df$filepath[1]))
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default: dirname(df$filepath[1]).
Will make a folder called "QC_STATS" with all results in this directory.}
}
\value{
invisible(NULL) (objects are stored to disc)
}
\description{
The ORFik QC uses the aligned files (usually bam files),
fastp and STAR log files
combined with annotation to create relevant statistics.\cr\cr
This report consists of several steps:\cr
1. From this report you will get a summary csv table, with distribution of
aligned reads and overlap counts over transcript regions like:
leader, cds, trailer, lincRNAs, tRNAs, rRNAs, snoRNAs etc. It will be called
STATS.csv. And can be imported with \code{\link{QCstats}} function.\cr
2. It will also make correlation plots and meta coverage plots,
so you get a good understanding of how good the quality of your NGS
data production + aligner step were.\cr
3. Count tables are produced, similar to HTseq count tables.
Over mrna, leader, cds and trailer separately. This tables
are stored as \code{\link{SummarizedExperiment}}, for easy loading into
DEseq, conversion to normalized fpkm values,
or collapsing replicates in an experiment.
And can be imported with \code{\link{countTable}} function.\cr
Everything will be outputed in the directory of your NGS data,
inside the folder ./QC_STATS/, relative to data location in 'df'.
You can specify new out location with out.dir if you want.\cr
To make a ORFik experiment, see ?ORFik::experiment \cr
To see some normal mrna coverage profiles of different RNA-seq protocols:
https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4310221/figure/F6/
}
\examples{
# Load an experiment
df <- ORFik.template.experiment()
# Run QC
# QCreport(df)
}
\seealso{
Other QC report: 
\code{\link{QCplots}()},
\code{\link{QCstats}()}
}
\concept{QC report}
