% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{QCstats.plot}
\alias{QCstats.plot}
\title{Make plot of ORFik QCreport}
\usage{
QCstats.plot(stats, output.dir = NULL)
}
\arguments{
\item{stats}{path to ORFik QC stats .csv file, or the experiment object.}

\item{output.dir}{NULL or character path, default: NULL, plot not saved to disc.
If defined saves plot to that directory with the name "/STATS_plot.png".}
}
\value{
ggplot object of the the statistics data
}
\description{
From post-alignment QC relative to annotation, make a plot for all samples.
Will contain among others read lengths, reads overlapping leaders,
cds, trailers, mRNA / rRNA etc.
}
\examples{
df <- ORFik.template.experiment()[3,]
## First make QC report
# QCreport(df)
## Now you can get plot
# QCstats.plot(df)
}
