% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{read.experiment}
\alias{read.experiment}
\title{Read ORFik experiment}
\usage{
read.experiment(file)
}
\arguments{
\item{file}{a .csv file following ORFik experiment style.}
}
\value{
an ORFik experiment
}
\description{
Read ORFik experiment
}
