% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{read.experiment}
\alias{read.experiment}
\title{Read ORFik \code{\link{experiment}}}
\usage{
read.experiment(file)
}
\arguments{
\item{file}{a .csv file following ORFik experiment style ("," as seperator)
, or a template data.frame from create.experiment()}
}
\value{
an ORFik \code{\link{experiment}}
}
\description{
The file must be csv and be a valid ORFik experiment
}
\examples{
# From file
\dontrun{
# Read from file
df <- read.experiment(filepath) # <- valid .csv file
}
# Read from (create.experiment() template)
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment from template
df <- read.experiment(template)

# To save it, do:
# save.experiment(df, file = "path/to/save/experiment.csv")
}
\seealso{
Other ORFik_experiment: 
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{libraryTypes}()},
\code{\link{outputLibs}()},
\code{\link{save.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
