% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readBam}
\alias{readBam}
\title{Custom bam reader}
\usage{
readBam(path, chrStyle = NULL)
}
\arguments{
\item{path}{a character path to .bam file}

\item{chrStyle}{a GRanges object, or a \code{\link{seqlevelsStyle}}
(Default: NULL) to get seqlevelsStyle from. Is chromosome 1
called chr1 or 1, is mitocondrial chromosome called MT or chrM etc.
Will use 1st seqlevel-style if more are present.
Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a \code{\link{GAlignments}} object of bam file
}
\description{
Only for single end reads
Safer version that handles the most important error done.
In the future will use a faster .bam loader for big .bam files in R.
}
\examples{
bam_file <- system.file("extdata", "ribo-seq.bam", package = "ORFik")
readBam(bam_file, "UCSC")
}
\seealso{
Other utils: 
\code{\link{bedToGR}()},
\code{\link{convertToOneBasedRanges}()},
\code{\link{fimport}()},
\code{\link{findFa}()},
\code{\link{fread.bed}()},
\code{\link{optimizeReads}()},
\code{\link{readWig}()}
}
\concept{utils}
