% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readWig}
\alias{readWig}
\title{Custom wig reader}
\usage{
readWig(path, chrStyle = NULL)
}
\arguments{
\item{path}{a character path to .bam file}

\item{chrStyle}{a GRanges object, or a character style (Default: NULL) to
get seqlevelsStyle from. Is chromosome 1 called chr1 or 1, is mitocondrial
chromosome called MT or chrM etc. Will use 1st seqlevel-
style if more are present. Like: c("NCBI", "UCSC") -> pick "NCBI"}
}
\value{
a GAlignment object of bam file
}
\description{
Given 2 wig files, first is forward second is reverse.
Merge them and return as GRanges object.
If they contain name reverse and forward, first and second order
does not matter, it will search for forward and reverse.
}
\seealso{
Other utils: \code{\link{bedToGR}},
  \code{\link{convertToOneBasedRanges}},
  \code{\link{findFa}}, \code{\link{fread.bed}},
  \code{\link{optimizeReads}}, \code{\link{readBam}}
}
\concept{utils}
