% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{reassignTSSbyCage}
\alias{reassignTSSbyCage}
\title{Reassign all Transcript Start Sites (TSS)}
\usage{
reassignTSSbyCage(fiveUTRs, cage, extension = 1000, filterValue = 1,
  cds = NULL)
}
\arguments{
\item{fiveUTRs}{(GRangesList) The 5' leaders or transcript sequences}

\item{cage}{Either a filePath for CageSeq file, or already loaded
CageSeq peak data as GRanges.}

\item{extension}{The maximum number of basses upstream of the TSS to search
for CageSeq peak.}

\item{filterValue}{The minimum number of reads on cage position,
for it to be counted as possible new tss.
(represented in score column in CageSeq data) to
If you already filtered, set it to 0.}

\item{cds}{(GRangesList) CDS of relative fiveUTRs, applicable only if you
want to extend 5' leaders downstream of CDS's, to allow upstream ORFs that
can overlap into CDS's.}
}
\value{
a GRangesList of newly assigned TSS for fiveUTRs, using CageSeq data.
}
\description{
Given a GRangesList of 5' UTRs or transcripts, reassign the start
postitions using max peaks from CageSeq data. A max peak is defined as new
TSS if it is within boundary of 5' leader range, specified by
\code{extension} in bp. A max peak must also be higher than minimum
CageSeq peak cutoff specified in \code{filterValue}. The new TSS will then
be the positioned where the cage read (with highest read count in the
interval).
}
\examples{
# example 5' leader, notice exon_rank column
fiveUTRs <- GenomicRanges::GRangesList(
  GenomicRanges::GRanges(seqnames = "chr1",
                         ranges = IRanges::IRanges(1000, 2000),
                         strand = "+",
                         exon_rank = 1))
names(fiveUTRs) <- "tx1"

# make fake CageSeq data from promoter of 5' leaders, notice score column
cage <- GenomicRanges::GRanges(
  seqnames = "chr1",
  ranges =  IRanges::IRanges(500, 510),
  strand = "+",
  score = 10)

# finally reassign TSS for fiveUTRs
reassignTSSbyCage(fiveUTRs, cage)

}
