% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{remove.experiments}
\alias{remove.experiments}
\title{Remove bam/bed/wig files load in R as variables}
\usage{
remove.experiments(df, envir = .GlobalEnv)
}
\arguments{
\item{df}{an ORFik experiment, to make it, see: ?experiment}

\item{envir}{environment to save to, default (.GlobalEnv)}
}
\value{
NULL (objects removed from envir specified)
}
\description{
Variable names defined by df, in envir defined
}
\examples{
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
# Output to .GlobalEnv with:
# outputLibs(df)
# Then remove them with: remove.experiments(df)
}
