% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{remove.experiments}
\alias{remove.experiments}
\title{Remove bam/bed/wig files load in R as variables}
\usage{
remove.experiments(df, envir = .GlobalEnv)
}
\arguments{
\item{df}{an ORFik experiment, to make it, see: ?experiment}

\item{envir}{environment to save to, default (.GlobalEnv)}
}
\description{
Variable names defined by df, in envir defined
}
