% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cage_annotations.R
\name{restrictTSSByUpstreamLeader}
\alias{restrictTSSByUpstreamLeader}
\title{Restrict extension of 5' UTRs to closest upstream leader end}
\usage{
restrictTSSByUpstreamLeader(fiveUTRs, shiftedfiveUTRs)
}
\arguments{
\item{fiveUTRs}{The 5' leader sequences as GRangesList}

\item{shiftedfiveUTRs}{The 5' leader sequences as GRangesList
shifted by CAGE}
}
\value{
GRangesList object of restricted fiveUTRs
}
\description{
Basicly this function restricts all startSites, to the upstream GRangesList
 objects end. Usually leaders, for CAGE.
}
