% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{reverseMinusStrandPerGroup}
\alias{reverseMinusStrandPerGroup}
\title{Reverse minus strand}
\usage{
reverseMinusStrandPerGroup(grl)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}
}
\value{
a \code{\link{GRangesList}}
}
\description{
Reverse minus strand per group in a GRangesList
}
