% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{RibosomeReleaseScore}
\alias{RibosomeReleaseScore}
\title{Ribosome Release Score (RRS)}
\usage{
RibosomeReleaseScore(grl, RFP, GtfOrThreeUtrs, RNA = NULL)
}
\arguments{
\item{grl}{a GRangesList object with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{GtfOrThreeUtrs}{if Gtf: a TxDb object of a gtf file,
if ThreeUtrs: a GrangesList of transcripts, called from:
threeUTRsByTranscript(Gtf, use.names = T)}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}
}
\value{
a named vector of numeric values of scores
}
\description{
is defined as (RPFs over ORF)/(RPFs over 3' utrs).
normalized by lengths
, if RNA is added as argument, normalize by RNA counts over areas too
to justify location of 3' utrs
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1016/j.cell.2013.06.009
}
