% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{ribosomeStallingScore}
\alias{ribosomeStallingScore}
\title{Ribosome Stalling Score (RSS)}
\usage{
ribosomeStallingScore(grl, RFP)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
with usually either leaders,
cds', 3' utrs or ORFs. ORFs are a special case, see argument tx_len}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}
}
\value{
a named vector of numeric values of scores
}
\description{
Is defined as (RPFs over ORF stop sites)/(RPFs over ORFs),
normalized by lengths
A pseudo-count of one was added to both the ORF and downstream sums.
See article: 10.1016/j.cels.2017.08.004
}
\details{
For a more accurate analysis see article:  10.1016/j.cels.2017.08.004.
}
\examples{
ORF <- GRanges(seqnames = "1",
ranges = IRanges(start = c(1, 10, 20),
                end = c(5, 15, 25)),
strand = "+")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25),"+")
ribosomeStallingScore(grl, RFP)
}
\seealso{
Other features: \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
