% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{riboTISCoverageProportion}
\alias{riboTISCoverageProportion}
\title{Find proportion of reads per position in ORF TIS window}
\usage{
riboTISCoverageProportion(grl, tx, footprints, onlyProportion = FALSE,
  average = FALSE, pShifted = TRUE, keep.names = FALSE,
  upStart = if (pShifted) 5 else 20, downStop = if (pShifted) 20 else
  5)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually new ORFs, but can also be
either leaders, cds', 3'UTRs.}

\item{tx}{a GrangesList of transcripts covering grl.}

\item{footprints}{ribo seq reads as GAlignment, GRanges
or GRangesList object.}

\item{onlyProportion}{a logical (FALSE), return whole data.frame or only
proportions}

\item{average}{a logical (FALSE), take average over coverage in all grl ?}

\item{pShifted}{a logical (TRUE), are riboseq reads p-shifted?}

\item{keep.names}{a logical(FALSE), only applies when onlyProportion
is TRUE.}
}
\value{
a data.frame with lengths by coverage / vector of proportions
}
\description{
Proportion defined as:
average count per position in -20,20 normalized by counts per gene
}
\details{
This pattern can be averaged on CDS's, to find other ORFs.
When detecting new ORFs, this CDS average can be used as a template.
}
