% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{save.experiment}
\alias{save.experiment}
\title{Save \code{\link{experiment}} to disc}
\usage{
save.experiment(df, file)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{file}{name of file to save df as}
}
\value{
NULL (experiment save only)
}
\description{
Save \code{\link{experiment}} to disc
}
\examples{
df <- ORFik.template.experiment()
## Save with:
#save.experiment(df, file = "path/to/save/experiment.csv")
## Identical (.csv not needed, can be added):
#save.experiment(df, file = "path/to/save/experiment")
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment.zf}()},
\code{\link{ORFik.template.experiment}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism,experiment-method}},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
