% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{save.experiment}
\alias{save.experiment}
\title{Save experiment to disc}
\usage{
save.experiment(df, file)
}
\arguments{
\item{df}{an ORFik experiment, to make it, see: ?experiment}

\item{file}{name of file to save df as}
}
\value{
NULL (experiment save only)
}
\description{
Save experiment to disc
}
