% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{save.experiment}
\alias{save.experiment}
\title{Save \code{\link{experiment}} to disc}
\usage{
save.experiment(df, file)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{file}{name of file to save df as}
}
\value{
NULL (experiment save only)
}
\description{
Save \code{\link{experiment}} to disc
}
\examples{
# 1. Pick directory
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
# Save with: save.experiment(df, file = "path/to/save/experiment.csv")
}
\seealso{
Other ORFik_experiment: 
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism.df}()},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{validateExperiments}()}
}
\concept{ORFik_experiment}
