% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{savePlot}
\alias{savePlot}
\title{Helper function for writing plots to disc}
\usage{
savePlot(plot, output = NULL, width = 200, height = 150, dpi = 150)
}
\arguments{
\item{plot}{the ggplot to save}

\item{output}{character string (NULL), if set, saves the plot as pdf or png
to path given. If no format is given, is save as pdf.}

\item{width}{width of output in mm}

\item{height}{height of output in mm}

\item{dpi}{(150) dpi of plot}
}
\value{
a ggplot object of the coverage plot, NULL if output is set,
then the plot will only be saved to location.
}
\description{
Helper function for writing plots to disc
}
\seealso{
Other coveragePlot: \code{\link{coverageHeatMap}},
  \code{\link{pSitePlot}}, \code{\link{windowCoveragePlot}}
}
\concept{coveragePlot}
