% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedExperiment_helpers.R
\name{scoreSummarizedExperiment}
\alias{scoreSummarizedExperiment}
\title{Helper function for makeSummarizedExperimentFromBam}
\usage{
scoreSummarizedExperiment(final, score = "transcriptNormalized",
  collapse = FALSE)
}
\arguments{
\item{final}{ranged summarized experiment object}

\item{score}{default: "transcriptNormalized"
(row normalized raw counts matrix),
alternative is "fpkm", "log2fpkm" or "log10fpkm"}

\item{collapse}{a logical (default FALSE), if TRUE all samples
within group}
}
\value{
a DEseq summerizedExperiment object
}
\description{
If txdb or gtf path is added, it is a rangedSummerizedExperiment
}
