% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{seqnamesPerGroup}
\alias{seqnamesPerGroup}
\title{Get list of seqnames per granges group}
\usage{
seqnamesPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
a character vector or Rle of seqnames(if seqnames == T)
}
\description{
Get list of seqnames per granges group
}
