% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{seqnamesPerGroup}
\alias{seqnamesPerGroup}
\title{get list of seqnames per granges group}
\usage{
seqnamesPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\description{
get list of seqnames per granges group
}
