% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grl_helpers.R
\name{seqnamesPerGroup}
\alias{seqnamesPerGroup}
\title{Get list of seqnames per granges group}
\usage{
seqnamesPerGroup(grl, keep.names = TRUE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
a character vector or Rle of seqnames(if seqnames == T)
}
\description{
Get list of seqnames per granges group
}
\examples{
gr_plus <- GRanges(seqnames = c("chr1", "chr1"),
                   ranges = IRanges(c(7, 14), width = 3),
                   strand = c("+", "+"))
gr_minus <- GRanges(seqnames = c("chr2", "chr2"),
                    ranges = IRanges(c(4, 1), c(9, 3)),
                    strand = c("-", "-"))
grl <- GRangesList(tx1 = gr_plus, tx2 = gr_minus)
seqnamesPerGroup(grl)
}
