% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{methods}
\name{show,experiment-method}
\alias{show,experiment-method}
\title{experiment show definition}
\usage{
\S4method{show}{experiment}(object)
}
\arguments{
\item{object}{an ORFik experiment}
}
\description{
Show a simplified version of experiment.
}
