% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{simpleLibs}
\alias{simpleLibs}
\title{Will make a simplified version of libraries}
\usage{
simpleLibs(df, out.dir = paste0(dirname(df$filepath[1]), "/bedo/"))
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default: dirname(df$filepath[1])}
}
\value{
NULL (saves files to disc)
}
\description{
A .obed file containing chromosome, start, stop, strand,
readWidth and number of duplicate reads
}
