% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{simpleLibs}
\alias{simpleLibs}
\title{Will make a simplified version of NGS libraries}
\usage{
simpleLibs(df, out.dir = paste0(dirname(df$filepath[1]), "/bedo/"))
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default: dirname(df$filepath[1]),
if it is NULL, it will just reassign R objects to simplified libraries.}
}
\value{
NULL (saves files to disc or R .GlobalEnv)
}
\description{
An .obed file containing chromosome, start, stop, strand,
readWidth and number of duplicate reads.
A bed file with 2 score columns
Gives a massive speedup when cigar strings are not needed.
}
