% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{simpleLibs}
\alias{simpleLibs}
\title{Converted format of NGS libraries}
\usage{
simpleLibs(
  df,
  out.dir = dirname(df$filepath[1]),
  addScoreColumn = TRUE,
  addSizeColumn = TRUE,
  must.overlap = NULL,
  method = "None",
  type = "ofst",
  reassign.when.saving = FALSE,
  envir = .GlobalEnv
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{out.dir}{optional output directory, default:
dirname(df$filepath[1]),
if it is NULL, it will just reassign R objects to simplified libraries.}

\item{addScoreColumn}{logical, default TRUE, if FALSE will not add
replicate numbers as score column, see ORFik::convertToOneBasedRanges.}

\item{addSizeColumn}{logical, default TRUE, if FALSE will not add
size (width) as size column, see ORFik::convertToOneBasedRanges.
Does not apply for .ofst or .bedoc.}

\item{must.overlap}{default (NULL), else a GRanges / GRangesList object, so
only reads that overlap (must.overlap) are kept. This is useful when you
only need the reads over transcript annotation or subset etc.}

\item{method}{character, default "None", the method to reduce ranges,
for more info see \code{\link{convertToOneBasedRanges}}}

\item{type}{a character of format, default "ofst".
Alternatives: "ofst", "wig","bedo" or "bedoc". Which format you want.
Will make a folder within out.dir with this name containing the files.}

\item{reassign.when.saving}{logical, default FALSE. If TRUE, will reassign
library to converted form after saving. Ignored when out.dir = NULL.}

\item{envir}{which environment to save to, default .GlobalEnv}
}
\value{
NULL (saves files to disc or R .GlobalEnv)
}
\description{
Export as either .ofst, .wig, .bedo or .bedoc files:\cr
Export files as .ofst for fastest load speed into R.\cr
Export files as .wig for use in IGV or other genome browsers.\cr
Export files as .bedo files: It is a bed file with 2 score columns.
Gives a massive speedup when cigar string and bam flags are not needed.\cr
Export files as .bedoc files: If cigar is needed, gives you replicates
and cigar, so a fast way to load a GAlignment object, other bam flags
are lost. If type is bedoc addSizeColumn and method will be ignored.
}
\details{
See \code{\link{export.ofst}}, \code{\link{export.wiggle}},
\code{\link{export.bedo}} and \code{\link{export.bedoc}}
for information on file formats.\cr
If libraries of the experiment are
already loaded into environment (default: .globalEnv) is will export
using those files as templates. If they are not in environment the
.ofst files from the bam files are loaded (unless you are converting
to .ofst then the .bam files are loaded).
}
\examples{
df <- ORFik.template.experiment()
#convertLibs(df)
# Keep only 5' ends of reads
#convertLibs(df, method = "5prime")
}
