% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{sortPerGroup}
\alias{sortPerGroup}
\title{sorts the GRangesList object grl}
\usage{
sortPerGroup(grl, equalSort = T)
}
\arguments{
\item{grl}{a GRangesList}

\item{equalSort}{a boolean, should minus strands be sorted from highest to lowest(T)}
}
\description{
sorts the GRangesList object grl
}
