% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_helpers.R
\name{splitIn3Tx}
\alias{splitIn3Tx}
\title{Create binned coverage of transcripts, split into the 3 parts.}
\usage{
splitIn3Tx(
  leaders,
  cds,
  trailers,
  reads,
  windowSize = 100,
  fraction = "1",
  weight = "score"
)
}
\arguments{
\item{leaders}{a \code{\link{GRangesList}} of leaders (5' UTRs)}

\item{cds}{a \code{\link{GRangesList}} of coding sequences}

\item{trailers}{a \code{\link{GRangesList}} of trailers (3' UTRs)}

\item{reads}{GRanges or GAlignment of reads}

\item{windowSize}{an integer (100), size of windows}

\item{fraction}{a character (1), info on reads (which read length,
or which type (RNA seq))}

\item{weight}{(default: 'score'), if defined a character name
of valid meta column in subject. GRanges("chr1", 1, "+", score = 5),
would mean score column tells that this alignment region was found 5 times.
ORFik .bedo files, contains a score column like this.
As do CAGEr CAGE files and many other package formats.
You can also assign a score column manually.}
}
\value{
a data.table with columns position, score
}
\description{
The 3 parts  of transcripts are the leaders, the cds' and trailers.
Per transcript part, bin them all to windowSize (default 100),
and make a data.table, rows are positions, useful for plotting with ORFik
and ggplot2.
}
