% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_naming.R
\name{stageNames}
\alias{stageNames}
\title{Get stage name variants}
\usage{
stageNames(zebrafish.stages = FALSE)
}
\arguments{
\item{zebrafish.stages}{logical, FALSE. If true, convert time points to stages.}
}
\value{
a data.table with 2 columns, the main name, and all name variants
of the main name in second column as a list.
}
\description{
Used to standardize nomeclature for experiments.\cr
Example: Find timepoints 2 hours, 4 hours etc.
Example: If using zebrafish stages as TRUE,
64Cell stage is same as 2 hours post fertilization,
so all 2hpf will be converted to 64Cell etc.
}
\references{
https://www.mbl.edu/zebrafish/files/2013/03/Kimmel_stagingseries1.pdf
}
\seealso{
Other experiment_naming: 
\code{\link{batchNames}()},
\code{\link{cellLineNames}()},
\code{\link{conditionNames}()},
\code{\link{fractionNames}()},
\code{\link{inhibitorNames}()},
\code{\link{libNames}()},
\code{\link{mainNames}()},
\code{\link{repNames}()},
\code{\link{tissueNames}()}
}
\concept{experiment_naming}
