% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.align.single}
\alias{STAR.align.single}
\title{Align single or paired end pair with STAR}
\usage{
STAR.align.single(
  file1,
  file2 = NULL,
  output.dir,
  index.dir,
  star.path = STAR.install(),
  fastp = install.fastp(),
  steps = "tr-ge",
  adapter.sequence = "auto",
  min.length = 15,
  trim.front = 0,
  alignment.type = "Local",
  max.cpus = min(90, detectCores() - 1),
  wait = TRUE,
  resume = NULL,
  max.multimap = 10,
  script.single = system.file("STAR_Aligner", "RNA_Align_pipeline.sh", package =
    "ORFik")
)
}
\arguments{
\item{file1}{library file, if paired must be R1 file. Allowed formats are:
(.fasta, .fastq, .fq, or.fa) with or without compression of .gz. This filename usually
 contains a suffix of .1}

\item{file2}{default NULL, set if paired end to R2 file. Allowed formats are:
(.fasta, .fastq, .fq, or.fa) with or without compression of .gz. This filename usually
 contains a suffix of .2}

\item{output.dir}{directory to save indices, default:
paste0(dirname(arguments[1]), "/STAR_index/"), where arguments is the
arguments input for this function.}

\item{index.dir}{path to STAR index folder. Path returned from ORFik function
STAR.index, when you created the index folders.}

\item{star.path}{path to STAR, default: STAR.install(),
if you don't have STAR installed at default location, it will install it there,
set path to a runnable star if you already have it.}

\item{fastp}{path to fastp trimmer, default: install.fastp(), if you
have it somewhere else already installed, give the path. Only works for
unix (linux or Mac OS), if not on unix, use your favorite trimmer and
give the output files from that trimmer as input.dir here.}

\item{steps}{a character, default: "tr-ge", trimming then genome alignment\cr
 steps of depletion and alignment wanted:
 The posible candidates you can use are:\cr
\itemize{
 \item{tr : }{trim reads}
 \item{co : }{contamination merged depletion}
 \item{ph : }{phix depletion}
 \item{rR : }{rrna depletion}
 \item{nc : }{ncrna depletion}
 \item{tR : }{trna depletion}
 \item{ge : }{genome alignment}
 \item{all: }{run steps: "tr-co-ge" or "tr-ph-rR-nc-tR-ge", depending on if you
 have merged contaminants}
}
 \cr
 If not "all", a subset of these ("tr-co-ph-rR-nc-tR-ge")\cr
 If co (merged contaminants) is used, non of the specific contaminants can be specified,
 since they should be a subset of co.\cr

 The step where you align to the genome is usually always included, unless you
 are doing pure contaminant analysis.
 For Ribo-seq and TCP(RCP-seq) you should do rR (ribosomal RNA depletion),
 so when you made the
 STAR index you need the rRNA step, either use rRNA from .gtf or manual download.
 (usually just download a Silva rRNA database
 for SSU&LSU at: https://www.arb-silva.de/) for your species.}

\item{adapter.sequence}{character, default: "auto". Auto detect adapter using fastp
adapter auto detection, checking first 1.5M reads. (auto detect adapter, is not
very reliable for Ribo-seq, so then you must include a manually specified,
else alignment will most likely fail!). You can manually assign adapter like:
"ATCTCGTATGCCGTCTTCTGCTTG" or "AAAAAAAAAAAAA". You can also specify one of the three
presets:\cr
\itemize{
 \item{illumina (standard for 100 bp sequencing): }{AGATCGGAAGAGC}
 \item{small_RNA (standard for ~50 bp sequencing): }{TGGAATTCTCGG}
 \item{nextera: }{CTGTCTCTTATA}
}
\cr If already trimmed or trimming not wanted: adapter.sequence = "disable"}

\item{min.length}{15, minimum length of reads to pass filter.}

\item{trim.front}{0, default trim 0 bases 5'. For Ribo-seq set use 0.
Ignored if tr (trim) is not one of the arguments in "steps"}

\item{alignment.type}{default: "Local": standard local alignment with soft-clipping allowed,
"EndToEnd" (global): force end-to-end read alignment, does not soft-clip.}

\item{max.cpus}{integer, default: min(90, detectCores() - 1),
number of threads to use. Default is minimum of 90 and maximum cores - 1. So if you
have 8 cores it will use 7.}

\item{wait}{a logical (not \code{NA}) indicating whether the \R
    interpreter should wait for the command to finish, or run it
    asynchronously.  This will be ignored (and the interpreter will
    always wait) if \code{intern = TRUE}. When running the command
    asynchronously, no output will be displayed on the \code{Rgui}
    console in Windows (it will be dropped, instead).}

\item{resume}{default: NULL, continue from step, lets say steps are "tr-ph-ge":
(trim, phix depletion, genome alignment) and resume is "ph", you will then use
the assumed already trimmed data and start / continue from there starting at phix,
usefull if something crashed. Like if you specified wrong STAR version, but the trimming
step was completed.}

\item{max.multimap}{numeric, default 10. If a read maps to more locations than specified,
will skip the read. Set to 1 to only get unique mapping reads. Only applies for
genome alignment step. The depletions are allowing for multimapping.}

\item{script.single}{location of STAR single file alignment script,
default internal ORFik file. You can change it and give your own if you
need special alignments.}
}
\value{
output.dir, can be used as as input in ORFik::create.experiment
}
\description{
Given a single NGS fastq/fasta library, or a paired setup of 2 mated
libraries. Run alignment and optionally remove contaminants.
}
\details{
Can only run on unix systems (Linux and Mac), and requires
minimum 30GB memory on genomes like human, rat, zebrafish etc.\cr
If for some reason the internal STAR alignment bash script will not work for you,
like if you have a very small genome. You can copy the internal alignment script,
edit it and give that as the Index script used for this function.\cr
The trimmer used is fastp (the fastest I could find), works on mac and linux.
If you want to use your own trimmer set file1/file2 to the location of
the trimmed files from your program.\cr
A note on trimming from creator of STAR about trimming:
"adapter trimming it definitely needed for short RNA sequencing.
For long RNA-seq, I would agree with Devon that in most cases adapter trimming
is not advantageous, since, by default, STAR performs local (not end-to-end) alignment,
i.e. it auto-trims." So trimming can be skipped for longer reads.
}
\examples{
# Use your own paths for annotation or the ORFik way

## use ORFik way:
output.dir <- "/Bio_data/references/Human"
# arguments <- getGenomeAndAnnotation("Homo sapiens", output.dir)
# index <- STAR.index(arguments, output.dir)
# STAR.align.single("data/raw_data/human_rna_seq/file1.bam", "data/processed/human_rna_seq",
#                    index)
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{STAR.multiQC}()},
\code{\link{STAR.remove.crashed.genome}()},
\code{\link{getGenomeAndAnnotation}()},
\code{\link{install.fastp}()}
}
\concept{STAR}
