% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.align.single}
\alias{STAR.align.single}
\title{Align single or paired end pair with STAR}
\usage{
STAR.align.single(
  file1,
  file2 = NULL,
  output.dir,
  index.dir,
  star.path = STAR.install(),
  fastp = install.fastp(),
  steps = "tr-ge",
  adapter.sequence = "auto",
  min.length = 15,
  trim.front = 0,
  alignment.type = "Local",
  max.cpus = min(90, detectCores() - 1),
  wait = TRUE,
  resume = NULL,
  script = "/export/valenfs/projects/Pipelines/STAR_Aligner/RNA_Align_pipeline.sh"
)
}
\arguments{
\item{file1}{library file, if paired must be R1 file}

\item{file2}{default NULL, set if paired end to R2 file}

\item{output.dir}{directory to save indices, default:
paste0(dirname(arguments[1]), "/STAR_index/"), where arguments is the
arguments input for this function.}

\item{index.dir}{path to STAR index}

\item{star.path}{path to STAR, default: STAR.install(),
if you don't have STAR installed at default location, it will install it there,
set path to a runnable star if you already have it.}

\item{fastp}{path to fastp trimmer, default: install.fastp(), if you
have it somewhere else already installed, give the path.}

\item{steps}{a character, default: "tr-ge", trimming --> genome alignment\cr
steps of depletion and alignment wanted:
If not "all", a subset of these ("tr-ph-rR-nc-tR-ge")\cr
In bash script it it reformated to this style:
(trimming and genome do: "tr-ge", write "all" to get all: "tr-ph-rR-nc-tR-ge")
tr: trim, ph: phix, rR: rrna, nc: ncrna, tR: trna, ge: genome)\cr
genome: the step where you align to the genome, this is usually always included.}

\item{adapter.sequence}{"auto"}

\item{min.length}{15, minimum length of reads to pass filter.}

\item{trim.front}{0, default trim 0 bases 5'. For Ribo-seq set use 0.
Ignored if tr (trim) is not one of the arguments in "steps"}

\item{alignment.type}{default: "Local": standard local alignment with soft-clipping allowed,
"EndToEnd" (global): force end-to-end read alignment, does not soft-clip.}

\item{max.cpus}{integer, default: min(90, detectCores() - 1),
number of threads to use. Default is minimum of 90 and maximum cores - 1}

\item{wait}{a logical (not \code{NA}) indicating whether the \R
    interpreter should wait for the command to finish, or run it
    asynchronously.  This will be ignored (and the interpreter will
    always wait) if \code{intern = TRUE}. When running the command
    asynchronously, no output will be displayed on the \code{Rgui}
    console in Windows (it will be dropped, instead).}

\item{resume}{default: NULL, continue from step, lets say steps are "tr-ph-ge":
(trim, phix depletion, genome alignment) and resume is "ph", you will use the trimmed
data and continue from there starting at phix, usefull if something crashed.}

\item{script}{location of STAR script}
}
\description{
If you want more than two files use: STAR.align.folder\cr
If genome aligner halts at .... loading genome, it means the star
index was aborted early, then you need to run:
STAR.remove.crashed.genome(), with the genome that crashed, and rerun.
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.index}()},
\code{\link{getGenomeAndFasta}()}
}
\concept{STAR}
