% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\name{STAR.remove.crashed.genome}
\alias{STAR.remove.crashed.genome}
\title{Remove crashed STAR genome}
\usage{
STAR.remove.crashed.genome(index.path, star.path = STAR.install())
}
\arguments{
\item{index.path}{path to index folder of genome}

\item{star.path}{path to STAR, default: STAR.install(),
if you don't have STAR installed at default location, it will install it there,
set path to a runnable star if you already have it.}
}
\value{
return value from system, 0 if all good.
}
\description{
This happens if you abort STAR run early, and it halts at: ..... loading genome
}
\examples{
# STAR.remove.crashed.genome(index.path = "/home/data/human_index/phix/)
}
\seealso{
Other STAR: 
\code{\link{STAR.align.folder}()},
\code{\link{STAR.align.single}()},
\code{\link{STAR.index}()},
\code{\link{STAR.install}()},
\code{\link{getGenomeAndFasta}()},
\code{\link{install.fastp}()}
}
\concept{STAR}
