% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ORFs.R
\name{startDefinition}
\alias{startDefinition}
\title{Returns start definition according to
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.}
\usage{
startDefinition(transl_table)
}
\arguments{
\item{transl_table}{numeric.  NCBI genetic code number for translation.}
}
\value{
A string of START sies separatd with "|".
}
\description{
Returns start definition according to
\url{http://www.ncbi.nlm.nih.gov/Taxonomy/taxonomyhome.html/index.cgi?chapter=tgencodes#SG1}
ncbi genetic code number for translation.
}
\examples{
startDefinition
startDefinition(1)

}
\seealso{
Other findORFs: \code{\link{findMapORFs}},
  \code{\link{findORFsFasta}}, \code{\link{findORFs}},
  \code{\link{stopDefinition}}
}
