% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riboseq_features.R
\name{startRegionCoverage}
\alias{startRegionCoverage}
\title{Start region coverage}
\usage{
startRegionCoverage(grl, RFP, tx = NULL, is.sorted = TRUE,
  upstream = 2L, downstream = 2L)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object
with usually either leaders, cds', 3' utrs or ORFs}

\item{RFP}{ribo seq reads as GAlignment, GRanges or GRangesList object}

\item{tx}{default NULL, a GRangesList of transcripts or (container region),
names of tx must contain all grl names. The names of grl can also be the
ORFik orf names. that is "txName_id"}

\item{is.sorted}{logical (TRUE), is grl sorted.}

\item{upstream}{an integer (2), relative region to get upstream from.}

\item{downstream}{an integer (2), relative region to get downstream from}
}
\value{
a numeric vector of counts
}
\description{
Get the number of reads in the start region of each ORF. If you want the
start codon coverage only, set upstream = 0. Standard is 2 upstream
and 2 downstream, a width 5 window centered at start site. since
p-shifting is not 100% accurate, this window is usually the reads from the
start site.
}
\details{
If tx is null, then upstream will be force to 0 and downstream to
a maximum of grl width. Since there is no reference for splicing.
}
\seealso{
Other features: \code{\link{computeFeaturesCage}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}},
  \code{\link{distToCds}}, \code{\link{distToTSS}},
  \code{\link{entropy}}, \code{\link{floss}},
  \code{\link{fpkm_calc}}, \code{\link{fpkm}},
  \code{\link{fractionLength}},
  \code{\link{initiationScore}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}}, \code{\link{orfScore}},
  \code{\link{rankOrder}},
  \code{\link{ribosomeReleaseScore}},
  \code{\link{ribosomeStallingScore}},
  \code{\link{startRegion}}, \code{\link{subsetCoverage}},
  \code{\link{translationalEff}}
}
\concept{features}
