% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{startRegionString}
\alias{startRegionString}
\title{Get -20,20 start region as DNA characters per GRanges group}
\usage{
startRegionString(grl, tx, faFile, groupBy = NULL)
}
\arguments{
\item{grl}{a GRangesList to find regions}

\item{tx}{a GRangesList of transcripts containing grl}

\item{faFile}{a FaFile from the fasta file, see ?FaFile.
Can also be path to fastaFile with fai file in same dir.}

\item{groupBy}{(NULL) column to group grl by, if NULL group by names(gr)}
}
\value{
a character vector of start regions
}
\description{
Get -20,20 start region as DNA characters per GRanges group
}
