% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORFs_helpers.R
\name{stopSites}
\alias{stopSites}
\title{Get the stop sites from a GRangesList of orfs grouped by orfs}
\usage{
stopSites(grl, asGR = FALSE, keep.names = FALSE, is.sorted = FALSE)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges:GRangesList]{GenomicRanges::GRangesList()}} object}

\item{asGR}{a boolean, return as GRanges object}

\item{keep.names}{if asGR is False, do you still want
to keep a named vector}

\item{is.sorted}{a speedup, if you know the ranges are sorted}
}
\value{
if asGR is False, a vector, if True a GRanges object
}
\description{
In ATGTTTTGC, get the position of the C.
}
\examples{
gr_plus <- GRanges(seqnames = c("chr1", "chr1"),
                   ranges = IRanges(c(7, 14), width = 3),
                   strand = c("+", "+"))
gr_minus <- GRanges(seqnames = c("chr2", "chr2"),
                    ranges = IRanges(c(4, 1), c(9, 3)),
                    strand = c("-", "-"))
grl <- GRangesList(tx1 = gr_plus, tx2 = gr_minus)
stopSites(grl, is.sorted = FALSE)

}
