% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_helpers.R
\name{strandBool}
\alias{strandBool}
\title{helper function to get a logical list of True/False, if GRangesList group have
+ strand = T, if - strand = F
Also checks for * strands, so a good check for bugs}
\usage{
strandBool(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} or GRanges object}
}
\value{
a logical vector
}
\description{
helper function to get a logical list of True/False, if GRangesList group have
+ strand = T, if - strand = F
Also checks for * strands, so a good check for bugs
}
