% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{subset_coverage}
\alias{subset_coverage}
\title{Subset GRanges to get coverage. GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.}
\usage{
subset_coverage(cov, y)
}
\arguments{
\item{cov}{A coverage object from coverage()}

\item{y}{GRanges object for which coverage should be extracted}
}
\value{
numeric vector of coverage of input GRanges object
}
\description{
Subset GRanges to get coverage. GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.
}
\examples{
#subset_coverage(x)

}
