% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{subsetToFrame}
\alias{subsetToFrame}
\title{Subset GRanges to get desired frame.}
\usage{
subsetToFrame(x, frame)
}
\arguments{
\item{x}{A tiled to size of 1 GRanges object}

\item{frame}{A numeric indicating which frame to extract}
}
\value{
GRanges object reduced to only first frame
}
\description{
Usually used for ORFs to get specific frame (0-2):
frame 0, frame 1, frame 2
}
\details{
GRanges object should be beforehand
tiled to size of 1. This subsetting takes account for strand.
}
