% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{symbols}
\alias{symbols}
\title{Get ORFik experiment QC folder path}
\usage{
symbols(x)
}
\arguments{
\item{x}{an ORFik \code{\link{experiment}}}
}
\value{
a data.table with gene id, gene symbols and tx ids (3 columns)
}
\description{
Get ORFik experiment QC folder path
}
