% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{te}
\alias{te}
\title{Translational efficiency}
\usage{
te(grl, RNA, RFP, tx_len)
}
\arguments{
\item{grl}{a GRangesList object with usually either leaders,
cds', 3' utrs or ORFs. ORFs is a special case, see argument tx_len}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{tx_len}{the transcript lengths of the transcripts
a named (tx names) vector.
Normally you call argument as: tx_len = txLen(Gtf)
fractionLength(grl, tx_len = txLen(Gtf))
See ?ORFik:::txLen
NB!!! if you used cage data, then
the tss for the the leaders have changed,
therefor the tx lengths have changed,
if so, call:
te(grl, RNA, RFP, tx_len = TxLen(Gtf, fiveUTRs))
where fiveUTRs are the changed cageLeaders as GRangesList.}
}
\description{
Uses Rna-seq and Ribo-seq to get te of grl
is defined as (density of RPFs within ORF)/(RNA expression).
See article: 10.1126/science.1168978
}
