% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_methods.R
\name{te}
\alias{te}
\title{Translational efficiency}
\usage{
te(grl, RNA, RFP, tx, with.fpkm = F, pseudoCount = 0)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object
can be either transcripts, 5' utrs, cds', 3' utrs or
ORFs are a special case (uORFs, potential new cds' etc).}

\item{RNA}{rna seq reads as GAlignment, GRanges
or GRangesList object}

\item{RFP}{ribo seq reads as GAlignment, GRanges
or GRangesList object}

\item{tx}{a GRangesList of the transcripts,
NB!!! if you used cage data, then
the tss for the the leaders have changed,
therefor the tx lengths have changed,
if so, call:
te(grl, RNA, RFP, tx= TxLen(Gtf, fiveUTRs))
where fiveUTRs are the changed cageLeaders.}

\item{with.fpkm}{logical F, if true return the fpkm values together with te}

\item{pseudoCount}{an integer, 0,
set it to 1 if you want to avoid NA and inf values.}
}
\value{
a numeric vector of fpkm ratios, if with.fpkm is TRUE
 ,return a data.table with te and fpkm values
}
\description{
Uses Rna-seq and Ribo-seq to get te of grl
is defined as (density of RPFs within ORF)/(RNA expression).
See article: 10.1126/science.1168978
}
\examples{
ORF <- GRanges(seqnames = "1",
ranges = IRanges(start = c(1, 10, 20),
                end = c(5, 15, 25)),
strand = "+")
grl <- GRangesList(tx1_1 = ORF)
RFP <- GRanges("1", IRanges(25, 25),"+")
RNA <- GRanges("1", IRanges(1, 50),"+")
tx <-  GRangesList(tx1 = GRanges("1", IRanges(1, 50),"+"))
# nb: grl must have same names as tx, + _1 etc
te <- te(grl, RNA, RFP, tx, with.fpkm = TRUE, pseudoCount = 1)
te$fpkmRFP
te$te
}
\seealso{
Other features: \code{\link{ORFDistToCds}},
  \code{\link{ORFScores}}, \code{\link{OrfRankOrder}},
  \code{\link{RibosomeReleaseScore}},
  \code{\link{RibosomeStallingScore}},
  \code{\link{computeFeatures}},
  \code{\link{disengagementScore}}, \code{\link{entropy}},
  \code{\link{floss}}, \code{\link{fpkm_calc}},
  \code{\link{fpkm}}, \code{\link{fractionLength}},
  \code{\link{insideOutsideORF}}, \code{\link{isInFrame}},
  \code{\link{isOverlapping}},
  \code{\link{kozakSequenceScore}},
  \code{\link{subsetCoverage}}
}
