% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{tile1}
\alias{tile1}
\title{Tile a GRangeslist by 1
This is not supported originally by GenomicRanges}
\usage{
tile1(grl)
}
\arguments{
\item{grl}{a GRangesList object}
}
\value{
a GRangesList grouped by original group, tiled to 1
}
\description{
Tile a GRangeslist by 1
This is not supported originally by GenomicRanges
}
