% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_helpers.R
\name{tile1}
\alias{tile1}
\title{Tile a GRangeslist by 1}
\usage{
tile1(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object}
}
\value{
a GRangesList grouped by original group, tiled to 1
}
\description{
Per group, sepereate the groups and split them on each position.
Returned sorted
This is not supported originally by GenomicRanges
As a precaution, this function requires the unlisted objects to
have names, so that regrouping will be done correctly.
}
\examples{
gr1 <- GRanges("1",
           ranges = IRanges(start = c(1, 10, 20),
           end = c(5, 15, 25)),
           strand = "+")

gr2 <- GRanges("1", ranges = IRanges(start = c(20, 30, 40),
                                      end = c(25, 35, 45)),
                     strand = "+")
names(gr1) = rep("tx1_1",3)
names(gr2) = rep("tx1_2",3)
grl <- GRangesList(tx1_1 = gr1, tx1_2 = gr2)
tile1(grl)
}
