% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{tile1}
\alias{tile1}
\title{Tile a GRangesList by 1}
\usage{
tile1(grl, sort.on.return = TRUE, matchNaming = TRUE)
}
\arguments{
\item{grl}{a \code{\link{GRangesList}} object with names}

\item{sort.on.return}{logical (T), should the groups be
sorted before return.}

\item{matchNaming}{logical (T), should groups keep unlisted names
and meta data.(This make the list very big, for > 100K groups)}
}
\value{
a GRangesList grouped by original group, tiled to 1
}
\description{
Will tile a GRangesList into single bp resolution, each group of the list
will be splited by positions of 1. Returned values are sorted.
This is not supported originally by GenomicRanges.
As a precaution, this function requires the unlisted objects to
have names.
}
\examples{
gr1 <- GRanges("1", ranges = IRanges(start = c(1, 10, 20),
                                     end = c(5, 15, 25)),
               strand = "+")
gr2 <- GRanges("1", ranges = IRanges(start = c(20, 30, 40),
                                     end = c(25, 35, 45)),
               strand = "+")
names(gr1) = rep("tx1_1", 3)
names(gr2) = rep("tx1_2", 3)
grl <- GRangesList(tx1_1 = gr1, tx1_2 = gr2)
tile1(grl)

}
