% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{tile1}
\alias{tile1}
\title{Tile a GRangeslist by 1}
\usage{
tile1(grl)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}} object}
}
\value{
a GRangesList grouped by original group, tiled to 1
}
\description{
Per group, sepereate the groups and split them on each position.
Returned sorted
This is not supported originally by GenomicRanges
}
