% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_plots.R
\name{transcriptWindow}
\alias{transcriptWindow}
\title{Make 100 bases size meta window for all libraries in experiment}
\usage{
transcriptWindow(
  leaders,
  cds,
  trailers,
  df,
  outdir = NULL,
  scores = c("sum", "zscore"),
  allTogether = TRUE,
  colors = rep("skyblue4", nrow(df)),
  title = "Coverage metaplot",
  windowSize = min(100, min(widthPerGroup(leaders, FALSE)), min(widthPerGroup(cds,
    FALSE)), min(widthPerGroup(trailers, FALSE))),
  returnPlot = is.null(outdir),
  dfr = NULL,
  idName = "",
  format = ".png",
  type = "default"
)
}
\arguments{
\item{leaders}{a \code{\link{GRangesList}} of leaders (5' UTRs)}

\item{cds}{a \code{\link{GRangesList}} of coding sequences}

\item{trailers}{a \code{\link{GRangesList}} of trailers (3' UTRs)}

\item{df}{an ORFik \code{\link{experiment}}}

\item{outdir}{directory to save to (default: NULL, no saving)}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}

\item{allTogether}{plot all coverage plots in 1 output? (defualt: TRUE)}

\item{colors}{Which colors to use, default (skyblue4)}

\item{title}{title of ggplot}

\item{windowSize}{size of binned windows, default: 100}

\item{returnPlot}{return plot from function, default is.null(outdir),
so TRUE if outdir is not defined.}

\item{dfr}{an ORFik \code{\link{experiment}} of RNA-seq to
normalize against. Will add RNA normalized to plot name if this is done.}

\item{idName}{A character ID to add to saved name of plot,
if you make several plots in the same folder,
and same experiment, like splitting transcripts in two groups like
targets / nontargets etc. (default: "")}

\item{format}{default (".png"), do ".pdf" if you want as pdf}

\item{type}{a character(default: "default"), load files in experiment
or some precomputed variant, either "bedo", "bedo" or "pshifted".
These are made with ORFik:::simpleLibs(), shiftFootprintsByExperiment()..}
}
\value{
NULL, or ggplot object if returnPlot is TRUE
}
\description{
Gives you binned meta coverage plots, either saved seperatly or
all in one.
}
\examples{
# Make ORFik experiment
dir <- system.file("extdata", "", package = "ORFik")
# 2. Pick an experiment name
exper <- "ORFik"
# 3. Pick .gff/.gtf location
txdb <- system.file("extdata", "annotations.gtf", package = "ORFik")
template <- create.experiment(dir = dir, exper, txdb = txdb,
                              viewTemplate = FALSE)
template$X5[6] <- "heart" # <- fix non unique row
# read experiment
df <- read.experiment(template)
loadRegions(df) # Load leader, cds and trailers as GRangesList
#transcriptWindow(leaders, cds, trailers, df, outdir = "directory_to_save")
}
\seealso{
Other experiment plots: 
\code{\link{transcriptWindow1}()},
\code{\link{transcriptWindowPer}()}
}
\concept{experiment plots}
