% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_plots.R
\name{transcriptWindow}
\alias{transcriptWindow}
\title{Make 100 bases size meta window for all libraries in experiment}
\usage{
transcriptWindow(leaders, cds, trailers, df, outdir, scores = c("sum",
  "zscore"), allTogether = FALSE, colors = rep("skyblue4", nrow(df)),
  windowSize = min(100, min(widthPerGroup(leaders, F)),
  min(widthPerGroup(cds, F)), min(widthPerGroup(trailers, F))),
  returnPlot = FALSE)
}
\arguments{
\item{leaders}{a \code{\link{GRangesList}} of leaders (5' UTRs)}

\item{cds}{a \code{\link{GRangesList}} of coding sequences}

\item{trailers}{a \code{\link{GRangesList}} of trailers (3' UTRs)}

\item{df}{an ORFik experiment, to make it, see: ?experiment}

\item{outdir}{directory to save to}

\item{scores}{scoring function (default: c("sum", "zscore"))}

\item{allTogether}{plot all coverage plots in 1? (defualt: FALSE)}

\item{colors}{Which colors to use, default (skyblue4)}

\item{windowSize}{size of binned windows, default: 100}

\item{returnPlot}{return plot from function, default False}
}
\value{
NULL, or ggplot object if returnPlot is TRUE
}
\description{
Gives you binned meta coverage plots, either saved seperatly or
all in one.
}
