% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_plots.R
\name{transcriptWindow1}
\alias{transcriptWindow1}
\title{Meta coverage over all transcripts}
\usage{
transcriptWindow1(
  df,
  outdir = NULL,
  scores = c("sum", "zscore"),
  colors = rep("skyblue4", nrow(df)),
  title = "Coverage metaplot",
  windowSize = 100,
  returnPlot = is.null(outdir),
  dfr = NULL,
  idName = "",
  format = ".png",
  type = "default"
)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{outdir}{directory to save to (default: NULL, no saving)}

\item{scores}{scoring function (default: c("sum", "zscore")),
see ?coverageScorings for possible scores.}

\item{colors}{Which colors to use, default (skyblue4)}

\item{title}{title of ggplot}

\item{windowSize}{size of binned windows, default: 100}

\item{returnPlot}{return plot from function, default is.null(outdir),
so TRUE if outdir is not defined.}

\item{dfr}{an ORFik \code{\link{experiment}} of RNA-seq to
normalize against. Will add RNA normalized to plot name if this is done.}

\item{idName}{A character ID to add to saved name of plot,
if you make several plots in the same folder,
and same experiment, like splitting transcripts in two groups like
targets / nontargets etc. (default: "")}

\item{format}{default (".png"), do ".pdf" if you want as pdf}

\item{type}{a character(default: "defualt"), load files in experiment
or some precomputed variant, either "bedo" or "pshifted".
These are made with ORFik:::simpleLibs()}
}
\value{
NULL, or ggplot object if returnPlot is TRUE
}
\description{
Given as single window
}
\seealso{
Other experiment plots: 
\code{\link{transcriptWindowPer}()},
\code{\link{transcriptWindow}()}
}
\concept{experiment plots}
