% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_helpers.R
\name{trim_detection}
\alias{trim_detection}
\title{Add trimming info to QC report}
\usage{
trim_detection(df, finals, out.dir)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}

\item{finals}{a data.table with current output from QCreport}

\item{out.dir}{optional output directory, default:
\code{dirname(df$filepath[1])}.
Will make a folder called "QC_STATS" with all results in this directory.}
}
\value{
a data.table of the update finals object with trim info
}
\description{
Only works if alignment was done using ORFik with STAR.
}
