% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR_multiQC.R
\name{trimming.table}
\alias{trimming.table}
\title{Create trimming table}
\usage{
trimming.table(trim_folder)
}
\arguments{
\item{trim_folder}{folder of trimmed files}
}
\value{
a data.table with 4 columns, raw_library (names of library),
 raw_reads (numeric, number of raw reads),
 trim_reads (numeric, number of trimmed reads),
 % trimmed (numeric, percentage of trimmed reads)
}
\description{
From fastp runs in ORFik alignment process
}
