% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_helpers.R
\name{txLen}
\alias{txLen}
\title{Get transcript lengths}
\usage{
txLen(Gtf = NULL, changedFiveUTRs = NULL)
}
\arguments{
\item{Gtf}{a TxDb object of a gtf file
NB! only add this if you dont use tx argument}

\item{changedFiveUTRs}{a GRangesList object of leaders.
NB! only add this if you used cage data or other things to change the
leaders, therefor we need it to update transcript lengths.}
}
\description{
A helper function for easy length retrieval
}
