% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{txNamesToGeneNames}
\alias{txNamesToGeneNames}
\title{Convert transcript names to gene names}
\usage{
txNamesToGeneNames(txNames, txdb)
}
\arguments{
\item{txNames}{character vector, the transcript names to convert.}

\item{txdb}{the transcript database to use or gtf/gff path to it.}
}
\value{
character vector of gene names
}
\description{
Works for ensembl etc.
}
