% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdbHelpers.R
\name{updateTxdbStartSites}
\alias{updateTxdbStartSites}
\title{Update start sites of leaders}
\usage{
updateTxdbStartSites(txList, fiveUTRs)
}
\arguments{
\item{txList}{a list, call of as.list(txdb)}

\item{fiveUTRs}{a GRangesList of 5' leaders}
}
\value{
a list, modified call of as.list(txdb)
}
\description{
Update start sites of leaders
}
