% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{upstreamOfPerGroup}
\alias{upstreamOfPerGroup}
\title{Get rest of objects upstream}
\usage{
upstreamOfPerGroup(tx, upstreamOf)
}
\arguments{
\item{tx}{a \code{\link{GRangesList}},
usually of Transcripts to be changed}

\item{upstreamOf}{a vector of integers, for each group in tx, where
is the new stop point of last valid exon.}
}
\value{
a GRangesList of upstream part
}
\description{
Per group get the part upstream of position
 defined in upstreamOf
 upstream(tx, ORFik:::stopSites(cds, asGR = F))
 will return the 5' utrs per transcript, usually used for interesting
 parts of the transcripts, like upstream open reading frames(uorf).
 downstreamOf +/- 1 is start/end site
 of transformed tx's, depending on strand
}
