% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_positions.R
\name{upstreamOfPerGroup}
\alias{upstreamOfPerGroup}
\title{Get rest of objects upstream (exclusive)}
\usage{
upstreamOfPerGroup(tx, upstreamOf, allowOutside = TRUE)
}
\arguments{
\item{tx}{a \code{\link{GRangesList}},
usually of Transcripts to be changed}

\item{upstreamOf}{a vector of integers, for each group in tx, where
is the the base after the new stop point of last valid exon.}

\item{allowOutside}{a logical (T), can upstreamOf extend outside
range of tx, can set boundary as a false hit, so beware.}
}
\value{
a GRangesList of upstream part
}
\description{
Per group get the part upstream of position
upstreamOfPerGroup(tx, startSites(cds, asGR = TRUE))
will return the 5' utrs per transcript, usually used for interesting
parts of the transcripts.
}
