% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{upstreamOfPerGroup}
\alias{upstreamOfPerGroup}
\title{get rest of objects upstream}
\usage{
upstreamOfPerGroup(tx, upstreamOf)
}
\arguments{
\item{tx}{a \code{\link[GenomicRanges]{GRangesList}},
usually of Transcripts to be changed}

\item{upstreamOf}{a vector of integers, for each group in tx, where
is the new start point of first valid exon.}
}
\description{
per group get the part upstream of position
 defined in upstreamOf
 upstream(tx, ORFik:::ORFStopSites(cds, asGR = F))
 will return the 5' utrs per transcript, usually used for interesting
 parts of the transcripts, like upstream open reading frames(uorf).
 downstreamOf +/- 1 is start/end site
 of transformed tx's, depending on strand
}
