% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{validateExperiments}
\alias{validateExperiments}
\title{Validate ORFik experiment
Check for valid non-empty files etc.}
\usage{
validateExperiments(df)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}
}
\value{
NULL (Stops if failed)
}
\description{
Validate ORFik experiment
Check for valid non-empty files etc.
}
