% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{validateExperiments}
\alias{validateExperiments}
\title{Validate ORFik experiment
Check for valid non-empty files etc.}
\usage{
validateExperiments(df)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}
}
\value{
NULL (Stops if failed)
}
\description{
Validate ORFik experiment
Check for valid non-empty files etc.
}
\seealso{
Other ORFik_experiment: 
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{libraryTypes}()},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()}
}
\concept{ORFik_experiment}
