% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{validateExperiments}
\alias{validateExperiments}
\title{Validate ORFik \code{\link{experiment}}}
\usage{
validateExperiments(df)
}
\arguments{
\item{df}{an ORFik \code{\link{experiment}}}
}
\value{
NULL (Stops if failed)
}
\description{
Check for valid existing, non-empty and all unique.
A good way to see if your experiment is valid.
}
\seealso{
Other ORFik_experiment: 
\code{\link{ORFik.template.experiment}()},
\code{\link{bamVarName}()},
\code{\link{create.experiment}()},
\code{\link{experiment-class}},
\code{\link{filepath}()},
\code{\link{libraryTypes}()},
\code{\link{organism.df}()},
\code{\link{outputLibs}()},
\code{\link{read.experiment}()},
\code{\link{save.experiment}()}
}
\concept{ORFik_experiment}
