% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_helpers.R
\name{validGRL}
\alias{validGRL}
\title{Helper Function to check valid GRangesList input}
\usage{
validGRL(class, type = "grl", checkNULL = FALSE)
}
\arguments{
\item{class}{as character vector the given class of
supposed GRangesList object}

\item{type}{a character vector, is it gtf, cds, 5', 3', for messages.}

\item{checkNULL}{should NULL classes be checked and return indeces of these?}
}
\description{
Helper Function to check valid GRangesList input
}
