% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validSeqlevels}
\alias{validSeqlevels}
\title{Helper function to find overlaping seqlevels}
\usage{
validSeqlevels(grl, reads)
}
\arguments{
\item{grl}{a GRangesList or GRanges object}

\item{reads}{a GRanges or GAlignment object}
}
\value{
a character vector of valid seqlevels
}
\description{
Useful to avoid warnings in bioC
}
