% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRanges_helpers.R
\name{widthPerGroup}
\alias{widthPerGroup}
\title{get list of widths per granges group}
\usage{
widthPerGroup(grl, keep.names = T)
}
\arguments{
\item{grl}{a \code{\link[GenomicRanges]{GRangesList}}}

\item{keep.names}{a boolean, keep names or not}
}
\value{
an integer vector (named/unnamed) of widths
}
\description{
get list of widths per granges group
}
