% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{windowCoveragePlot}
\alias{windowCoveragePlot}
\title{Get coverage window plot of reads}
\usage{
windowCoveragePlot(coverage, output = NULL, scoring = "zscore",
  colors = c("skyblue4", "orange"), title = "Coverage metaplot",
  type = "transcript")
}
\arguments{
\item{coverage}{a data.table, e.g. output of scaledWindowCoverage}

\item{output}{character string (NULL), if set, saves the plot as pdf or png
to path given. If no format is given, is save as pdf.}

\item{scoring}{character vector (zscore), either of zScore,
transcriptNormalized, sum, mean, median, NULL. Set NULL if already scored.}

\item{colors}{character vector colors to use in plot, will fix automaticly,
using binary splits with colors c('skyblue4', 'orange').}

\item{title}{a character (metaplot) (what is the title of plot?)}

\item{type}{a character (transcript), what should legends say is
the whole region? Transcript, gene, non coding rna etc.}
}
\value{
a ggplot object of the coverage plot, NULL if output is set,
then the plot will only be saved to location.
}
\description{
Spanning a region like a transcripts, plot how the reads distribute.
}
\details{
If you return this function without assigning it and output is NULL,
it will automaticly plot the figure in your session. If output is assigned,
no plot will be shown in session.

Remember if you want to change anything like colors, just return the
ggplot object, and reassign like: obj + scale_color_brewer() etc.
}
\examples{
library(data.table)
coverage <- data.table(position = seq(20), score = cumsum(seq(20)))
windowCoveragePlot(coverage)
# See vignette for a more practical example

}
\seealso{
Other coveragePlot: \code{\link{coverageHeatMap}},
  \code{\link{pSitePlot}}, \code{\link{savePlot}}
}
\concept{coveragePlot}
