% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges_helpers.R
\name{windowPerGroup}
\alias{windowPerGroup}
\title{Get window region of tx around point of gr}
\usage{
windowPerGroup(gr, tx, downstream = 0L, upstream = 0L)
}
\arguments{
\item{gr}{a GRanges object (startSites and others, must be single point)}

\item{tx}{a GRangesList of transcripts or (container region)}

\item{downstream}{an integer, relative region to get downstream from}

\item{upstream}{an integer vector, relative region to get upstream from.}
}
\value{
a GRanges, or GRangesList object if any group had > 1 exon.
}
\description{
If downstreamFrom is 20, it means the window will start -20 downstream of
gr start site. It will keep exon structure of tx
}
\examples{
# find 2nd codon of an ORF on a spliced transcript
ORF <- GRanges("1", c(3), "+") # start site
names(ORF) <- "tx1_1" # ORF 1 on tx1
tx <- GRangesList(tx1 = GRanges("1", c(1,3,5,7,9,11,13), "+"))
windowPerGroup(ORF, tx, 5, -3) # <- 2nd codon

}
\seealso{
Other ExtendGenomicRanges: \code{\link{asTX}},
  \code{\link{coveragePerTiling}},
  \code{\link{overlapsToCoverage}},
  \code{\link{pmapFromTranscriptF}},
  \code{\link{reduceKeepAttr}}, \code{\link{tile1}},
  \code{\link{txSeqsFromFa}}
}
\concept{ExtendGenomicRanges}
