% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{yAxisScaler}
\alias{yAxisScaler}
\title{Scale y axis correctly}
\usage{
yAxisScaler(covPos, increments.y = "auto")
}
\arguments{
\item{covPos}{a levels object from a factor of y axis}

\item{increments.y}{increments of y axis, default "auto".
Or a numeric value < max position & > min position.}
}
\value{
a character vector from the seq() function, aligned to 0.
}
\description{
Works for all coverage plots.
}
